% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AddAreaPositionStoxLanding}
\alias{AddAreaPositionStoxLanding}
\title{Append position to landings data}
\usage{
AddAreaPositionStoxLanding(
  StoxLandingData,
  AreaPosition,
  resolution = c("Area", "Location"),
  latColName = NULL,
  lonColName = NULL
)
}
\arguments{
\item{StoxLandingData}{landing data, see \code{\link[RstoxData]{StoxLandingData}}}

\item{AreaPosition}{coordinates for Area and SubArea codes, see \code{\link[RstoxFDA]{AreaPosition}}}

\item{resolution}{character(), defaults to Area, specify what resolution to use: 'Area' or 'Location'. See details.}

\item{latColName}{name of the column which should be appended for latitudes. Defaults to "Latitude".}

\item{lonColName}{name of the column which should be appended for longitudes. Defaults to "Longitude".}
}
\value{
\code{\link[RstoxData]{StoxLandingData}} with columns for latitude and longitude appended.
}
\description{
StoX function
 Appends a position to landings data, based on Area and Location codes.
}
\details{
When 'resolution' is specified as 'Area' Area is looked up from 'AreaPosition', using the row where 'Location' is missing.
 When 'resolution' is specified as 'Location', 'Area' and 'Location' is looked up 'AreaPosition'.
}
