% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AddGearGroupStoxBiotic}
\alias{AddGearGroupStoxBiotic}
\title{Add Gear group to StoxBioticData}
\usage{
AddGearGroupStoxBiotic(StoxBioticData, Translation)
}
\arguments{
\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} data which will be annotated.}

\item{Translation}{Translation table (\code{\link[RstoxData]{Translation}}). See details.}
}
\value{
StoxBioticData with column 'GearGroup' appended. See \code{\link[RstoxData]{StoxBioticData}}.
}
\description{
Adds a column to StoxBioticData with gear groups
}
\details{
The provided Translation should maps values ('Value' in Translation) for the variable 'Gear' ('VariableName' in Translation) in 'StoxBioticData' to a gear group ('NewValue').
}
