% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AddGearGroupStoxLanding}
\alias{AddGearGroupStoxLanding}
\title{Add Gear group to StoxLandingData}
\usage{
AddGearGroupStoxLanding(StoxLandingData, Translation)
}
\arguments{
\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} data which will be annotated.}

\item{Translation}{Translation table (\code{\link[RstoxData]{Translation}}). See details.}
}
\value{
StoxLandingData with column 'GearGroup' appended. See \code{\link[RstoxData]{StoxLandingData}}.
}
\description{
Adds a column to StoxLandingData with gear groups
}
\details{
The provided Translation should maps values ('Value' in Translation) for the variable 'Gear' ('VariableName' in Translation) in 'StoxBioticData' to a gear group ('NewValue').
}
