% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landingsAnalyses.R
\name{addLogbookColumns}
\alias{addLogbookColumns}
\title{Add columns from logbooks}
\usage{
addLogbookColumns(
  landings,
  logbooks,
  logbookColumns,
  tripIdCol = "tripid",
  catchIdCol,
  speciesColLand = "Art FAO (kode)",
  speciesColLog = "FANGSTART_FAO"
)
}
\arguments{
\item{landings}{\code{\link[data.table]{data.table}} containing landings records, and columns identifying the trip and catch ('tripIdCol', 'catchIdCol').}

\item{logbooks}{\code{\link[data.table]{data.table}} containing logbook records, and columns identifying the trip and catch ('tripIdCol', 'catchIdCol').}

\item{logbookColumns}{character() vector that identifies columns in logbooks that are to be added to 'landings'}

\item{tripIdCol}{character() that identifies the column in 'landings' and 'logbooks' that contain trip ids}

\item{catchIdCol}{character() that identifies the column in 'landings' and 'logbooks' that contain catch ids}

\item{speciesColLand}{character() that identifies the column in 'landings' that identify species landed}

\item{speciesColLog}{character() that identifies the column in 'logbooks' that identify species caught.}
}
\description{
Add columns to landings from corresponding logbook entries.
}
\details{
'landings' should already imputed with information about individual catches.
 The columns 'tripIdCol' and 'catchIdCol' must be set on 'landings' and 'logbooks'. This may be achieved by \code{\link[RstoxFDA]{appendTripIdLogbooks}}, \code{\link[RstoxFDA]{appendTripIdLandings}} and \code{\link[RstoxFDA]{imputeCatchesLandings}}.
 
 'landings' may contain records with no correspond logbook entires ('tripIdCol' contains NAs).
 The added columns will have NA for these records.
}
\seealso{
\code{\link[RstoxFDA]{makeTripIds}}, \code{\link[RstoxFDA]{appendTripIdLandings}}, \code{\link[RstoxFDA]{appendTripIdLogbooks}}, and \code{\link[RstoxFDA]{imputeCatchesLandings}} for obtaining 'landings' and 'logbooks' with trip-ids and catch-ids.

Other logbook functions: 
\code{\link{appendTripIdLandings}()},
\code{\link{appendTripIdLogbooks}()},
\code{\link{calculateLogbookPartitionByTrip}()},
\code{\link{imputeCatchesLandings}()},
\code{\link{logbookAdjustment}()},
\code{\link{makeTripIds}()},
\code{\link{sourceLogbookColumns}()}

Other landings functions: 
\code{\link{ReadLandingFDA}()},
\code{\link{ReportFdaLandings}()},
\code{\link{ReportFdaSampling}()},
\code{\link{appendTripIdLandings}()},
\code{\link{imputeCatchesLandings}()},
\code{\link{logbookAdjustment}()},
\code{\link{makeTripIds}()},
\code{\link{plotBubbleMap}()},
\code{\link{sourceLogbookColumns}()},
\code{\link{tabulateFisheries}()}
}
\concept{landings functions}
\concept{logbook functions}
