% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AddPeriodStoxBiotic}
\alias{AddPeriodStoxBiotic}
\title{Add Period to StoxBioticData}
\usage{
AddPeriodStoxBiotic(StoxBioticData, TemporalDefinition)
}
\arguments{
\item{StoxBioticData}{\code{\link[RstoxData]{StoxLandingData}} data which will be annotated with period.}

\item{TemporalDefinition}{\code{\link[RstoxFDA]{TemporalDefinition}} definiton of temporal category.}
}
\value{
StoxBioticData with column appended. See \code{\link[RstoxData]{StoxBioticData}}.
\code{\link[RstoxFDA]{DefinePeriod}} for configuring the temporal definition,
\code{\link[RstoxFDA]{AddStratumStoxLanding}} for similar function for landing data,
\code{\link[RstoxFDA]{PrepareRecaEstimate}} for use of 'Period' as an effect in Reca-estimation,
and \code{\link[RstoxFDA]{ReportFdaSampling}} for use of 'Period' as an aggregation variable when comparing sampling with landed volume.
}
\description{
Add a column 'Period' to 'Station' on StoxBioticData with a temporal category, such as 'quarter'.
}
\details{
'Period' will be added based on the column 'DateTime' on the table 'Station' in 'StoxBioticData'.

Temporal definitions (\code{\link[RstoxFDA]{TemporalDefinition}}) may be produced by
\code{\link[RstoxFDA]{DefinePeriod}}
}
\concept{StoX-functions}
\concept{temporal coding functions}
