% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalyticalBaselineFunctions.R
\name{AddPsuStratificationVariables}
\alias{AddPsuStratificationVariables}
\title{Add Stratification columns to 'PSUSamplingParametersData'}
\usage{
AddPsuStratificationVariables(
  PSUSamplingParametersData,
  StratificationVariables,
  StratificationVariablesTable = data.table::data.table()
)
}
\arguments{
\item{PSUSamplingParametersData}{Sampling parameters stratification variables should be added to}

\item{StratificationVariables}{name of variables to add}

\item{StratificationVariablesTable}{value-combinations for the variables to add to each stratum}
}
\value{
\code{\link[RstoxFDA]{PSUSamplingParametersData}}
}
\description{
Add additional variables to encode strata and its correspondance with census data (e.g. landings data).
}
\details{
\code{\link[RstoxFDA]{PSUSamplingParametersData}} provide sampling parameters by strata.
Optionally, it may also contain additional variables that encode the stratification in terms of variables
available in other data sources, such as \code{\link[RstoxData]{StoxLandingData}}. This function allows
such variables to be added, if not already present.

More detailed encoding of stratification is useful for
encoding the sampling frame of the design provided by 'PSUSamplingParametersData'. By encoding all strata
in terms of variables that are available in census-data, the correspondance between sampling frame and
target population can be encoded. This information will be available in downstream estimates (e.g.
\code{\link[RstoxFDA]{AnalyticalPopulationEstimate}}) and allow for pragmatic inference to
out-of-frame strata (via \code{\link[RstoxFDA]{ExtendAnalyticalSamplingFrameCoverage}}).
}
\concept{Analytical estimation}
\concept{StoX-functions}
