% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AddStratumStoxBiotic}
\alias{AddStratumStoxBiotic}
\title{Adds Stratum to StoxBioticData}
\usage{
AddStratumStoxBiotic(StoxBioticData, StratumPolygon)
}
\arguments{
\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} data which will be annotated. Needs postions appended. See details.}

\item{StratumPolygon}{Definition of spatial strata. See \code{\link[RstoxBase]{StratumPolygon}}}

\item{columnName}{Name of the column that will be appended. Defaults to 'Stratum'.}
}
\value{
StoxBioticData with column appended to data.table 'Station'. See \code{\link[RstoxData]{StoxBioticData}}.
}
\description{
Adds a column to StoxBioticData with the spatial strata each row belongs to.
}
\details{
The strata are added to the new column 'Stratum'.
}
