% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalyticalBaselineFunctions.R
\name{AggregateAnalyticalEstimate}
\alias{AggregateAnalyticalEstimate}
\title{Aggregate Analytical Estimate across strata}
\usage{
AggregateAnalyticalEstimate(
  AnalyticalPopulationEstimateData,
  RetainStrata = character(),
  AggregateStratumName = character()
)
}
\arguments{
\item{AnalyticalPopulationEstimateData}{\code{\link[RstoxFDA]{AnalyticalPopulationEstimateData}} with estimates to aggregate.}

\item{RetainStrata}{strata that should not be included in the aggregation. Must correspond to 'Stratum' in 'AnalyticalPopulationEstimateData'}

\item{AggregateStratumName}{name to use for the aggregated stratum}
}
\value{
\code{\link[RstoxFDA]{AnalyticalPopulationEstimateData}} with aggregated estimates.
}
\description{
Aggregates estimate across strata. One can optinally exclude some strata from being aggregated (using the argument 'RetainStrata').
Retained strata are returned unchanged.

Aggregation of means and frequencies across strata, requires information about strata size (abundance in strata).
For sampling programs where these cannot be directly estimated, they must be inferred before application of this function.
Otherwise means and frequencies will be NA. Consider for example ratio-estimation by strata (\code{\link[RstoxFDA]{AnalyticalRatioEstimate}})

Specifically, a new stratum \eqn{s'} (named according to 'AggregateStratumName') is defined based on the strata in \eqn{S}, where \eqn{S} contains all strata in 'AnalyticalPopulationEstimateData',
except those explicitly excluded by the argument 'RetainStrata'. The estimated parameters are obtained as follows:
\describe{
\item{Abundance:}{
The estimate of the total number of individuals in domain \eqn{d} and stratum \eqn{s'}:
\deqn{\hat{N}^{(s',d)} = \sum_{s \in S} \hat{N}^{(s,d)}}
with co-variance:
\deqn{\widehat{CoVar}(\hat{N}^{(s',d_{1})}, \hat{N}^{(s',d_{2})}) = \sum_{s \in S} \widehat{CoVar}(\hat{N}^{(s,d_{1})}, \hat{N}^{(s,d_{2})})}

where \eqn{\hat{N}^{(s,d)}} and \eqn{\widehat{CoVar}(\hat{N}^{(s,d_{1})}, \hat{N}^{(s,d_{2})})} are provided by the argument 'AnalyticalPopulationEstimateData'.
}
\item{Frequency:}{
The estimate of the fraction of individuals in stratum \eqn{s'} that are in domain \eqn{d}:
\deqn{ \hat{f}^{(s',d)} = \frac{\hat{N}^{(s',d)}}{\hat{N}^{(s')}} }
with co-variance:
\deqn{ \widehat{CoVar}(\hat{f}^{(s',d_{1})}, \hat{f}^{(s',d_{2})}) = \frac{1}{(\hat{N}^{(s')})^{2}}\widehat{CoVar}(\hat{N}^{(s',d_{1})}, \hat{N}^{(s',d_{2})})}

where \eqn{\hat{N}^{(s')}} is the total estimated abundance in stratum, across all domains.
}
\item{Total:}{
The estimate of the total value of some variable \eqn{v} in domain \eqn{d} and stratum \eqn{s'}.
\deqn{\hat{t}^{(s',d,v)}=\sum_{s \in S}\hat{t}^{(s,d,v)}}
with co-variance:
\deqn{\widehat{CoVar}(\hat{t}^{(s',d_{1},v_{1})}, \hat{t}^{(s',d_{2},v_{2})}) = \sum_{s \in S} \widehat{CoVar}(\hat{t}^{(s,d_{1},v_{1})}, \hat{t}^{(s,d_{2},v_{2})})}

where \eqn{\hat{t}^{(s,d,v)}} and \eqn{\widehat{CoVar}(\hat{t}^{(s,d_{1},v_{1})}, \hat{t}^{(s,d_{2},v_{2})})} are provided by the argument 'AnalyticalPopulationEstimateData'.
}

\item{Mean:}{
The estimate of the mean value of some variable \eqn{v} in domain \eqn{d} and stratum \eqn{s'}:
\deqn{\hat{\mu}^{(s',d,v)} = \frac{1}{\hat{N}^{(s',d)}}\sum_{s \in S} \hat{\mu}^{(s,d,v)}\hat{N}^{(s,d)}}
with co-variance:
\deqn{\widehat{CoVar}(\hat{\mu}^{(s',d_{1},v_{1})}, \hat{\mu}^{(s',d_{2},v_{2})}) = \sum_{s \in S} \widehat{CoVar}(\hat{\mu}^{(s,d_{1},v_{1})}, \hat{\mu}^{(s,d_{2},v_{2})}) z^{(s, d_{1}, d_{2})}}
where
\deqn{z^{(s, d_{1}, d_{2})}=\frac{\hat{N}^{(s,d_{1})}\hat{N}^{(s,d_{2})}}{\hat{N}^{(s',d_{1})}\hat{N}^{(s',d_{2})}}}
}
}
}
\concept{Analytical estimation}
