% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{AnalyticalPopulationEstimateData}
\alias{AnalyticalPopulationEstimateData}
\title{Analytical Population Estimate Data}
\description{
Analytical estimates for each PSU
 
 Abundance
 \describe{
   \item{Stratum}{}
   \item{Domain}{}
   \item{PSU}{}
   \item{Total}{}
   \item{TotalVar}{}
   \item{Frequency}{}
   \item{FrequencyVar}{}
 }
 
 Variables
 \describe{
   \item{Stratum}{}
   \item{Domain}{}
   \item{PSU}{}
   \item{Variable}{}
   \item{Total}{}
   \item{TotalVar}{}
   \item{Mean}{}
   \item{MeanVar}{}
 }
 
 StratificationVariables
 \describe{
  \item{Stratum}{}
  \item{StratificationVariables}{}
 }
 
 DomainVariables
 \describe{
  \item{Stratum}{}
  \item{Domain}{}
  \item{DomainVariables}{}
 }
}
\concept{Analytical estimation}
\concept{Data types}
