% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{AnalyticalPSUEstimateData}
\alias{AnalyticalPSUEstimateData}
\title{Analytical PSU Estimate Data}
\description{
Analytical estimates for each PSU
 
 Abundance
 \describe{
   \item{Stratum}{Identifier for stratification of individuals}
   \item{Domain}{Identifier of domains for individuals}
   \item{SampleId}{Identgier for Primary sampling unit}
   \item{Abundance}{Total number in Stratum and domain. Unsampled strata is reported as NA}
   \item{Frequency}{Frequency in Domain within Stratum. Unsampled strata is reported as NA}
 }
 
 Variables
 \describe{
   \item{SampleId}{}
   \item{Stratum}{}
   \item{Domain}{}
   \item{Variable}{}
   \item{Total}{. Unsampled strata is reported as NA}
   \item{Mean}{. Unsampled strata is reported as NA}
 }
 
 DomainVariables
 \describe{
  \item{Domain}{}
  \item{<DomainVariables>}{}
 }
 
 StratificationVariables
 \describe{
  \item{Stratum}{}
  \item{<StratificationVariables>}{}
 }
}
\concept{Analytical estimation}
\concept{Data types}
