% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AppendGearStoxBiotic}
\alias{AppendGearStoxBiotic}
\title{Append Gear to StoxBioticData}
\usage{
AppendGearStoxBiotic(
  StoxBioticData,
  UnifiedVariableDefinition,
  columnName = "UnifiedGear"
)
}
\arguments{
\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} data which will be annotated.}

\item{UnifiedVariableDefinition}{\code{\link[RstoxFDA]{UnifiedVariableDefinition}} unified gear definition.}

\item{columnName}{character(), defaults to 'UnifiedGear', name of the appended column.}
}
\value{
StoxBioticData with column appended. See \code{\link[RstoxData]{StoxBioticData}}.
}
\description{
StoX function
 Appends a column to StoxBioticData with a unified gear definition
 that are also defined for for other formats, such as StoxLandingData.
}
