% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AppendGearStoxLanding}
\alias{AppendGearStoxLanding}
\title{Append Gear to StoxLandingData}
\usage{
AppendGearStoxLanding(
  StoxLandingData,
  UnifiedVariableDefinition,
  columnName = "UnifiedGear"
)
}
\arguments{
\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} data which will be annotated.}

\item{UnifiedVariableDefinition}{\code{\link[RstoxFDA]{UnifiedVariableDefinition}} unified gear definition.}

\item{columnName}{character(), defaults to 'UnifiedGear', name of the appended column.}
}
\value{
StoxLandingData with column appended. See \code{\link[RstoxData]{StoxLandingData}}.
}
\description{
Appends a column to StoxLandingData with a unified gear definition
 that are also defined for for other formats, such as StoxBioticData.
}
