% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metierAnnotation.R
\name{appendMetier}
\alias{appendMetier}
\title{Annotate metier}
\usage{
appendMetier(
  data,
  metiertable,
  gearColumn,
  targetColumn = NULL,
  meshSizeColumn = NULL,
  selectivityDeviceColumn = NULL,
  selectivityDeviceMeshSizeColumn = NULL,
  metierColName = "metier"
)
}
\arguments{
\item{data}{\code{\link[data.table]{data.table}} with data to be annotated}

\item{metiertable}{\code{\link[RstoxFDA]{MetierTable}} with metier definitions, or object that can be read with \code{\link[RstoxFDA]{readMetierTable}}}

\item{gearColumn}{character() identifies the column in 'data' that encodes gear. Gear definition must match metiertable$gearcode}

\item{targetColumn}{character(), optional, identifies the column in 'data' that encodes target species. Definition must match metiertable$target}

\item{meshSizeColumn}{integer(), optional, identifies the column in 'data' that encodes the mesh size of the gear.}

\item{selectivityDeviceColumn}{character(), optional, identifies the column in 'data' that encodes selectivitydevices mounted on gear. Definition must mathc metiertable$selectivityDevice}

\item{selectivityDeviceMeshSizeColumn}{integer(), optional, identifies the column in 'data' that encodes the mesh size of any mounted selectivity device.}

\item{metierColName}{character() name of the column that should be appended to 'data'}
}
\value{
\code{\link[data.table]{data.table}} 'data' with the column 'metierColName' appended (character).
}
\description{
annotates metier to data table
}
\details{
Annotates data with metier, given a metier definition (\code{\link[RstoxFDA]{MetierTable}}) and a selection of data-columns that idenitfies metier-defining variables.
 Metier names are defined by the column 'metier' in 'metiertable', and does not have to reflect the information actually used for annotation.
 For example, metiers of the type "OTB_DEF", indicating both gear and target assemblage, can be configured based only on gear codes.

 For the metier-defining variables 'gearcode', 'target' and 'selectivityDevice' (see \code{\link[RstoxFDA]{MetierTable}}), missing values are matches with missing values.
 That is, a metier that is defined with a missing value for 'target'
 will be annotated to otherwise matching data which has a missing value in the 'targetColumn'.
}
\examples{
 data(metier4table)
 data(activityCensus)

 # annotate metier lvl 4 on the cencus based on gear, and compare with finer gear declaration.
 annotated <- appendMetier(activityCensus,
          metier4table,
          "gearNS",
          metierColName = "metier4")
 table(annotated$metier4)

 data(metier5table)
 # annotate metier lvl 5 on COD-catches based on only gear,
 # and compare with declarations for shrimp fisheries
 annotated <- appendMetier(activityCensus[activityCensus$species=="COD"],
          metier5table,
          "gearNS",
          metierColName = "metier5")
 annotatedShrimp <- annotated[annotated$targetFAO \%in\% c("PAN", "PRA"),]
 table(paste(annotatedShrimp$gearFAO, annotatedShrimp$targetFAO, sep="/"),
      annotatedShrimp$metier5)

 data(metier6table)
 # annotate metier lvl 6 on COD-catches based on gear and mesh size,
 # and compare with lvl 5 annotations from last example
 annotated <- appendMetier(activityCensus[activityCensus$species=="COD"],
          metier6table,
          "gearNS",
          meshSizeColumn = "meshSize",
          metierColName = "metier6")
 annotated <- appendMetier(annotated,
           metier5table,
           "gearNS",
           metierColName = "metier5")
 annotatedShrimp <- annotated[annotated$targetFAO \%in\% c("PAN", "PRA"),]
 table(annotatedShrimp$metier5, annotatedShrimp$metier6)
 
}
\seealso{
Other gear coding functions: 
\code{\link{AddGearGroupStoxBiotic}()},
\code{\link{AddGearGroupStoxLanding}()},
\code{\link{GearGroupFdirTable}},
\code{\link{SetShortGearBiotic}()},
\code{\link{metier4table}},
\code{\link{metier5table}},
\code{\link{metier6table}},
\code{\link{readMetierTable}()}
}
\concept{gear coding functions}
