% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.R
\name{appendPosition}
\alias{appendPosition}
\title{Append positions}
\usage{
appendPosition(
  table,
  areaPolygons,
  areaName,
  latColName,
  lonColName,
  StratumName = "StratumName"
)
}
\arguments{
\item{table}{data.table to be annotated.}

\item{areaPolygons}{\code{\link[sf]{sf}} data.frames with area names identified in the column 'StratumName'}

\item{areaName}{name of column that identifies the area in 'table'}

\item{latColName}{name of the latitdue column to be appended to 'table'}

\item{lonColName}{name of the longitude column to be appended to 'table'}

\item{StratumName}{name of the column in 'areaPolygons' that identifies the area.}
}
\value{
'table' with the positions appended in the columns 'latColName' and 'lonColName'.
}
\description{
Appends columns with positions to a data table, based on an area code.
 
 Positions are centroids of areas. For oddly shaped (concave) areas, this can have unintended effects. 
 The centroid may for instance lie outside the area.
}
\concept{spatial coding functions}
