% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AppendPositionLanding}
\alias{AppendPositionLanding}
\title{Append position to landings data}
\usage{
AppendPositionLanding(
  StoxLandingData,
  AreaCodePosition,
  resolution = c("Area", "Location"),
  latColName = "Latitude",
  lonColName = "Longitude"
)
}
\arguments{
\item{StoxLandingData}{landing data, see \code{\link[RstoxData]{StoxLandingData}}}

\item{AreaCodePosition}{coordinates for Area and Location codes, see \code{\link[RstoxFDA]{AreaCodePosition}}}

\item{resolution}{character(), defaults to Area, specify what resolution to use: 'Area' or 'Location'. See details.}

\item{latColName}{character(), defaults to Latitude, name of the latitude column that will be appended.}

\item{lonColName}{character(), defaults to Longitude, name of the longitude column that will be appended.}
}
\value{
\code{\link[RstoxData]{StoxLandingData}} with columns for latitude and longitude appended.
}
\description{
StoX function
 Appends a position to landings data, based on Area and Location codes.
}
\details{
When 'resolution' is specified as 'Area' the midpoint of the Area will be assigned to the landings.
 When 'resolution' is specified as 'Location' the midpoint of the Location will be assigned to the landings.
}
