% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AppendStratumStoxBiotic}
\alias{AppendStratumStoxBiotic}
\title{Appends Stratum to StoxBioticData}
\usage{
AppendStratumStoxBiotic(StoxBioticData, StratumPolygon, columnName = "Stratum")
}
\arguments{
\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} data which will be annotated.}

\item{StratumPolygon}{definition of spatial strata. See \code{\link[RstoxBase]{StratumPolygon}}}

\item{columnName}{character(), defaults to 'Stratum', name of the appended column}
}
\value{
StoxBioticData with column appended. See \code{\link[RstoxData]{StoxBioticData}}.
}
\description{
StoX function
 Appends a column to StoxBioticData with the spatial strata each row belongs to
}
