% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AppendStratumStoxLanding}
\alias{AppendStratumStoxLanding}
\title{Appends Stratum to StoxLandingData}
\usage{
AppendStratumStoxLanding(
  StoxLandingData,
  StratumPolygon,
  columnName = "Stratum",
  latColumn = "Latitude",
  lonColumn = "Longitude"
)
}
\arguments{
\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} data which will be annotated. Needs postions appended. See details.}

\item{StratumPolygon}{definition of spatial strata. See \code{\link[RstoxBase]{StratumPolygon}}}

\item{columnName}{character(), defaults to 'Stratum', name of the appended column.}

\item{latColumn}{character(), defaults to 'Latitdue', identifies the column in StoxLandingData with latitudes.}

\item{lonColumn}{character(), defaults to 'Longitude', identifies the column in StoxLandingData with longitudes.}
}
\value{
StoxLandingData with column appended. See \code{\link[RstoxData]{StoxLandingData}}.
}
\description{
StoX function
 Appends a column to StoxLandingData with the spatial strata each row belongs to.
}
\details{
\code{\link[RstoxData]{StoxLandingData}} does not contain columns for positions,
 these need to be appended before calling this function, and identified with the parameters 'latColumn' and 'lonColumn'.
 \code{\link[RstoxFDA]{AppendPositionLanding}} may be used to append positions.
}
\seealso{
\code{\link[RstoxFDA]{AppendPositionLanding}} for appending positions to \code{\link[RstoxData]{StoxLandingData}}.
}
