% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AppendTemporalStoxLanding}
\alias{AppendTemporalStoxLanding}
\title{Append Temporal Categories to StoxLandingData}
\usage{
AppendTemporalStoxLanding(
  StoxLandingData,
  TemporalDefinition,
  columnName = "TemporalCategory"
)
}
\arguments{
\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} data which will be annotated.}

\item{TemporalDefinition}{\code{\link[RstoxFDA]{TemporalDefinition}} definiton of temporal category.}

\item{columnName}{character(), defaults to 'TemporalCategory', name of the appended column.}
}
\value{
StoxLandingData with column appended. See \code{\link[RstoxData]{StoxLandingData}}.
}
\description{
StoX function
 Appends a column to StoxLandingData with a temporal category, such as 'quarter',
 that are also defined for for other formats, such as StoxBioticData.
}
