% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbookAnalyses.R
\name{appendTripIdLogbooks}
\alias{appendTripIdLogbooks}
\title{Assing trip ID to logbooks}
\usage{
appendTripIdLogbooks(
  logbooks,
  tripIds,
  timeCol = "STARTTIDSPUNKT",
  vesselIdCol = "RC",
  tripIdCol = "tripid",
  verbose = T
)
}
\arguments{
\item{logbooks}{\code{\link[data.table]{data.table}} containing logbook records}

\item{tripIds}{\code{\link[data.table]{data.table}} with columns 'vesselId', 'time' and 'tripId'}

\item{timeCol}{character() that identifies a POSIXct column in 'logbooks' that contain the time of operation for each logbook record. Default compatible with \code{\link[RstoxData]{readErsFile}}.}

\item{vesselIdCol}{character() that identifies a column in 'logbooks' that contain the vessel id (e.g. radio call signal) of the reporting vessel. Default compatible with \code{\link[RstoxData]{readErsFile}}.}

\item{tripIdCol}{character() that identifies the column name to append to 'logbooks'}

\item{verbose}{logical() wheter to message progress}
}
\value{
'logbooks' with the column identified by 'tripIDcol' added
}
\description{
Assings a trip ID to logbooks based on vessel identifier and a known last time of operation on a trip.
}
\details{
'tripid' need not be complete in the sense that vessels not found in 'tripid' will be assigned a trip ID of NA,
 and any fishing operations after the last trip for a vessel will be assinged a trip ID of NA.
 This function will however assign all the fishing operations of a vessel preceeding the first trip time to the first trip.

 Runnning time is proportional to the number of trips in 'logbooks', so it is advantagous to annotate after other relevant filtering.
}
\examples{
 \dontrun{
   #make trip ids from landings and assign them to logbooks
   lssfile <- "" #set appropriately
   logbfile <- "" #set appropriately
   landings <- RstoxData::readLssFile(lssfile)
   logbooks <- RstoxData::readErsFile(logbfile)
   tripIds <- makeTripIds(landings)
   logbooksWtripIds <- appendTripIdLogbooks(logbooks, tripIds)
 }
}
\seealso{
Other logbook functions: 
\code{\link{addLogbookColumns}()},
\code{\link{appendTripIdLandings}()},
\code{\link{calculateLogbookPartitionByTrip}()},
\code{\link{convertToErsData}()},
\code{\link{imputeCatchesLandings}()},
\code{\link{logbookAdjustment}()},
\code{\link{makeTripIds}()},
\code{\link{readLstFile}()},
\code{\link{sourceLogbookColumns}()}
}
\concept{logbook functions}
