% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxFunctions.R
\name{AreaCodePosition}
\alias{AreaCodePosition}
\title{Area Code Positions (AreaCodePosition)}
\description{
Table (\code{\link[data.table]{data.table}}) defining a position for area codes.
}
\details{
\describe{
  \item{Area}{Area code. (key)}
  \item{Location}{optional subdivision of 'Area'}
  \item{Latitude}{WGS84 Latitude, decimal degrees}
  \item{Longitude}{WGS84 Longitude, decimal degrees}
 }
 If location is provided, the case for missing location is also encoded.
}
