% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metierAnnotation.R
\name{assignMetier}
\alias{assignMetier}
\title{Annotate metier}
\usage{
assignMetier(
  data,
  metiertable,
  gearColumn,
  targetColumn = NULL,
  meshSizeColumn = NULL,
  selectivityDeviceColumn = NULL,
  selectivityDeviceMeshSizeColumn = NULL,
  metierColName = "metier"
)
}
\arguments{
\item{data}{\code{\link[data.table]{data.table}} with data to be annotated}

\item{metiertable}{\code{\link[RstoxFDA]{MetierTable}} with metier definitions, or object that can be read with \code{\link[RstoxFDA]{readMetierTable}}}

\item{gearColumn}{character() identifies the column in 'data' that encodes gear. Gear definition must match metiertable$gearcode}

\item{targetColumn}{character(), optional, identifies the column in 'data' that encodes target species. Definition must match metiertable$target}

\item{meshSizeColumn}{integer(), optional, identifies the column in 'data' that encodes the mesh size of the gear.}

\item{selectivityDeviceColumn}{character(), optional, identifies the column in 'data' that encodes selectivitydevices mounted on gear. Definition must mathc metiertable$selectivityDevice}

\item{selectivityDeviceMeshSizeColumn}{integer(), optional, identifies the column in 'data' that encodes the mesh size of any mounted selectivity device.}

\item{metierColName}{character() name of the column that should be appended to 'data'}
}
\value{
\code{\link[data.table]{data.table}} 'data' with the column 'metierColName' appended (character).
}
\description{
annotates metier to data table
}
\details{
For the metier-defining variables 'gearcode', 'target' and 'selectivityDevice' (see \code{\link[RstoxFDA]{MetierTable}}), missing values are matches with missing values.
 That is, a metier that is defined with a missing value for 'target'
 will be annotated to otherwise matching data which has a missing value in the 'targetColumn'.
}
