% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalyticalBaselineFunctions.R
\name{AssignPSUSamplingParameters}
\alias{AssignPSUSamplingParameters}
\title{Assign PSU Sampling Parameters}
\usage{
AssignPSUSamplingParameters(
  PSUSamplingParametersData,
  StoxBioticData,
  DataRecordsId,
  DefinitionMethod = c("MissingAtRandom")
)
}
\arguments{
\item{PSUSamplingParametersData}{~\code{\link[RstoxFDA]{PSUSamplingParametersData}} with sampling parameters for PSU selection}

\item{StoxBioticData}{~\code{\link[RstoxData]{StoxBioticData}} with data records for responding PSUs.}

\item{DataRecordsId}{name of Variable in ~\code{\link[RstoxData]{StoxBioticData}} that represent records of sampled PSUs}

\item{DefinitionMethod}{The method for dealing with non-response, e.g. 'MissingAtRandon'}
}
\value{
~\code{\link[RstoxFDA]{PSUSamplingParametersData}}
}
\description{
Assigns data records to PSU Sampling Parameters and provides non-response adjustments for
selected PSUs that was not sampled.
}
\details{
Some sampling parameters provided in ~\code{\link[RstoxFDA]{PSUSamplingParametersData}} are only
interpretable for sampling with complete response. This function adjusts these parameters, removes non-respondents from the
~\code{\link[RstoxFDA]{PSUSamplingParametersData}}, and checks that all responding PSUs are present in data records.

If any respondants (rows of the SelectionTable of PSUSamplingParametersData that does not have NA for SamplingUnitId) are not
found in 'SamplingUnitId', execution halts with an error.

Response after selection can generally be considered a process that modifies the sampling parameters that are set by design.
Typically sample size, InclusionProbabilities and normalized SamplingWeights need to be adjusted as non-respondents are removed,
since these are depend of the entire sample, not just the sampling unit they are assigned to.
SelectionProbabilites are by definition set for a single draw of a single sampling unit from the population and are valid even
when response is not complete.

Treatment of non-response requires some assumption about systematic differences between respondents and non-respondents.
These assumptions are specified via the argument 'DefinitionMethod' and the following options are available:
\describe{
\item{MissingAtRandom}{A response propensity is estimated for each stratum as the fraction of the sample resonding, and sample size (n) and InclusionProbability are adjusted by multiplying with this propensity. Sampling weights are adjusted by dividing them with their sum over repsondents in a stratum.}
}
}
\concept{Analytical estimation}
\concept{StoX-functions}
