% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landingsAnalyses.R
\name{assignTripIdLandings}
\alias{assignTripIdLandings}
\title{Assign trip IDs}
\usage{
assignTripIdLandings(
  landings,
  tripIds = NULL,
  vesselIdCol = "Radiokallesignal (seddel)",
  lastCatchCol = "Siste fangstdato",
  tripIdCol = "tripid"
)
}
\arguments{
\item{landings}{\code{\link[data.table]{data.table}} containing landings records}

\item{vesselIdCol}{character() that identifies a column in 'landings' that contain the vessel id (e.g. radio call signal) of the landing vessel. Default compatible with \code{\link[RstoxData]{readLssFile}}.}

\item{lastCatchCol}{character() that identifies a POSIXct column in 'landings' that contain the time of last catch for each record. Default compatible with \code{\link[RstoxData]{readLssFile}}.}

\item{\code{\link[data.table]{data.table}}}{with columns 'vesselId', 'time' and 'tripId'}

\item{tripIDcol}{character() that identifies the column name to append to 'landings'}
}
\value{
\code{\link[data.table]{data.table}} with columns 'vesselId' and 'time'
}
\description{
Assigns trip IDs to landings based on vessel identifier and date of last catch.
}
\details{
if 'tripIds' is NULL, trip IDs will be constructed from landings using \code{\link[RstoxFDA]{makeTripIds}}
}
\examples{
 \dontrun{
   #merge mesh size from first operation on a trip into landings
   lssfile <- "" #set appropriately
   logbfile <- "" #set appropriately
   landings <- RstoxData::readLssFile(lssfile)
   logbooks <- RstoxData::readErsFile(logbfile)
   tripIds <- makeTripIds(landings)
   logbooksWtripIds <- assignTripIdLogbooks(logbooks, tripIds)
   landingsWtripIds <- assignTripIdLandings(landings, tripIds)

   firstCatch <- logbooksWtripIds[!duplicated(logbooksWtripIds$tripid),]
   landingsWmeshSize <- merge(landingsWtripIds, firstCatch[,c("tripid", "MASKEVIDDE")], by="tripid", all.x=T)
 }
}
