% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbookAnalyses.R
\name{calculateLogbookPartitionByTrip}
\alias{calculateLogbookPartitionByTrip}
\title{Calcualte catch partition}
\usage{
calculateLogbookPartitionByTrip(
  logbooks,
  groupCols,
  tripCol = "tripid",
  speciesCol = "FANGSTART_FAO",
  weightCol = "RUNDVEKT"
)
}
\arguments{
\item{logbooks}{\code{\link[data.table]{data.table}} with logbooks}

\item{groupCols}{character() vector of names identifying the grouping columns in 'logbooks'}

\item{tripCol}{character() identyfing the column in 'logbooks' that identify a trip}

\item{speciesCol}{character() identyfing the column in 'logbooks' that specify species.}

\item{weigthCol}{character() identifying the column in 'logbboks' that specify the live weight of the species.}
}
\value{
\code{\link[RstoxFDA]{TripPartition}}
}
\description{
Partitions total catch of each species for each trip on provided grouping variables.
 NAs in weights are ignored (treated as zero).
}
\details{
Default parameters are compatible \code{\link[RstoxData]{readErsFile}} with tripids annotated by \code{\link[RstoxFDA]{assignTripIdLogbooks}}.

 For other partitionings of logbooks, consider \code{\link[RstoxData]{tabulateFisheries}}.
}
\examples{
 \dontrun{
  #make calculate fractions in each area for each trip
  lssfile <- "" #set appropriately
  logbfile <- "" #set appropriately
  landings <- RstoxData::readLssFile(lssfile)
  logbooks <- RstoxData::readErsFile(logbfile)
  tripIds <- makeTripIds(landings)
  logbooksWtripIds <- assignTripIdLogbooks(logbooks, tripIds)

  logbooksWtripIds$mainArea <- substring(logbooksWtripIds$LOKASJON_START,1,2)
  fractions <- calculateLogbookPartitionByTrip(logbooksWtripIds, "mainArea")
 }
}
