% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatchLotteryExample-datadoc.R
\docType{data}
\name{CatchLotteryExample}
\alias{CatchLotteryExample}
\title{Data from the Norwegian catch lottery sampling program.}
\format{
\code{\link[RstoxData]{StoxBioticData}}
}
\usage{
data(CatchLotteryExample)
}
\description{
Example of data formatted as \code{\link[RstoxData]{StoxBioticData}}
Hauls are primary sampling units, selected by Poission sampling with selection probabilities proportional to the catch size.
The data contain North Sea herring samples from catch lottery sampling in 2022.
}
\examples{
 RstoxFDA::plotArea(RstoxFDA::CatchLotteryExample$Station, 
      areaDef=RstoxFDA::mainareaFdir2018, 
      latCol = "Latitude", 
      lonCol = "Longitude", 
      areaLabels = TRUE)
}
\concept{Analytical estimation}
\keyword{datasets}
