% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatchLotteryLandingExample-datadoc.R
\docType{data}
\name{CatchLotteryLandingExample}
\alias{CatchLotteryLandingExample}
\title{Landings for the Norwegian North Sea Herring fisheries in 2022.}
\format{
\code{\link[RstoxData]{StoxLandingData}}
}
\usage{
data(CatchLotteryLandingsExample)
}
\description{
Example of data formatted as \code{\link[RstoxData]{StoxLandingData}}
The data set contains landings of herring caugth in 2022 and sold as North Sea Herring, from Norwegian vessels larger than 15 m.
}
\details{
Samples from the Catch Lottery covering this fishery are provided in \code{\link[RstoxFDA]{CatchLotteryExample}}
}
\examples{
 #report destination (country landed) of landed catch in tonnes
 RstoxFDA::CatchLotteryLandingExample$Landing[,list(weightT=sum(RoundWeight)/1000), by=c("CountryLanding")]
}
\concept{Analytical estimation}
\keyword{datasets}
