% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatchLotterySamplingExample-datadoc.R
\docType{data}
\name{CatchLotterySamplingExample}
\alias{CatchLotterySamplingExample}
\title{Sampling parameters from the Norwegian catch lottery sampling program.}
\format{
\code{\link[RstoxFDA]{PSUSamplingParametersData}}
}
\usage{
data(CatchLotterySamplingExample)
}
\description{
Example of data formatted as \code{\link[RstoxFDA]{PSUSamplingParametersData}}
Hauls are primary sampling units, selected by Poission sampling with selection probabilities proportional to the catch size.
The data contain sampling parameters for North Sea herring samples from catch lottery sampling in 2022.
}
\details{
The corresponding samples are provided in \code{\link[RstoxFDA]{CatchLotteryExample}}
}
\examples{
 #all selected PSU that where actuall sampled are provided in CatchLotteryExample
 sum(!is.na(CatchLotterySamplingExample$SelectionTable$SamplingUnitId))
 sum(CatchLotterySamplingExample$SelectionTable$SamplingUnitId \%in\%
       CatchLotteryExample$Haul$HaulKey)
}
\concept{Analytical estimation}
\keyword{datasets}
