% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxFunctions.R
\name{ClassificationError}
\alias{ClassificationError}
\title{Stock classification error (ClassificationError)}
\description{
Table (\code{\link[data.table]{data.table}})
defining probabilities of misclassifying stock membership of a fish (e.g. from otholith).
}
\details{
The stock classification system is designed for coastal and atlantic cod as they are determined at IMR.
The two stock of interest are classified by code 1, and 5.
Code 2 signifies the same as code 1, but indicate less certainty.
Code 4 signifies the same as code 5, but indicate less certainty.

The classification error specifies the probability of misclassifying between some of these classifications.


 \describe{
  \item{ptype1.CC}{numeric() [0,1]. Probability of classifying a type 1 specimen as type 1.}
  \item{ptype1.S}{numeric() [0,1]. Probability of classifying a type 5 specimen as type 1.}
  \item{ptype2.CC}{numeric() [0,1]. Probability of classifying a type 2 specimen as type 2.}
  \item{ptype2.S}{numeric() [0,1]. Probability of classifying a type 4 specimen as type 2.}
  \item{ptype4.CC}{numeric() [0,1]. Probability of classifying a type 2 specimen as type 4.}
  \item{ptype4.S}{numeric() [0,1]. Probability of classifying a type 4 specimen as type 4.}
  \item{ptype5.CC}{numeric() [0,1]. Probability of classifying a type 1 specimen as type 5.}
  \item{ptype5.S}{numeric() [0,1]. Probability of classifying a type 5 specimen as type 5.}
 }

 The data table contains only one row
}
