% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalyticalBaselineFunctions.R
\name{CollapseStrata}
\alias{CollapseStrata}
\title{Collapse strata in sampling design}
\usage{
CollapseStrata(IndividualSamplingParametersData, RetainStrata = character())
}
\arguments{
\item{IndividualSamplingParametersData}{\code{\link[RstoxFDA]{IndividualSamplingParametersData}} with sampling parameters for sample selection}

\item{RetainStrata}{character() with the names of stratification variables to retain. Stratification variables not specified here will be removed.}
}
\value{
\code{\link[RstoxFDA]{IndividualSamplingParametersData}} with simplified stratification
}
\description{
Transforms a sampling design to an equivalent one with simpler stratification
}
\details{
The sampling information in \code{\link[RstoxFDA]{IndividualSamplingParametersData}} allows for specification of several stratification variables.
This function facilitates removal of some of these stratificaiton vairables, redefining strata to a coarser stratification scheme that still has known
stratum sizes.

If all stratification variables are removed, all samples will be assigned to a single stratum named 'All'.
}
\concept{Analytical estimation}
