% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{ConvertLengthBiotic}
\alias{ConvertLengthBiotic}
\title{Convert lengths Biotic}
\usage{
ConvertLengthBiotic(
  BioticData,
  LengthConversionTable,
  TargetLengthMeasurement = character()
)
}
\arguments{
\item{BioticData}{\code{\link[RstoxData]{BioticData}} for which lengths should be converted}

\item{LengthConversionTable}{\code{\link[RstoxFDA]{LengthConversionTable}} with parameters for converting lengths to 'TargetLengthMeasurement'.}

\item{TargetLengthMeasurement}{The desired length measurement. Typically the code for 'total length'.}
}
\value{
\code{\link[RstoxData]{BioticData}} with converted lengths.
}
\description{
Convert lengths to approximate values for a desired length measurement. Typically 'total length'.
}
\details{
Different length measurements may be defined for fish, such as 'total length', 'standard length' or 'fork length'.
For partially processed fish, only some measurements may be available, such as 'head-length'.
When desired length measurment is not measured,
'total length' or some other length measurement may be approximated by a regression model.

For records where the variable 'catchcategory' on the table 'catchsample' matches the 'Species' variable in 'LengthConversionTable'
this function converts lengths ('length' on the 'Individual'-table) to desired measurement type ('TargetLengthMeasurement')
by matching the 'lengthmeasurement' variable on the corresponding 'catchsample'-table to 'MeasurementType' in 'LengthConversionTable'
and applying the formula L_{desired} = 0.01 \* Alpha + Beta \* L_{MeasurementType}.
The factor of 0.01 is applied because lengths in Biotic is defined i m, while 'Alpha' in 'LengthConversionTable' is defined in cm.

After conversion the 'lengthmeasurement'-variable will be changed to reflect the converted length.
That is, they will be set to 'TargetLengthMeasurement'.

These variables are on these tables in data originating from NMDBiotic (http://www.imr.no/formats/nmdbiotic/).
For bioticdata that does not conform to this, no modifications are done.
}
\seealso{
\code{\link[RstoxFDA]{DefineLengthConversionParameters}} for configuration of length parameters.

Other nmdbiotic functions: 
\code{\link{ConvertWeightBiotic}()},
\code{\link{ListBioticDifference}()},
\code{\link{SetAreaPositionsBiotic}()},
\code{\link{SetShortGearBiotic}()},
\code{\link{SetStartDateBiotic}()},
\code{\link{SetTimeBiotic}()}

Other parameter conversion functions: 
\code{\link{ConvertWeightBiotic}()},
\code{\link{DefineLengthConversionParameters}()},
\code{\link{DefineWeightConversionFactor}()}
}
\concept{nmdbiotic functions}
\concept{parameter conversion functions}
