% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecaOuputputConversion.R
\name{convertRecaData}
\alias{convertRecaData}
\title{Convert RecaData}
\usage{
convertRecaData(
  RecaData,
  nSamples = as.integer(NA),
  burnin = as.integer(NA),
  thin = as.integer(NA),
  resultdir = as.character(NA),
  lgamodel = as.character(NA),
  delta.age = as.numeric(NA),
  seed = as.numeric(NA),
  fitfile = as.character(NA),
  predictfile = as.character(NA)
)
}
\arguments{
\item{RecaData}{\code{\link[RstoxFDA]{RecaData}}, as prepared by \code{\link[RstoxFDA]{PrepareRecaEstimate}}.}

\item{nSamples}{nSamples as specified in \code{\link[Reca]{eca.estimate}}}

\item{burnin}{burnin as specified in \code{\link[Reca]{eca.estimate}}}

\item{thin}{thin as specified in \code{\link[Reca]{eca.estimate}}}

\item{resultdir}{resultdir as specified in \code{\link[Reca]{eca.estimate}}}

\item{lgamodel}{lgamodel as specified in \code{\link[Reca]{eca.estimate}}}

\item{delta.age}{delta.age as specified in \code{\link[Reca]{eca.estimate}}}

\item{seed}{seed as specified in \code{\link[Reca]{eca.estimate}}}

\item{fitfile}{fitfile as specified in \code{\link[Reca]{eca.estimate}}}

\item{predictfile}{predictfile as specified in \code{\link[Reca]{eca.estimate}}}
}
\value{
list with 5 members:
 \describe{
  \item{AgeLength}{Data prepared for the argument AgeLength to \code{\link[Reca]{eca.estimate}}}
  \item{WeightLength}{Data prepared for the argument WeightLength to \code{\link[Reca]{eca.estimate}}}
  \item{Landings}{Data prepared for the argument Landings to \code{\link[Reca]{eca.estimate}}}
  \item{GlobalParameters}{Data prepared for the argument GlobalParameters to \code{\link[Reca]{eca.estimate}}, Completeness of this data structure depend on whether optional parameters where provided.}
  \item{CovariateMaps}{Not input to Reca. Nested list, providing mapping between integer encoding and character encoding of the levels of categorical variables.}
 }
}
\description{
Converts \code{\link[RstoxFDA]{RecaData}} prepared for StoX-Reca to the input format accepted by \code{\link[Reca]{eca.estimate}}.
 This facilitates preparing input data with a StoX-project for more cusomized programs directly interfacing Reca.
}
\details{
The StoX-function interfacing Reca-parameterisation (\code{\link[RstoxFDA]{RunRecaModels}}) accepts a data format
 \code{\link[RstoxFDA]{RecaData}} that is not exactly compatible with the format accepted by \code{\link[Reca]{eca.estimate}}.
 This inconvenience is introduced to make sure that output from \code{\link[RstoxFDA]{PrepareRecaEstimate}} adheres to
 restrictions in the StoX-user interface, so that the results are inspectable there.
 
 Reca-input include model parameters and options, but \code{\link[RstoxFDA]{PrepareRecaEstimate}} only sets parameters that have
 implications for data formatting, other parameters are subsequently set in for instance \code{\link[RstoxFDA]{RunRecaModels}}. 
 The optional arguments to this function can be used to set those parameters that is handled by \code{\link[RstoxFDA]{RunRecaModels}} in the StoX-workflow.
 
 Reca requires categorical variables to be encoded as consecutive integers, starting with 1. In order to keep track of how this correspond to other representations, 
 this function provides the mapping in the list 'CovariateMaps'. CovariateMaps is not an input to Reca, but only provided to assist book-keeping.
}
\concept{Reca functions}
