% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{convertToErsData}
\alias{convertToErsData}
\title{Convert logbook data}
\usage{
convertToErsData(LstLogbookData, timestring = "12:00:00")
}
\arguments{
\item{LstLogbookData}{\code{\link[RstoxFDA]{LstLogbookData}} to be converted}

\item{timestring}{string representing the time of day (UTC) to assume for fishing operations. Format: \%H:\%M:\%S.}
}
\value{
\code{\link[data.table]{data.table}} formatted as return from \code{\link[RstoxData]{readErsFile}}.
}
\description{
Convert \code{\link[RstoxFDA]{LstLogbookData}} to the format provided by \code{\link[RstoxData]{readErsFile}}.
Not all records can be filled, and some are left blank.

\code{\link[RstoxFDA]{LstLogbookData}} contains records for fishing date, but not exact time, which is required for conversion of time information.
In order to set start time of fishing operations, the a time of day need to be provided via the parameter 'timestring'
UTC times are assumed.
}
\seealso{
Other logbook functions: 
\code{\link{activityCensus}},
\code{\link{addLogbookColumns}()},
\code{\link{appendTripIdLandings}()},
\code{\link{appendTripIdLogbooks}()},
\code{\link{calculateLogbookPartitionByTrip}()},
\code{\link{imputeCatchesLandings}()},
\code{\link{logbookAdjustment}()},
\code{\link{makeTripIds}()},
\code{\link{readLstFile}()},
\code{\link{sourceLogbookColumns}()}

Other IO functions: 
\code{\link{ReadLandingFDA}()},
\code{\link{readFdirLandingsArchive}()},
\code{\link{readLstFile}()}
}
\concept{IO functions}
\concept{logbook functions}
