% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{ConvertWeightsBiotic}
\alias{ConvertWeightsBiotic}
\title{Convert weights Biotic}
\usage{
ConvertWeightsBiotic(
  BioticData,
  WeightConversionTable,
  TargetProductType = character()
)
}
\arguments{
\item{BioticData}{\code{\link[RstoxData]{BioticData}} for which weights should be converted}

\item{WeightConversionTable}{\code{\link[RstoxFDA]{WeightConversionTable}} with parameters for converting weights to 'TargetProductType'.}

\item{TargetProductType}{The desired producttype. Typically the code for Round fish.}
}
\value{
\code{\link[RstoxData]{BioticData}} with converted weights.
}
\description{
Convert weights to round weights.
}
\details{
Different products of fish may be measured and round weight (live weight)
 or some other product type may be approximated by multiplying with a scalar conversion factor.
 
 For records where the variable 'catchcategory' on the table 'catchsample' matches the 'Species' variable in 'Translation'
 this function converts weights to desired product type ('TargetProductType') by matching 'producttype'-variables to 'ProductType' in 'Translation'.
 The following weights are converted:
 \itemize{
  \item{'catchweight' based on 'catchproducttype' both on the table 'catchsample'}
  \item{'lengthsampleweight' based on 'sampleproducttype' both on the table 'catchsample'}
  \item{'individualweight' based on 'individualproducttype' both on the table 'individual'} 
 }
 
 After conversion the 'producttype'-variables will be changed to reflect the approximate weight.
 That is they will be set to 'TargetProductType'
 
 If 'WeightConversionTable' contains NAs for 'WeightFactor', converted weights will be NA.
 This is useful for dealing with species whoose weights are not needed in subsequent analysis. 
 
 For weights that are missing (NA), no conversion will be attempted.
 
 The variables 'catchvolume', 'lengthsamplevolume', and 'individualvolume'
 are also dependent on product type. These are set to NA as weight is converted.
 
 These variables are on these tables in data originating from NMDBiotic (http://www.imr.no/formats/nmdbiotic/).
 For bioticdata that does not conform to this, no modifications are done.
}
