% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{DefineAreaCodePosition}
\alias{DefineAreaCodePosition}
\title{Define Area Code Positions}
\usage{
DefineAreaCodePosition(
  processData,
  resourceFilePath,
  encoding = "UTF-8",
  useProcessData = F
)
}
\arguments{
\item{processData}{data.table() as returned from this function}

\item{resourceFilePath}{path to resource file}

\item{encoding}{encoding of resource file}

\item{useProcessData}{logical() Bypasses execution of function, and simply returns argument 'processData'}
}
\value{
\code{\link[RstoxFDA]{AreaCodePosition}}.
}
\description{
StoX function
 Define positions for areas of a spatial coding system.
 Definitions are read from a resource file.
}
\details{
Definitions are read from a tab separated file with headers. Columns defined as:
 \describe{
 \item{Column 1: 'Area'}{Area code (key)}
 \item{Column 2: 'Location'}{optional subdivision of area. If provided, encode the case for missing location should be encoded as well.}
 \item{Column 3: 'Latitude'}{WGS84 Latitude, decimal degress}
 \item{Column 4: 'Longitude'}{WGS84 Longitude, decimal degress}
 }
}
