% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{DefineAreaPosition}
\alias{DefineAreaPosition}
\title{Define Area Code Positions}
\usage{
DefineAreaPosition(
  processData,
  DefinitionMethod = c("ResourceFile", "StratumPolygon"),
  FileName = NULL,
  StratumPolygon = NULL,
  UseProcessData = F
)
}
\arguments{
\item{processData}{data.table() as returned from this function}

\item{DefinitionMethod}{'ResourceFile' or 'StratumPolygon', see details.}

\item{FileName}{path to resource file}

\item{StratumPolygon}{\code{\link[RstoxBase]{StratumPolygon}} to extract area positions from}

\item{UseProcessData}{logical() Bypasses execution of function, and simply returns argument 'processData'}
}
\value{
\code{\link[RstoxFDA]{AreaPosition}}.
}
\description{
StoX function
 Define positions for areas of a spatial coding system.
}
\details{
For DefinitionMethod ResourceFile:
 Definitions are read from a tab separated, UTF-8 encoded file with headers. Columns defined as:
 \describe{
 \item{Column 1: 'Area'}{Area code (key)}
 \item{Column 2: 'Location'}{optional subdivision of area. If provided, positions for missing locations should be encoded as well.}
 \item{Column 3: 'Latitude'}{WGS84 Latitude, decimal degrees}
 \item{Column 4: 'Longitude'}{WGS84 Longitude, decimal degrees}
 }
 
 For DefinitionMethod StratumPolygon:
 Definitions are extracted from a \code{\link[RstoxBase]{StratumPolygon}}:
 Area is derived from the column 'polygonName' in 'StratumPolygon'.
 Location is encoded as missing.
 Latitude and Longitude are the coordinates set for each polygon in 'StratumPolygon'.
}
