% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{DefineClassificationError}
\alias{DefineClassificationError}
\title{Define Classification Error (for stock splitting)}
\usage{
DefineClassificationError(
  processData,
  resourceFilePath,
  encoding = "UTF-8",
  useProcessData = F
)
}
\arguments{
\item{processData}{data.table() as returned from this function}

\item{resourceFilePath}{path to resource file}

\item{encoding}{encoding of resource file}

\item{useProcessData}{logical() Bypasses execution of function, and simply returns argument 'processData'}
}
\value{
Classification Error Matrix, see: \code{\link[RstoxFDA]{ClassificationError}}.
}
\description{
StoX function.
 Defines probabilities for misclassifying when determining stock membership of a specimen.
}
\details{
Definitions are read from a tab separated file with headers. Columns defined as:
 \describe{
  \item{Column 1 : ptype1.CC}{Probability of classifying a type 1 specimen as type 1.}
  \item{Column 2: ptype1.S}{Probability of classifying a type 5 specimen as type 1.}
  \item{Column 3: ptype2.CC}{Probability of classifying a type 2 specimen as type 2.}
  \item{Column 4: ptype2.S}{Probability of classifying a type 4 specimen as type 2.}
  \item{Column 5: ptype4.CC}{Probability of classifying a type 2 specimen as type 4.}
  \item{Column 6: ptype4.S}{Probability of classifying a type 4 specimen as type 4.}
  \item{Column 7: ptype5.CC}{Probability of classifying a type 1 specimen as type 5.}
  \item{Column 8: ptype5.S}{Probability of classifying a type 5 specimen as type 5.}
 }
 see \code{\link[RstoxFDA]{ClassificationError}} for further explanation on the coding system.
}
