% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxFunctions.R
\name{DefineGear}
\alias{DefineGear}
\title{Define gear}
\usage{
DefineGear(processData, resourceFilePath, encoding = "latin1",
  useProcessData = F)
}
\arguments{
\item{processData}{data.table() as returned from this function}

\item{resourceFilePath}{path to resource file}

\item{encoding}{encoding of resource file}

\item{useProcessData}{logical() Bypasses execution of function, and simply returns argument 'processData'}
}
\value{
Unified variable definition, see: \code{\link[RstoxFDA]{UnifiedVariableDefinition}}.
}
\description{
Define a unified categorical variable 'Gear', and its correspondance to gear codes in
 data formats \code{\link[RstoxData]{StoxBioticData}} and \code{\link[RstoxData]{StoxLandingData}}.
 Definitions are read from a resource file.
}
\details{
Definitions are stored with in a tab separated file with headers. Columns defined as:
 \describe{
 \item{Column 1: 'UnifiedVariable'}{Unified value (key)}
 \item{Column 2: 'Source'}{The format for which the unified value is defined(key)}
 \item{Column 3: 'Definition'}{A comma separated list of values, any of which will be defined as the unified value in the format.}
 }
}
