% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{DefineLengthConversionParameters}
\alias{DefineLengthConversionParameters}
\title{Define Length Conversion Parameters}
\usage{
DefineLengthConversionParameters(
  processData,
  DefinitionMethod = c("ResourceFile"),
  FileName = character(),
  UseProcessData = F
)
}
\arguments{
\item{processData}{\code{\link[RstoxFDA]{LengthConversionTable}} as returned from this function}

\item{DefinitionMethod}{'ResourceFile'. See details.}

\item{FileName}{path to resource file}

\item{UseProcessData}{If TRUE, bypasses execution of function and returns existing 'processData'}
}
\value{
\code{\link[RstoxFDA]{LengthConversionTable}}
}
\description{
Define length conversion parameters for approximating one type of length measurement from measurments of another type,
based on a linear regression fit between measurement types. E.g. calculating 'total length' from measured 'fork length'.
}
\details{
For DefinitionMethod 'ResourceFile':
Definitions are read from a tab separated, UTF-8 encoded file with headers. Columns defined as:
\describe{
\item{Column 1: 'Description'}{Free-text description of the product}
\item{Column 2: 'Species'}{Identifier for the species that the conversion applies to}
\item{Column 3: 'MeasurmentType'}{Identifier for the type of length measurement of the measured length}
\item{Column 4: 'Alpha'}{scalar value representing the intercept (in cm) of a linear regression fit between length measurements.}
\item{Column 5: 'Beta'}{scalar value representing the slope of a linear regression fit between length measurements.}
}
See also \code{\link[RstoxFDA]{LengthConversionTable}}
}
\seealso{
\code{\link[RstoxFDA]{ConvertLengthBiotic}} for applying length conversion to data

Other parameter conversion functions: 
\code{\link{ConvertLengthBiotic}()},
\code{\link{ConvertWeightBiotic}()},
\code{\link{DefineWeightConversionFactor}()},
\code{\link{FDIR.factors.VIII.2022}},
\code{\link{convertCodes}()}

Other StoX-functions: 
\code{\link{AddAreaPositionStoxLanding}()},
\code{\link{AddGearGroupStoxBiotic}()},
\code{\link{AddGearGroupStoxLanding}()},
\code{\link{AddPeriodStoxBiotic}()},
\code{\link{AddPeriodStoxLanding}()},
\code{\link{AddStratumStoxBiotic}()},
\code{\link{AddStratumStoxLanding}()},
\code{\link{ConvertLengthBiotic}()},
\code{\link{ConvertWeightBiotic}()},
\code{\link{DefineAgeErrorMatrix}()},
\code{\link{DefineAreaPosition}()},
\code{\link{DefineCarNeighbours}()},
\code{\link{DefinePeriod}()},
\code{\link{DefineStockSplittingParameters}()},
\code{\link{DefineWeightConversionFactor}()},
\code{\link{FilterAgeLengthOutliersStoxBiotic}()},
\code{\link{FilterWeightLengthOutliersStoxBiotic}()},
\code{\link{ListBioticDifference}()},
\code{\link{LoadFdaStratumPolygon}()},
\code{\link{ParameterizeRecaModels}()},
\code{\link{PrepareRecaEstimate}()},
\code{\link{ReadLandingFDA}()},
\code{\link{ReportFdaLandings}()},
\code{\link{ReportFdaSOP}()},
\code{\link{ReportFdaSampling}()},
\code{\link{ReportParameterConvergence}()},
\code{\link{ReportRecaCatchAtAgeCovariance}()},
\code{\link{ReportRecaCatchAtAge}()},
\code{\link{ReportRecaCatchAtLengthAndAge}()},
\code{\link{ReportRecaCatchAtLength}()},
\code{\link{ReportRecaCatchStatistics}()},
\code{\link{ReportRecaLengthAtAge}()},
\code{\link{ReportRecaParameterStatistics}()},
\code{\link{ReportRecaWeightAtAge}()},
\code{\link{RunRecaModels}()},
\code{\link{SetAreaPositionsBiotic}()},
\code{\link{SetShortGearBiotic}()},
\code{\link{SetStartDateBiotic}()},
\code{\link{SetTimeBiotic}()}
}
\concept{StoX-functions}
\concept{parameter conversion functions}
