% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalyticalBaselineFunctions.R
\name{DefineMultiStageSamplingParameters}
\alias{DefineMultiStageSamplingParameters}
\title{Define Sampling Design Parameters}
\usage{
DefineMultiStageSamplingParameters(
  processData,
  DefinitionMethod = c("ResourceFile", "AdHocStoxBiotic"),
  FileName = character(),
  StoxBioticData,
  SamplingUnitId,
  StratificationColumns,
  UseProcessData = F
)
}
\arguments{
\item{processData}{\code{\link[RstoxFDA]{MultiStageSamplingParametersData}} as returned from this function.}

\item{DefinitionMethod}{'ResourceFile' or 'AdHocStoxBiotic'}

\item{FileName}{path to resource file}

\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} Sample data to construct design parameters from}

\item{SamplingUnitId}{name of column in 'StoxBioticData' that identifies the sampling unit the design is constructed for.}

\item{StratificationColumns}{name of any column (at the same table as 'SamplingUnitId') that are to be used to define Strata for sampling.}

\item{UseProcessData}{If TRUE, bypasses execution of function and returns existing 'processData'}
}
\value{
\code{\link[RstoxFDA]{MultiStageSamplingParametersData}}
}
\description{
Define sampling design parameters for intermediate sampling units in multi-stage sampling.
}
\details{
The DefintionMethod 'ResourceFile' reads design parameters from a tab delimited file with headers corresponding to those listed in
\code{\link[RstoxFDA]{MultiStageSamplingParametersData}}. The data is provided as one table, so that the information in 'sampleTable' is repeated for each entry in 'selectionTable'.
Any columns not named in \code{\link[RstoxFDA]{MultiStageSamplingParametersData}} are assumed to be stratification variables.
The conditions listed for the variables in \code{\link[RstoxFDA]{MultiStageSamplingParametersData}} are checked upon reading the data, and
execution halts with error if any are violated.

The DefinitionMethod 'AdHocStoxBiotic' constructs Sampling Design Parameters from data,
assuming equal probability sampling with fixed sample size, selection with replacement and complete response.
This is a reasonable approximation if within-strata sampling is approximately simple random selections,
non-response is believed to be at random, and only a small fraction of the strata is sampled,
so that with and without replacement sampling probabilities are approximately equal.
}
\concept{Analytical estimation}
\concept{StoX-functions}
