% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{DefinePeriod}
\alias{DefinePeriod}
\title{Define Periods}
\usage{
DefinePeriod(
  processData,
  TemporalCategory = c("Quarter", "Month", "Custom"),
  CustomPeriods = character(),
  UseProcessData = F
)
}
\arguments{
\item{processData}{\code{\link[RstoxFDA]{TemporalDefinition}} as returned from this function}

\item{TemporalCategory}{type of temporal category: 'Quarter', 'Month' or 'Custom', defaults to 'Quarter'}

\item{CustomPeriods}{provided if temporalCategory is 'Custom', vector of strings formatted as DD-MM or DD-MM-YYYY, giving the start date of each temporal category.}

\item{UseProcessData}{Bypasses execution of function, if TRUE, and simply returns argument 'ProcessData'}
}
\value{
Temporal Categories, see: \code{\link[RstoxFDA]{TemporalDefinition}}.
}
\description{
Define periods for grouping data based on date.
}
\details{
The 'TemporalCategory'-options 'Quarter' and 'Month' produce seasonal definitions.
 Seasonal definitions include dates
 based on day and month, irrespective of year. Seasonal definitions also wrap around so that 
 January is considered to follow December. 
 
 In order to make custom seasonal definitions,
 use the TemporalCategory'-option 'Custom' without providing years in 'CustomPeriods': "DD-MM".
 In order to make non-seasonal definitions for Quarter or Month, provide the as 'CustomPeriods'
 for all years of interest. If years are provided categories are automatically extended to the entire year, if necesesarry.
 That is a category starting with 1. Jan is added if not present, and likewise a category ending with 31. Dec.
}
