% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalyticalBaselineFunctions.R
\name{DefinePSUCoInclusionProbabilities}
\alias{DefinePSUCoInclusionProbabilities}
\title{Define PSU Co-Inclusion Probabilities}
\usage{
DefinePSUCoInclusionProbabilities(PSUSamplingParametersData)
}
\arguments{
\item{PSUSamplingParametersData}{\code{\link[RstoxFDA]{PSUSamplingParametersData}}}
}
\value{
\code{\link[RstoxFDA]{PSUCoInclusionProbabilities}}
}
\description{
Computes co-inclusion probabilites for a selection of Primary Selection Units
}
\details{
The method for calculating co-inclusion probabilites depend on the method of selection, which is encded
in \code{\link[RstoxFDA]{PSUSamplingParametersData}}. Only selection methods 'Possion' and 'FSWR' are currently supported.
}
