% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalyticalBaselineFunctions.R
\name{DefinePSUSamplingParameters}
\alias{DefinePSUSamplingParameters}
\title{Define PSU Sampling Design Parameters}
\usage{
DefinePSUSamplingParameters(
  processData,
  DefinitionMethod = c("ResourceFile", "AdHocStoxBiotic"),
  FileName = character(),
  StoxBioticData,
  SamplingUnitId = character(),
  StratificationColumns = character(),
  UseProcessData = F
)
}
\arguments{
\item{processData}{\code{\link[RstoxFDA]{PSUSamplingParametersData}} as returned from this function.}

\item{DefinitionMethod}{'ResourceFile' or 'AdHocStoxBiotic'}

\item{FileName}{path to resource file}

\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} Sample data to construct design parameters from}

\item{SamplingUnitId}{name of column in 'StoxBioticData' that identifies the Primary Sampling Unit the design is constructed for.}

\item{StratificationColumns}{name of any column (at the same table as 'SamplingUnitId') that are to be used to define Strata for sampling.}

\item{UseProcessData}{If TRUE, bypasses execution of function and returns existing 'processData'}
}
\value{
\code{\link[RstoxFDA]{PSUSamplingParametersData}}
}
\description{
Define sampling parameters for Primary Sampling Units in multi-stage sampling.
}
\details{
The DefintionMethod 'ResourceFile' reads sampling parameters from a tab delimited file with headers corresponding to those listed in
\code{\link[RstoxFDA]{PSUSamplingParametersData}}. The data is provided as one table, so that the information in 'sampleTable' is repeated for each entry in 'selectionTable'.
Any columns not named in \code{\link[RstoxFDA]{PSUSamplingParametersData}} are assumed to be stratification variables.
The conditions listed for the variables in \code{\link[RstoxFDA]{PSUSamplingParametersData}} are checked upon reading the data, and
execution halts with error if any are violated. Consult the examples in this documentation to see how the resource is formatted
with ae stratification variable 'Species'.

The DefinitionMethod 'AdHocStoxBiotic' constructs Sampling Design Parameters from data,
assuming equal probability sampling with fixed sample size, selection with replacement and complete response.
These is a reasonable approximation if within-strata sampling is approximately simple random selections,
the sample intensitiy is low (only a small fraction of the population is sampled),
and non-response is believed to be at random.
}
\examples{
 # embedded example file:
 exampleFile <- system.file("testresources", 
                        "lotteryParameters", 
                        "lotteryDesignNSHstrata.txt", package="RstoxFDA")
 
 # Read example file with StoX
 PSUSamplingParametersData <- RstoxFDA::DefinePSUSamplingParameters(DefinitionMethod="ResourceFile", 
                          FileName=exampleFile)
 
 # Read example file as flat table, to illustrate formatting
 FlatSamplingParametersData <- read.csv(exampleFile, sep="\t")

}
\concept{Analytical estimation}
\concept{StoX-functions}
