% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalyticalBaselineFunctions.R
\name{DefineSamplingDesignParameters}
\alias{DefineSamplingDesignParameters}
\title{Define Sampling Design Parameters}
\usage{
DefineSamplingDesignParameters(
  processData,
  DefinitionMethod = c("ResourceFile", "AdHocStoxBiotic"),
  FileName = character(),
  StoxBioticData,
  SamplingUnitId,
  StratificationColumns,
  OrderColumn,
  UseProcessData = F
)
}
\arguments{
\item{processData}{\code{\link[RstoxFDA]{SamplingDesignParametersData}} as returned from this function.}

\item{DefinitionMethod}{'ResourceFile' or 'AdHocStoxBiotic'}

\item{FileName}{path to resource file}

\item{UseProcessData}{If TRUE, bypasses execution of function and returns existing 'processData'}
}
\value{
\code{\link[RstoxFDA]{SamplingDesignParametersData}}
}
\description{
Define sampling design parameters for use in analytical estimation.
}
\details{
The DefintionMethod 'ResourceFile' reads design parameters from a tab delimited file with headers corresponding to those listed in
\code{\link[RstoxFDA]{SamplingDesignParametersData}}. The data is provided as one table, so that the information in 'sampleTable' is repeated for each entry in 'selectionTable'.
Any columns not named in \code{\link[RstoxFDA]{SamplingDesignParametersData}} are assumed to be stratification variables.
The conditions listed for the variables in \code{\link[RstoxFDA]{SamplingDesignParametersData}} are checked upon reading the data, and
execution halts with error if any are violated.

The DefinitionMethod 'AdHocStoxBiotic' constructs Sampling Design Parameters from data,
assuming equal probability non-finite sampling with fixed sample size, selection without replacement and complete response.
}
\concept{Analytical estimation}
\concept{StoX-functions}
