% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalyticalBaselineFunctions.R
\name{DefineSamplingHierarchy}
\alias{DefineSamplingHierarchy}
\title{Extend IndividualSamplingParametersData to reflect selection from PSU by specifying intermediate selection.
Not yet implemented.}
\usage{
DefineSamplingHierarchy(
  IndividualSamplingParametersData,
  Hierarchy = character(),
  Stratification = character(),
  StrataSizes = character(),
  SelectionMetod = character(),
  CollapseStrata = character()
)
}
\value{
\code{\link[RstoxFDA]{IndividualSamplingParametersData}}
}
\description{
Extend IndividualSamplingParametersData to reflect selection from PSU by specifying intermediate selection.
Not yet implemented.
}
\concept{Analytical estimation}
