% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{DefineStockSplittingParameters}
\alias{DefineStockSplittingParameters}
\title{Define Stock Splitting Parameters}
\usage{
DefineStockSplittingParameters(
  processData,
  DefinitionMethod = c("ResourceFile", "FunctionParameters"),
  FileName = character(),
  StockNameCC = character(),
  StockNameS = character(),
  ProbabilityType1As1 = numeric(),
  ProbabilityType5As1 = numeric(),
  ProbabilityType2As2 = numeric(),
  ProbabilityType4As2 = numeric(),
  ProbabilityType2As4 = numeric(),
  ProbabilityType4As4 = numeric(),
  ProbabilityType1As5 = numeric(),
  ProbabilityType5As5 = numeric(),
  UseProcessData = F
)
}
\arguments{
\item{processData}{data.table() as returned from this function}

\item{DefinitionMethod}{'ResourceFile' or 'FunctionParameters', see details.}

\item{FileName}{path to resource file}

\item{StockNameCC}{Variable for \code{\link[RstoxFDA]{StockSplittingParameters}}}

\item{StockNameS}{Variable for \code{\link[RstoxFDA]{StockSplittingParameters}}}

\item{ProbabilityType1As1}{Variable for \code{\link[RstoxFDA]{StockSplittingParameters}}}

\item{ProbabilityType5As1}{Variable for \code{\link[RstoxFDA]{StockSplittingParameters}}}

\item{ProbabilityType2As2}{Variable for \code{\link[RstoxFDA]{StockSplittingParameters}}}

\item{ProbabilityType4As2}{Variable for \code{\link[RstoxFDA]{StockSplittingParameters}}}

\item{ProbabilityType2As4}{Variable for \code{\link[RstoxFDA]{StockSplittingParameters}}}

\item{ProbabilityType4As4}{Variable for \code{\link[RstoxFDA]{StockSplittingParameters}}}

\item{ProbabilityType1As5}{Variable for \code{\link[RstoxFDA]{StockSplittingParameters}}}

\item{ProbabilityType5As5}{Variable for \code{\link[RstoxFDA]{StockSplittingParameters}}}

\item{UseProcessData}{If TRUE, bypasses execution of function and returns existing 'processData'}
}
\value{
\code{\link[RstoxFDA]{StockSplittingParameters}}.
}
\description{
Defines parameters for the stock-splitting analysis in Reca, including
parameters for misclassifying when determining stock membership of a specimen.
}
\details{
For DefinitionMethod 'ResourceFile', definitions are read from a UTF-8 encoded tab separated file with headers and one row
with headers corresponding to column names in \code{\link[RstoxFDA]{StockSplittingParameters}}.
see \code{\link[RstoxFDA]{StockSplittingParameters}} for further explanation on the coding system.

For DefinitionMethod 'FunctionParameters' defintions are constructed from parameters to this function.
}
\seealso{
\code{\link[RstoxFDA]{PrepareRecaEstimate}} for use of stock-splitting parameters in Reca-estimation.
}
\concept{StoX-Reca functions}
\concept{StoX-functions}
