% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{DefineStockSplittingParamteres}
\alias{DefineStockSplittingParamteres}
\title{Define Stock Splitting Parameters}
\usage{
DefineStockSplittingParamteres(
  processData,
  DefinitionMethod = c("ResourceFile", "FunctionParameters"),
  FileName = character(),
  StockNameCC = character(),
  StockNameS = character(),
  ProbabilityType1As1 = numeric(),
  ProbabilityType5As1 = numeric(),
  ProbabilityType2As2 = numeric(),
  ProbabilityType4As2 = numeric(),
  ProbabilityType2As4 = numeric(),
  ProbabilityType4As4 = numeric(),
  ProbabilityType1As5 = numeric(),
  ProbabilityType5As5 = numeric(),
  UseProcessData = F
)
}
\arguments{
\item{processData}{data.table() as returned from this function}

\item{FileName}{path to resource file}

\item{StockNameCC}{Variable for \code{\link[RstoxFDA]{StockSplittingParamteres}}}

\item{StockNameS}{Variable for \code{\link[RstoxFDA]{StockSplittingParamteres}}}

\item{ProbabilityType1As1}{Variable for \code{\link[RstoxFDA]{StockSplittingParamteres}}}

\item{ProbabilityType5As1}{Variable for \code{\link[RstoxFDA]{StockSplittingParamteres}}}

\item{ProbabilityType2As2}{Variable for \code{\link[RstoxFDA]{StockSplittingParamteres}}}

\item{ProbabilityType4As2}{Variable for \code{\link[RstoxFDA]{StockSplittingParamteres}}}

\item{ProbabilityType2As4}{Variable for \code{\link[RstoxFDA]{StockSplittingParamteres}}}

\item{ProbabilityType4As4}{Variable for \code{\link[RstoxFDA]{StockSplittingParamteres}}}

\item{ProbabilityType1As5}{Variable for \code{\link[RstoxFDA]{StockSplittingParamteres}}}

\item{ProbabilityType5As5}{Variable for \code{\link[RstoxFDA]{StockSplittingParamteres}}}

\item{UseProcessData}{logical() Bypasses execution of function, and simply returns argument 'processData'}
}
\value{
\code{\link[RstoxFDA]{StockSplittingParamteres}}.
}
\description{
Defines parameters for the stock-splitting analysis in Reca, including
 parameters for misclassifying when determining stock membership of a specimen.
}
\details{
For DefinitionMethod 'ResourceFile', definitions are read from a UTF-8 encoded tab separated file with headers and one row
 with headers corresponding to column names in \code{\link[RstoxFDA]{StockSplittingParamteres}}.
 see \code{\link[RstoxFDA]{StockSplittingParamteres}} for further explanation on the coding system.
}
