% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{DefineTemporalCategories}
\alias{DefineTemporalCategories}
\title{Define Temporal Categories}
\usage{
DefineTemporalCategories(processData, temporalCategory = c("Quarter",
  "Month", "Custom"), customPeriods = NULL, seasonal = T,
  years = NULL, encoding = "UTF-8", useProcessData = F)
}
\arguments{
\item{processData}{data.table() as returned from this function}

\item{temporalCategory}{character(), defaults to 'Quarter', type of temporal category: 'Quarter', 'Month' or 'Custom'}

\item{customPeriods}{character(), provided if temporalCategory is 'Custom', vector of strings formatted as DD-MM, giving the start date of each temporal category.}

\item{seasonal}{logical(), defaults to TRUE, whether the categories are defined as seasons or periods, that is, whether e.g. Q1 of 2015 is condiedered the same category as Q1 2016.}

\item{years}{integer() vector, optional, provide if non-seasonal categories should specify year.}

\item{encoding}{encoding of resource file}

\item{useProcessData}{logical() Bypasses execution of function, and simply returns argument 'processData'}
}
\value{
Temporal Categories, see: \code{\link[RstoxFDA]{TemporalCategories}}.
}
\description{
Define temporal categories for grouping data based on date.
}
\details{
A seasonal definition has the effect of making caegories independent of year,
 and of making season cross year boundaries of 1st of January is not a the start of any category.
}
