% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{DefineWeightConversionFactor}
\alias{DefineWeightConversionFactor}
\title{Define Weight Conversion Factors}
\usage{
DefineWeightConversionFactor(
  processData,
  DefinitionMethod = c("ResourceFile"),
  FileName,
  UseProcessData = F
)
}
\arguments{
\item{processData}{\code{\link[RstoxFDA]{WeightConversionTable}} as returned from this function}

\item{DefinitionMethod}{'ResourceFile'. See details.}

\item{FileName}{path to resource file}

\item{UseProcessData}{Bypasses execution of function, if TRUE, and simply returns argument 'processData'}
}
\value{
\code{\link[RstoxFDA]{WeightConversionTable}}
}
\description{
Define weight conversion factors for calculating approximate weights for a product type (e.g. round fish)
 from the weight of other fish products, such as gutted fish.
}
\details{
For DefinitionMethod 'ResourceFile':
 Definitions are read from a tab separated, UTF-8 encoded file with headers. Columns defined as:
 \describe{
 \item{Column 1: 'Description'}{Free-text description of the product}
 \item{Column 2: 'Species'}{Identifier for the species that the conversion applies to}
 \item{Column 3: 'ProductType'}{Identifier for the type of product that the conversion applies to}
 \item{Column 4: 'WeightFactor'}{scalar value that weights for the given 'ProductType' can be multiplied with to approximate the desired product type (w.g. round fish).}
 }
 
 Missing values for WeightFactor are interpreted as NA, and will result in NA for weights after conversion.
}
