% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierDetection.R
\name{filterLogLinear}
\alias{filterLogLinear}
\title{Filter length-weight outliers}
\usage{
filterLogLinear(
  individuals,
  logalfa,
  beta,
  sigma,
  kAl,
  kAu = kAl,
  weightCol = "IndividualRoundWeight",
  lengthCol = "IndividualTotalLength"
)
}
\arguments{
\item{individuals}{\code{\link[data.table]{data.table}} of fish records}

\item{logalfa}{The alfa parameter (on a log scale) for the log-linear model}

\item{beta}{The beta parameter for the log-linear model}

\item{sigma}{The standard deviation of weight for the log-linear model}

\item{kAl}{Number of standard deviations (on a log scale) that defines the lower limit of the acceptable region}

\item{kAu}{Number of standard deviations (on a log scale) that defines the upper limit of the acceptable region}

\item{weightCol}{name of column in 'individuals' that contain fish weight (in a unit corresponding to alfa and beta). Default correspond to the Individual level of \code{\link[RstoxData]{StoxBioticData}}, but see details.}

\item{lengthCol}{name of column in 'individuals' that contain fish length (in a unit corresponding to alfa and beta). Default correspond to the Individual level of \code{\link[RstoxData]{StoxBioticData}}}
}
\value{
\code{\link[data.table]{data.table}}, like individuals, but with some records removed.
}
\description{
Removes fish records that fall outside an acceptable weight-length region
 defined by the log-linear model
 
 weight=alfa\*length^beta\*exp(epsilon); epsilon~N(0,sigma^2)
 
 with alfa=exp(logalfa), logalfa and other parameters corresponding to arguments to this function.
}
\details{
This function is intended to provide the same filtering that is offered in ECA 3.x and ECA 4.x
 for removing outliers based on a log-linear weight-length model, and
 function arguments are named to correspond to the naming convention used in ECA.
 
 Records are removed if their weights that outside the range from:
 alfa\*L^beta\*exp(kAu\*sigma)
 to
 alfa\*L^beta\*exp(-kAl\*sigma)

 any records with missing length or weight is not removed.
 
 Note that kAl and kAu are given on a log scale, so that the acceptable region
 is not symmetric around the growth curve when kAl=kAu.
}
\seealso{
Other data QA functions: 
\code{\link{FilterAgeLengthOutliersStoxBiotic}()},
\code{\link{FilterWeightLengthOutliersStoxBiotic}()},
\code{\link{ListBioticDifference}()},
\code{\link{ReportFdaSOP}()},
\code{\link{filterVonBsigma}()}
}
\concept{data QA functions}
