% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GearGroupFdirTable-datadoc.R
\docType{data}
\name{GearGroupFdirTable}
\alias{GearGroupFdirTable}
\title{Gear groups}
\format{
\code{\link[RstoxFDA]{MetierTable}} with column 'gearcode' identifying gear codes used in Norwegian fisheries data (derived from NS 9400)
}
\usage{
data(GearGroupFdirTable)
}
\description{
Example of metier table for assigning gear groups in NS9400 / Norwegian Directorate of fisheries
}
\examples{
data(GearGroupFdirTable)
data(activityCensus)
annotated <- appendMetier(activityCensus,
                          GearGroupFdirTable,
                          "gearNS",
                          metierColName = "Hovedgruppe Redskap")
table(annotated$gearFAO, annotated$"Hovedgruppe Redskap")
}
\seealso{
Other gear coding functions: 
\code{\link{AddGearGroupStoxBiotic}()},
\code{\link{AddGearGroupStoxLanding}()},
\code{\link{SetShortGearBiotic}()},
\code{\link{appendMetier}()},
\code{\link{metier4table}},
\code{\link{metier5table}},
\code{\link{metier6table}},
\code{\link{readMetierTable}()}
}
\concept{gear coding functions}
\keyword{datasets}
