% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecaWrap.R
\name{getCovariateMap}
\alias{getCovariateMap}
\title{Make Covariate Map}
\usage{
getCovariateMap(covariate, samples, landings)
}
\arguments{
\item{covariate}{character() name of covariate to make map for}

\item{samples}{data.table() the samples the covariate mapping should be defined for (must contain column specified by 'covariate')}

\item{landings}{data.table() the landings the covariate mapping should be defined for (may contain column specified by 'covariate')}
}
\value{
list() where the nth member represents the RECA covariate value 'n', so that for a covarate 'cov', the integer i is used for value a, when covariateMaps[[cov]][[i]] == a
}
\description{
Constructs a covariate map for R-ECA
}
\details{
RECA require all covariate to be formatted as integers.
 In order to maintain connection to other formats for the covariates, they should be converted back and forth with a pre-defined covariate map
}
\examples{
 data(catchsamples)
 data(landings)
 catchsamples$Metier5 <- catchsamples$LEmetier5
 landings$Metier5 <- landings$FishingActivityCategoryEuropeanLvl5
 getCovariateMap("Metier5", catchsamples, landings)
}
\seealso{
Other Reca functions: 
\code{\link{catchsamples}},
\code{\link{convertRecaData}()},
\code{\link{getLandings}()},
\code{\link{landings}},
\code{\link{makeAgeTracesRECA}()},
\code{\link{makeResultTableRECA}()},
\code{\link{plotAgeTraces}()},
\code{\link{plotCatchAtAge}()},
\code{\link{prepRECA}()},
\code{\link{rEcaDataReport}()},
\code{\link{recaDataExample}},
\code{\link{recaPrediction}},
\code{\link{runRECA}()},
\code{\link{sanitizeRecaInput}()},
\code{\link{tabulateFisheries}()}
}
\concept{Reca functions}
