% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecaWrap.R
\name{getLandings}
\alias{getLandings}
\title{Formats landings for R-ECA.}
\usage{
getLandings(
  landings,
  covariates,
  covariateMaps,
  date = NULL,
  month = NULL,
  quarter = NULL
)
}
\arguments{
\item{landings}{data.table() with total landings (as in \code{\link[RstoxFDA]{prepRECA}}, and \code{\link[RstoxFDA]{rEcaDataReport}}), each row corresponding to one cell. Contains columns:
\describe{
 \item{LiveWeightKG}{numeric(). Total landings (Live/Round weight in Kg) for the cell}
 \item{...}{Additional columns to be used as covariates. These define each cell. Type of covariate must be sepcified in 'fixedEffects', 'randomEffects' or 'carEffect'}
}}

\item{covariates}{character() vector of covariates used in model}

\item{covariateMaps}{list() mapping covariate values from integers used in RECA to values used in 'landings'. For a covarate 'cov', the integer i is used for value a, when covariateMaps[[cov]][[i]] == a}

\item{date}{POSIXct() vector, matching the number of rows in 'landings', date of catch, see details.}

\item{month}{integer() vector, matching the number of rows in 'landings', month of catch (1 for January, etc.), see details.}

\item{quarter}{integer() vector, vector, matching the number of rows in 'landings', quarter of catch (1 for Q1, etc.), see details.}
}
\value{
Landings object as required by \code{\link[Reca]{eca.predict}}
}
\description{
Prepares a Landings object as required by \code{\link[Reca]{eca.predict}}.
}
\details{
The parameters 'date', 'month', and 'quarter' are used to set the temporal resolution for catch at age prediction.
 Provide exactly one of these, and set the other ones to NULL.
 Temporal resolution need not match any temporal covariate used.
 One can for example run with month, even if Quarter is a covariate in the model.
}
\examples{
 data(catchsamples)
 data(landings)
 catchsamples$Metier5 <- catchsamples$LEmetier5
 landings$Metier5 <- landings$FishingActivityCategoryEuropeanLvl5
 covMap <- getCovariateMap("Metier5", catchsamples, landings)
 getLandings(landings, c("Metier5"), covMap, month=landings$Month)
}
