% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAsubAreas-datadoc.R
\docType{data}
\name{GSAsubArea}
\alias{GSAsubArea}
\title{GFCM sub areas (mixed area codes)}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with area names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}}.
}
\usage{
data(GSAsubArea)
}
\description{
General Fisheries Commission for the Mediterranean - GFCM area coding system (FAO Major Fishing Area 37) 
Polygons are defined on for Divisions, and the Geographic Subareas are annotated in the column F_SUBAREA.
}
\details{
Polygons are defined in WGS84 coordinates (unprojected).

The data contains the following columns with rows for each area:
\describe{
 \item{F_AREA}{FAO Major fishing Area (37)}
 \item{F_SUBAREA}{The GSAs (Geographical subareas)}
 \item{F_DIVISION}{The GFCM divisions, subdividing the GSAs.}
}
}
\examples{
 # plot divisions
 data(GSAsubArea)
 plotArea(areaDef=GSAsubArea)
 
 # plot subareas
 RstoxFDA::plotArea(areaDef=GSAsubArea, areaNameCol = "F_SUBAREA")
}
\seealso{
Other area code polygons: 
\code{\link{ICESareas}},
\code{\link{ICESdivision}},
\code{\link{ICESrectangles}},
\code{\link{ICESsubArea}},
\code{\link{ICESsubDivision}},
\code{\link{ICESunit}},
\code{\link{NAFOareas}},
\code{\link{locationsFdir2017}},
\code{\link{locationsFdir2018}},
\code{\link{mainareaFdir2017}},
\code{\link{mainareaFdir2018}},
\code{\link{portcodes2020}}
}
\concept{area code polygons}
\keyword{datasets}
