% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICESareas-datadoc.R
\docType{data}
\name{ICESareas}
\alias{ICESareas}
\title{ICES areas}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with area names identified in the column 'polygonName'. See \code{\link[RstoxBase]{StratumPolygon}}.
}
\usage{
data(ICESareas)
}
\description{
Definition for FAO Northeast Atlantic (Major Fishing Area 27) area coding system used in ICES data calls.
Polygons are defined on different levels of aggregation, either Subarea, Division, Subdivision or Unit.
}
\details{
In addtion to columns idenitfying the polygons in FAO nomenclature, a column with the area in square kilometers is included.
Polygons are derived from shapefiles provided by ICES web-portals.

Polygons are defined in WGS84 coordinates (unprojected).
}
\examples{
 # plot ICES areas
 data(ICESareas)
 sp::plot(ICESareas)
}
\keyword{datasets}
