% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICESrectangles-datadoc.R
\docType{data}
\name{ICESrectangles}
\alias{ICESrectangles}
\title{ICES rectangles}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with rectange names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}}.
}
\usage{
data(ICESrectangles)
}
\description{
Definitions for ICES rectangles.
ICES rectangles are defined by grids made up from integer longitudes, integer latitudes, and a bisection between each pair of latitudes.
They derive their name from being rectangular in mercator projection.
}
\details{
The data contains the following columns with rows for each rectangle:
\describe{
 \item{ICESNAME}{name of the rectangle in standard ICES notation}
 \item{SOUTH}{southern border of rectangle (latitude)}
 \item{WEST}{western border of rectangle (longitude)}
 \item{NORTH}{northern border of rectangle (latitdue)}
 \item{EAST}{eastern border of rectangle (longitude)}
 \item{Area_Full}{Code for full ICES area, including the FAO area code (27).}
 \item{Area_27}{Code for full ICES area, excluding the FAO area code (27).}
}

The data also contains columns for the individul components of the full ICES area code:
<Major_FA>.<SubArea>.<Division>.<SubDivision>.<Unit>

Polygons are derived from shapefiles provided by ICES web-portals.

Polygons are defined in WGS84 coordinates (unprojected).
}
\examples{
 # plot ICES areas alongside main areas
 data(ICESrectangles)
 sp::plot(ICESrectangles)
}
\concept{area code polygons}
\keyword{datasets}
