% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICESsubDivision-datadoc.R
\docType{data}
\name{ICESsubDivision}
\alias{ICESsubDivision}
\title{ICES areas}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with area names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}}.
}
\usage{
data(ICESsubDivision)
}
\description{
Definition ICES SubDivisions as they have been defined from 2018 inclusive.
}
\details{
Polygons are derived from shapefiles provided by ICES web-portals, and has been edited with some simplifications.
Notably detailed coast-lines have been removed, in favor of drawing area borders on land-mass.

Polygons are defined in WGS84 coordinates (unprojected).
}
\examples{
 # plot ICES areas
 data(ICESsubDivision)
 plotArea(areaDef=ICESsubDivision)
}
\seealso{
Other area code polygons: 
\code{\link{ICESareas}},
\code{\link{ICESdivision}},
\code{\link{ICESrectangles}},
\code{\link{ICESsubArea}},
\code{\link{ICESunit}},
\code{\link{NAFOareas}},
\code{\link{locationsFdir2017}},
\code{\link{locationsFdir2018}},
\code{\link{mainareaFdir2017}},
\code{\link{mainareaFdir2018}}
}
\concept{area code polygons}
\keyword{datasets}
