% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landingsAnalyses.R
\name{imputeCatchesLandings}
\alias{imputeCatchesLandings}
\title{Impute catches to landings}
\usage{
imputeCatchesLandings(
  landings,
  catchPartition,
  tripIdCol = "tripid",
  speciesCol = "Art FAO (kode)",
  valueColumns = c("Bruttovekt", "Produktvekt", "Rundvekt")
)
}
\arguments{
\item{landings}{\code{\link[data.table]{data.table}} containing landings records, and a column identifying the trip that caught the landed catch ('tripIdCol').}

\item{catchPartition}{\code{\link[RstoxFDA]{TripPartition}} partitioning the catch of each trip on grouping variables.}

\item{tripIdCol}{character() that identifies the column in 'landings' that contain trip ids corresponding to those in 'catchPartition'}

\item{speciesCol}{character() that identifies the column in 'landings' that contain species ids corresponding to those in 'catchPartition'}

\item{valueColumns}{character() vector with names of columns that should be redistributed over the imputed landings}
}
\value{
\code{\link[data.table]{data.table}} with 'landings' that have artificial landings imputed and 'valueColumns' redistributed to reflect the 'logbookPartition'.
}
\description{
Imputes landings to represent individual catches of species from a trip, and redistributes selected values over the imputed landings.
 This mimics a situation where landings are reported for each catch rather than each trip.
 This is useful to obtain a higher resolution on some grouping variables and consolidate logbook records with landing records, 
 while preserving data for landings not reported in logbooks, and preseriving the total value equal to that in landings.
}
\details{
Important: columns in landings that are not in 'valueColumns' and are not in the group definition ('catchPartition') will
 be duplicated exactly for the imputed catches. 
 That means that any weight-column not in 'valueColumns' will result in duplicated total weights for that column.
 That also means that any identifiers / keys may be copied and thus be rendered non-unique in the imputed landings.
 Landings are imputed and redistributed line-by-line, so that the total proportions of categorical variables not in the group definition ('catchPartition') are preserved for the columns in 'valueColumns'.
 
 Landings for which no partition can be found in 'catchPartition' are left untouched.
 Catchpartitions with no corresponding landings are ignored.
 
 Default values for parameters 'speciesCol' and 'valueColumns' are suitable for data read with \code{\link[RstoxData]{readLssFile}}.
 Default values for 'tripIdCol' corresponds to defaults for \code{\link[RstoxFDA]{makeTripIds}} and \code{\link[RstoxFDA]{appendTripIdLandings}}.
}
\seealso{
\code{\link[RstoxFDA]{calculateLogbookPartitionByTrip}} for calculation of 'catchPartition'
 \code{\link[RstoxFDA]{makeTripIds}} and \code{\link[RstoxFDA]{appendTripIdLandings}} for obtaining 'landings' with trip-ids.
}
