% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.AreaCodePosition}
\alias{is.AreaCodePosition}
\title{Check if argument is AreaCodePosition}
\usage{
is.AreaCodePosition(AreaCodePosition)
}
\arguments{
\item{AreaCodePosition}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conformed to specification for \code{\link[RstoxFDA]{AreaCodePosition}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{AreaCodePosition}}
}
