% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.AreaPosition}
\alias{is.AreaPosition}
\title{Check if argument is AreaPosition}
\usage{
is.AreaPosition(AreaPosition)
}
\arguments{
\item{AreaPosition}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conforms to specification for \code{\link[RstoxFDA]{AreaPosition}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{AreaPosition}}
}
