% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.CarNeighbours}
\alias{is.CarNeighbours}
\title{Check if argument is CarNeighbours}
\usage{
is.CarNeighbours(CarNeighbours)
}
\arguments{
\item{CarNeighbours}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conformed to specification for \code{\link[RstoxFDA]{CarNeighbours}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{CarNeighbours}}
}
