% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.ClassificationError}
\alias{is.ClassificationError}
\title{Check if argument is ClassificationError}
\usage{
is.ClassificationError(ClassificationError)
}
\arguments{
\item{ClassificationError}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conformed to specification for \code{\link[RstoxFDA]{ClassificationError}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{ClassificationError}}
}
