% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.LengthConversionTable}
\alias{is.LengthConversionTable}
\title{Check if argument is LengthConversionTable}
\usage{
is.LengthConversionTable(LengthConversionTable)
}
\arguments{
\item{LengthConversionTable}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conforms to specification for \code{\link[RstoxFDA]{LengthConversionTable}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{LengthConversionTable}}
}
