% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.ParameterConvergenceData}
\alias{is.ParameterConvergenceData}
\title{Checks if argument is \code{\link[RstoxFDA]{ParameterConvergenceData}}}
\usage{
is.ParameterConvergenceData(ParameterConvergenceData)
}
\arguments{
\item{ParameterConvergenceData}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conforms to specification for \code{\link[RstoxFDA]{ParameterConvergenceData}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{ParameterConvergenceData}}
}
\concept{Data types}
