% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.ParameterizationSummaryData}
\alias{is.ParameterizationSummaryData}
\title{Checks if argument is \code{\link[RstoxFDA]{ParameterizationSummaryData}}}
\usage{
is.ParameterizationSummaryData(ParameterizationSummaryData)
}
\arguments{
\item{ParameterizationSummaryData}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conforms to specification for \code{\link[RstoxFDA]{ParameterizationSummaryData}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{ParameterizationSummaryData}}
}
