% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.RecaCatchAtAge}
\alias{is.RecaCatchAtAge}
\title{Check if argument is RecaCatchAtAge}
\usage{
is.RecaCatchAtAge(RecaCatchAtAge)
}
\arguments{
\item{RecaCatchAtAge}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conforms to specification for \code{\link[RstoxFDA]{RecaCatchAtAge}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{RecaCatchAtAge}}
}
