% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.RecaData}
\alias{is.RecaData}
\title{Check if argument is RecaData}
\usage{
is.RecaData(RecaData)
}
\arguments{
\item{RecaData}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conformed to specification for \code{\link[RstoxFDA]{RecaData}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{RecaData}}
}
