% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.RecaResult}
\alias{is.RecaResult}
\title{Check if argument is RecaResult}
\usage{
is.RecaResult(RecaResult)
}
\arguments{
\item{RecaResult}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conformed to specification for \code{\link[RstoxFDA]{RecaResult}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{RecaResult}}
}
