% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.ReportFdaByAgeData}
\alias{is.ReportFdaByAgeData}
\title{Checks if argument is \code{\link[RstoxFDA]{ReportFdaByAgeData}}}
\usage{
is.ReportFdaByAgeData(ReportFdaByAgeData)
}
\arguments{
\item{ReportFdaByAgeData}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conforms to specification for \code{\link[RstoxFDA]{ReportFdaByAgeData}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{ReportFdaByAgeData}}
}
