% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.ReportFdaData}
\alias{is.ReportFdaData}
\title{Checks if argument is \code{\link[RstoxFDA]{ReportFdaData}}}
\usage{
is.ReportFdaData(ReportFdaData)
}
\arguments{
\item{ReportFdaData}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conforms to specification for \code{\link[RstoxFDA]{ReportFdaData}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{ReportFdaData}}
}
\concept{Data types}
