% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.ReportFdaSamplingData}
\alias{is.ReportFdaSamplingData}
\title{Check if argument is ReportFdaSamplingData}
\usage{
is.ReportFdaSamplingData(ReportFdaSamplingData)
}
\arguments{
\item{ReportFdaSamplingData}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conforms to specification for \code{\link[RstoxFDA]{ReportFdaSamplingData}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{ReportFdaSamplingData}}
}
\concept{Data types}
