% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.ReportFdaSOP}
\alias{is.ReportFdaSOP}
\title{Checks if argument is \code{\link[RstoxFDA]{ReportFdaSOP}}}
\usage{
is.ReportFdaSOP(ReportFdaSOP)
}
\arguments{
\item{ReportFdaSOP}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conforms to specification for \code{\link[RstoxFDA]{ReportFdaSOP}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{ReportFdaSOP}}
}
\concept{Data types}
