% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.ReportRecaCatchAtAgeData}
\alias{is.ReportRecaCatchAtAgeData}
\title{Checks if argument is \code{\link[RstoxFDA]{ReportRecaCatchAtAgeData}}}
\usage{
is.ReportRecaCatchAtAgeData(ReportRecaCatchAtAgeData)
}
\arguments{
\item{ReportRecaCatchAtAgeData}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conforms to specification for \code{\link[RstoxFDA]{ReportRecaCatchAtAgeData}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{ReportRecaCatchAtAgeData}}
}
