% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.ReportRecaData}
\alias{is.ReportRecaData}
\title{Checks if argument is \code{\link[RstoxFDA]{ReportRecaData}}}
\usage{
is.ReportRecaData(ReportRecaData)
}
\arguments{
\item{ReportRecaData}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conforms to specification for \code{\link[RstoxFDA]{ReportRecaData}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{ReportRecaData}}
}
