% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.StockSplittingParameters}
\alias{is.StockSplittingParameters}
\title{Check if argument is StockSplittingParameters}
\usage{
is.StockSplittingParameters(StockSplittingParameters)
}
\arguments{
\item{StockSplittingParameters}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conforms to specification for \code{\link[RstoxFDA]{StockSplittingParameters}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{StockSplittingParameters}}
}
\concept{Data types}
