% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.StockSplittingParamteres}
\alias{is.StockSplittingParamteres}
\title{Check if argument is StockSplittingParamteres}
\usage{
is.StockSplittingParamteres(StockSplittingParamteres)
}
\arguments{
\item{StockSplittingParamteres}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conforms to specification for \code{\link[RstoxFDA]{StockSplittingParamteres}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{StockSplittingParamteres}}
}
