% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.TemporalDefinition}
\alias{is.TemporalDefinition}
\title{Check if argument is TemporalDefinition}
\usage{
is.TemporalDefinition(TemporalDefinition)
}
\arguments{
\item{TemporalDefinition}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conformed to specification for \code{\link[RstoxFDA]{TemporalDefinition}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{TemporalDefinition}}
}
