% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.Translation}
\alias{is.Translation}
\title{Checks if argument is \code{\link[RstoxData]{Translation}}}
\usage{
is.Translation(Translation)
}
\arguments{
\item{Translation}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conforms to specification for \code{\link[RstoxData]{Translation}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxData]{Translation}}
}
