% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.UnifiedVariableDefinition}
\alias{is.UnifiedVariableDefinition}
\title{Check if argument is UnifiedVariableDefinition}
\usage{
is.UnifiedVariableDefinition(UnifiedVariableDefinition)
}
\arguments{
\item{UnifiedVariableDefinition}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conformed to specification for \code{\link[RstoxFDA]{UnifiedVariableDefinition}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{UnifiedVariableDefinition}}
}
