% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.WeightConversionTable}
\alias{is.WeightConversionTable}
\title{Check if argument is WeightConversionTable}
\usage{
is.WeightConversionTable(WeightConversionTable)
}
\arguments{
\item{WeightConversionTable}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conforms to specification for \code{\link[RstoxFDA]{WeightConversionTable}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{WeightConversionTable}}
}
