% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{KommunePolygons}
\alias{KommunePolygons}
\title{Kommune polygons}
\description{
\code{\link[sp]{SpatialPolygonsDataFrame}} with area names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}}.
 
 Polygons are defined in WGS84 coordinates (unprojected).

 The data contains the following columns with rows for each area:
  \describe{
   \item{id}{Offical code for the muncipality (kommunenummer)}
   \item{name}{Official name for the muncipality}
   \item{StratumName}{Equal to id}
  }
}
\seealso{
Other Data types: 
\code{\link{AgeErrorMatrix}},
\code{\link{AreaPosition}},
\code{\link{CarNeighbours}},
\code{\link{LandingData}},
\code{\link{LandingsArchiveData}},
\code{\link{LengthConversionTable}},
\code{\link{LocodeTable}},
\code{\link{LstLogbookData}},
\code{\link{MetierTable}},
\code{\link{ParameterConvergenceData}},
\code{\link{ParameterizationSummaryData}},
\code{\link{RecaCatchAtAge}},
\code{\link{RecaData}},
\code{\link{RecaParameterData}},
\code{\link{RecaResult}},
\code{\link{ReportFdaByAgeData}},
\code{\link{ReportFdaCatchAtAgeCovarianceData}},
\code{\link{ReportFdaCatchAtAgeData}},
\code{\link{ReportFdaCatchAtLengthAndAgeData}},
\code{\link{ReportFdaCatchAtLengthData}},
\code{\link{ReportFdaLandingData}},
\code{\link{ReportFdaLengthAtAgeData}},
\code{\link{ReportFdaSamplingData}},
\code{\link{ReportFdaSopData}},
\code{\link{ReportFdaSummaryData}},
\code{\link{ReportFdaWeightAtAgeData}},
\code{\link{StockSplittingParameters}},
\code{\link{TemporalDefinition}},
\code{\link{TripPartition}},
\code{\link{WeightConversionTable}},
\code{\link{is.AgeErrorMatrix}()},
\code{\link{is.AreaPosition}()},
\code{\link{is.CarNeighbours}()},
\code{\link{is.LengthConversionTable}()},
\code{\link{is.ParameterConvergenceData}()},
\code{\link{is.ParameterizationSummaryData}()},
\code{\link{is.RecaCatchAtAge}()},
\code{\link{is.RecaData}()},
\code{\link{is.RecaParameterData}()},
\code{\link{is.RecaResult}()},
\code{\link{is.ReportFdaByAgeData}()},
\code{\link{is.ReportFdaSOP}()},
\code{\link{is.ReportFdaSamplingData}()},
\code{\link{is.StockSplittingParameters}()},
\code{\link{is.TemporalDefinition}()},
\code{\link{is.Translation}()},
\code{\link{is.WeightConversionTable}()}
}
\concept{Data types}
