% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kommuner2022-datadoc.R
\docType{data}
\name{kommuner2022}
\alias{kommuner2022}
\title{Norwegian municipalities (kommune)}
\format{
\code{\link[sf]{sf}} with area names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}} (v.2).
}
\usage{
data(kommuner2022)
}
\description{
Definitions of Norwegian municipalities (kommune), current in 2022.
providing their borders, their official name and their official code.
}
\details{
Norwegian landings data often list landing sites by identifiers for municipalities (kommunenummer).
This resource can be used to infer location code that are associated with polygons or coordinates,
such as UN/LOCODES.

The Norwegian municipalities have been subject to revision, and other resources may have to be obtained
to completely handle data from other years.
}
\examples{
# map kommune to locodes
data(kommuner2022)
data(portcodes2020)
nocodes <- portcodes2020[portcodes2020$Country=="NO" & !is.na(portcodes2020$latitude),]
locodemap <- RstoxFDA::appendAreaCode(nocodes, kommuner2022, 
        colName="kommune", 
        latName = "latitude", 
        lonName = "longitude")[,c("kommune", "Location")]
# Note that kommune does not uniquely identify locode:
locodemap[locodemap$kommune==4601,]
}
\concept{area code polygons}
\keyword{datasets}
