% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings-datadoc.R
\docType{data}
\name{landings}
\alias{landings}
\title{Landings of Haddock}
\format{

}
\usage{
data(landings)
}
\description{
Aggregated landings of Haddock by Norwegian vessels in 2018 as reported in salesnotes.
Gear-codes are converted from national standard and should be considered approximate.
}
\examples{
data(landings)
sum(landings$OfficialLandingsWeight)
}
\seealso{
Other Reca functions: 
\code{\link{catchsamples}},
\code{\link{convertRecaData}()},
\code{\link{getCovariateMap}()},
\code{\link{getLandings}()},
\code{\link{makeAgeTracesRECA}()},
\code{\link{makeResultTableRECA}()},
\code{\link{plotAgeTraces}()},
\code{\link{plotCatchAtAge}()},
\code{\link{prepRECA}()},
\code{\link{rEcaDataReport}()},
\code{\link{recaDataExample}},
\code{\link{recaPrediction}},
\code{\link{runRECA}()},
\code{\link{sanitizeRecaInput}()},
\code{\link{tabulateFisheries}()}

Other landings functions: 
\code{\link{ReadLandingFDA}()},
\code{\link{ReportFdaLandings}()},
\code{\link{ReportFdaSampling}()},
\code{\link{addLogbookColumns}()},
\code{\link{appendTripIdLandings}()},
\code{\link{imputeCatchesLandings}()},
\code{\link{logbookAdjustment}()},
\code{\link{makeTripIds}()},
\code{\link{plotBubbleMap}()},
\code{\link{sourceLogbookColumns}()},
\code{\link{tabulateFisheries}()}
}
\concept{Reca functions}
\concept{landings functions}
\keyword{datasets}
