% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{LengthConversionTable}
\alias{LengthConversionTable}
\title{Length Conversion Table (LengthConversionTable)}
\description{
Length conversion parameters realting different length measurements.
}
\details{
Length conversion factors relating different length measurements, such as 'standard length' and 'fork length'
 based on a linear regression fit between length measurements:
 L1 = alpha + beta \* L2,
 where L1 and L2 are different length measurements
 and 'alpha' and 'beta' are species-specific coefficients.
 
 \code{\link[data.table]{data.table}} with columns:
 \describe{
 \item{'Description'}{character: Free-text description of the product}
 \item{'Species'}{character: Identifier for the species that the conversion applies to}
 \item{'MeasurmentType'}{character: Identifier for the type of length measurement for the independent variable (L2 above)}
 \item{'Alpha'}{numeric: scalar value representing the intercept (in cm) of a linear regression fit between length measurements.}
 \item{'Beta'}{numeric: scalar value representing the slope of a linear regression fit between length measurements.}
 }
}
