% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalyticalBaselineFunctions.R
\name{LiftStrata}
\alias{LiftStrata}
\title{Unify strata for AnalyticalPSUEstimateData}
\usage{
LiftStrata(AnalyticalPSUEstimateData)
}
\arguments{
\item{AnalyticalPSUEstimateData}{\code{\link[RstoxFDA]{AnalyticalPSUEstimateData}} with estimates for all strata for each PSU that has positive abundance.}
}
\value{
\code{\link[RstoxFDA]{AnalyticalPSUEstimateData}} with zeroes and NaNs inferred for strata that have zero abundance in a PSU.
}
\description{
Populates unreported strata with zeroes and NaNs.
}
\details{
AnalyticalPSUEstimateData may be provided with stratification for each Primary Sampling Unit. These estimates cannot be combined into population level estimates
unless the same strata are provided for all Primary Sampling Units (PSUs). See \code{\link[RstoxFDA]{AnalyticalPopulationEstimate}}. This may not be the case for two reasons. The stratification may be local to each PSU,
and no common stratification exists between PSUs. For instance, stratified sub-sampling of a catch may have taken place where the criteria for
stratification was only used for that particular catch. Otherwise, some strata may not be estimated for some samples, simply because they did not exist there.
For instance stratification by length groups can use a common length stratification scheme for all PSUs, but not all of these strata will exist for each catch.

If strata are omitted because they do not exist for a particular Sample, Abundance, Frequencies and Totals can be inferred to be zero, while Means are undefined.
Application of this functions introduces zeros (Abundance, Frequency and Total) and NaN (Means) for all missing strata for all samples in 'AnalyticalPSUEstimateData'

In the case where strata definitions are local, the stratification are only useful for the calculation of sampling parameters. If this is the case
consider using the function \code{\link[RstoxFDA]{CollapseStrata}} before producing estimates for each sample.
}
\concept{Analytical estimation}
