% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{LoadFdaStratumPolygon}
\alias{LoadFdaStratumPolygon}
\title{Load common area definitions}
\usage{
LoadFdaStratumPolygon(
  processData,
  StrataSystem = c("FDIR.2017", "FDIR.2018", "ICES.2018", "ICES.SubArea.2018",
    "ICES.Division.2018", "ICES.SubDivision.2018", "ICES.Unit.2018",
    "ICES.Rectangles.2018", "NAFO", "NAFO.FDIR.2017", "NAFO.FDIR.2018"),
  UseProcessData = F
)
}
\arguments{
\item{processData}{\code{\link[RstoxBase]{StratumPolygon}} as returned from this function}

\item{StrataSystem}{the strata system to load}

\item{UseProcessData}{If TRUE, bypasses execution of function and returns existing 'processData'}
}
\value{
\code{\link[RstoxBase]{StratumPolygon}} with the desired strata definition.
}
\description{
Loads standard area definitions often used in analysis of fisheries.
}
\details{
This function has the same effect as \code{\link[RstoxBase]{DefineStratumPolygon}}, but loads
the polygons from the RstoxFDA package, rather than requiring them as a resource file.

The options available are:
\describe{
\item{"FDIR.2017"}{Main areas defined by the Norwegian Directorate of Fisheries, as they where defined to 2017 inclusive. Derived from \code{\link[RstoxFDA]{mainareaFdir2017}}.}
\item{"FDIR.2018"}{Main areas defined by the Norwegian Directorate of Fisheries, as they have been defined from 2018 inclusive. Derived from \code{\link[RstoxFDA]{mainareaFdir2018}}.}
\item{"ICES.2018"}{ICES areas as they have been defined from 2018 inclusive. Areas are provided with finest available spatial resolution (Sub-area, division, sub-division or unit), and the full naming convention is used (e.g. 27.3.d.27 or 27.4.b). Derived from \code{\link[RstoxFDA]{ICESareas}}.}
\item{"ICES.SubArea.2018"}{ICES areas as they have been defined from 2018 inclusive. Areas are provided as 'sub-areas' where they are defined, and the full naming convention is used (e.g. 27.3). Derived from \code{\link[RstoxFDA]{ICESsubArea}}.}
\item{"ICES.Division.2018"}{ICES areas as they have been defined from 2018 inclusive. Areas are provided as 'divisions' where they are defined, and the full naming convention is used (e.g. 27.3.d). Derived from \code{\link[RstoxFDA]{ICESdivision}}.}
\item{"ICES.SubDivision.2018"}{ICES areas as they have been defined from 2018 inclusive. Areas are provided as 'sub-divisions' where they are defined, and the full naming convention is used (e.g. 27.3.d.27). Derived from \code{\link[RstoxFDA]{ICESsubDivision}}.}
\item{"ICES.Unit.2018"}{ICES areas as they have been defined from 2018 inclusive. Areas are provided as 'unit' where they are defined, and the full naming convention is used (e.g. 27.3.d.28.1). Derived from \code{\link[RstoxFDA]{ICESunit}}.}
\item{"ICES.Rectangles.2018"}{ICES rectangles (e.g. 15C2) as they have been defined from 2018 inclusive. Derived from \code{\link[RstoxFDA]{ICESrectangles}}.}
\item{"NAFO"}{NAFO areas. NAFO naming convention is used. Derived from \code{\link[RstoxFDA]{NAFOareas}}.}
\item{"NAFO.FDIR.2017"}{NAFO areas combined with FDIR.2017. The naming convention of the Norwegian Directorate of fisheries is used.}
\item{"NAFO.FDIR.2018"}{NAFO areas combined with FDIR.2018. The naming convention of the Norwegian Directorate of fisheries is used.}
}

FDIR.2017 corresponds to the area coding referred to as system 2 at IMR.
FDIR.2018 corresponds to the area coding referred to as system 10 at IMR.
}
\concept{StoX-functions}
\concept{spatial coding functions}
