% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationsFdir2017-datadoc.R
\docType{data}
\name{locationsFdir2017}
\alias{locationsFdir2017}
\title{Location codes (FDIR to 2017 incl.)}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with location names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}}.
}
\usage{
data(locationsFdir2017)
}
\description{
Definition for location coding system defined by the Norwegian directorate of Fisheries up until 2017 (inclusive).
Polygons are defined in WGS84 coordinates (unprojected).
}
\examples{
 # compare locations and ICES rectangles in Barents Sea
 RstoxFDA::plotAreaComparison(
    RstoxFDA::ICESrectangles, 
    RstoxFDA::locationsFdir2017, 
    xlim=c(20,30), 
    ylim=c(70,75), 
    linetype2 = "dotted", 
    polygonColor1 = "green", 
    areaLabels2 = T)
}
\concept{area code polygons}
\keyword{datasets}
