% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landingsAnalyses.R
\name{logbookAdjustment}
\alias{logbookAdjustment}
\title{Adjust landings by logbooks}
\usage{
logbookAdjustment(
  landings,
  logbooks,
  gearCodes = character(),
  gearCol = "REDSKAP_FAO",
  timeCol = "STARTTIDSPUNKT",
  dateLastCatchCol = "Siste fangstdato",
  locCol = "LOKASJON_START",
  areaCol = "Hovedområde (kode)",
  speciesColLog = "FANGSTART_FAO",
  speciesColLand = "Art FAO (kode)",
  weightColLog = "RUNDVEKT",
  valueColLand = c("Bruttovekt", "Produktvekt", "Rundvekt"),
  lineIdCol = "Linjenummer"
)
}
\arguments{
\item{landings}{landings as returned by e.g. \code{\link[RstoxData]{readLssFile}}}

\item{logbooks}{logbooks as returned by e.g. \code{\link[RstoxData]{readErsFile}}}

\item{gearCodes}{character() with gear-codes which logbook cleaning should be applied for (as they appear in 'gearCol'). If not provided, cleaning is applied to all gear codes}

\item{gearCol}{character() that identifies the column in 'logbook' that contain gear codes.}

\item{timeCol}{character() that identifies the column in 'logbook' that contain the time columns to use for inferring quarter.}

\item{dateLastCatchCol}{character() that identifies the column in 'landings' that contain the date of last catch.}

\item{locCol}{character() that identifies the column in 'logbook' that contain area and location info.}

\item{areaCol}{character() that identifies the column in 'landings' that contain the area code.}

\item{speciesColLog}{character() that identifies the column in 'logbook' that contain information about species of catch.}

\item{speciesColLand}{character() that identifies the column in 'landings' that contain information about species of catch.}

\item{weightColLog}{character() that identifies the column in 'logbook' that contain the weights that redistribution of weights should be based on.}

\item{valueColLand}{character() vector that identifies the columns in 'landings' that contain values that are to be redistributed over imputed landings (weight and value columns)}

\item{lineIdCol}{character() that identifies the column in 'landings' that contain the identifier for each line on a sales note.}
}
\value{
'landings' with catches redistributed over more sales-note lines corresponding to logbook-catch records / fishing operations.
}
\description{
imputes landings and redistributes catch within quarter and area to mimic sales notes for each individual fishing operation (logbook record).
 Applicabe to formats parsed by \code{\link[RstoxData]{readLssFile}} and \code{\link[RstoxData]{readErsFile}}
}
\details{
As there is some variation in naming conventions used in the supported formats, 
 relevant column names are paramaterized so that they can be changed to the appropriate format dialect.
 This may also be utilized to choose between coding systems. For instance 'gearCol' can be specified as 
 "REDSKAP_FAO" or "REDSKAP_NS" to use FAO-gear codes or NS-9400 gear codes, respectively.
 
 Sales-note lines identifiers are made unique after imputation and redistibution ('lineIdCol').
}
