% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainareaFdir2017-datadoc.R
\docType{data}
\name{mainareaFdir2017}
\alias{mainareaFdir2017}
\title{Main areas (FDIR to 2017 incl.)}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with area names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}}.
}
\usage{
data(mainareaFdir2017)
}
\description{
Definition for area coding system defined by the Norwegian directorate of Fisheries up until 2017 (inclusive).
Polygons are defined in WGS84 coordinates (unprojected).
}
\examples{
data(mainareaFdir2017)
data(mainareaFdir2018)
sp::plot(mainareaFdir2017)
sp::plot(mainareaFdir2018, border="blue", add=TRUE)
}
\seealso{
Other area code polygons: 
\code{\link{GSAsubArea}},
\code{\link{ICESareas}},
\code{\link{ICESdivision}},
\code{\link{ICESrectangles}},
\code{\link{ICESsubArea}},
\code{\link{ICESsubDivision}},
\code{\link{ICESunit}},
\code{\link{NAFOareas}},
\code{\link{locationsFdir2017}},
\code{\link{locationsFdir2018}},
\code{\link{mainareaFdir2018}}
}
\concept{area code polygons}
\keyword{datasets}
