% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainareaFdir2018-datadoc.R
\docType{data}
\name{mainareaFdir2018}
\alias{mainareaFdir2018}
\title{Main areas (FDIR from 2018 incl.)}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with area names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}}.
}
\usage{
data(mainareaFdir2018)
}
\description{
Definition for area coding system defined by the Norwegian directorate of Fisheries (2018 revision).
This revision has been in use for their fishery statistics as of 2018 (inclusive).
Polygons are defined in WGS84 coordinates (unprojected).
}
\examples{
 data(mainareaFdir2018)
 sp::plot(mainareaFdir2018)
}
\seealso{
Other area code polygons: 
\code{\link{GSAsubArea}},
\code{\link{ICESareas}},
\code{\link{ICESdivision}},
\code{\link{ICESrectangles}},
\code{\link{ICESsubArea}},
\code{\link{ICESsubDivision}},
\code{\link{ICESunit}},
\code{\link{NAFOareas}},
\code{\link{locationsFdir2017}},
\code{\link{locationsFdir2018}},
\code{\link{mainareaFdir2017}}
}
\concept{area code polygons}
\keyword{datasets}
