% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecaReport.R
\name{makeAgeTracesRECA}
\alias{makeAgeTracesRECA}
\title{Catch at age traces}
\usage{
makeAgeTracesRECA(prediction, unit = "millions", plusGroup = NULL)
}
\arguments{
\item{prediction}{as returned by \code{\link[Reca]{eca.predict}} or \code{\link[RstoxFDA]{runRECA}}.}

\item{unit}{unit of traced estimates. See details.}

\item{plusGroup}{Fish this age or older will be grouped in one trace.}
}
\value{
data.table() with one column for each age (or plus group), and one row for each iteration of \code{\link[Reca]{eca.predict}}.
}
\description{
Compiles traces for age groups. That is the value for total catch at age in each iteration of \code{\link[Reca]{eca.predict}}
}
\details{
The number of iterations of \code{\link[Reca]{eca.predict}}
 is determined by the parameter 'nSamples' to \code{\link[Reca]{eca.estimate}} or \code{\link[RstoxFDA]{runRECA}}
 and the parameter 'caa.burnin' (different from parameter 'burnin') to \code{\link[Reca]{eca.predict}} or \code{\link[RstoxFDA]{runRECA}}

 parameter 'unit' supports:
 \describe{
  \item{number}{Catch at age as number of fish}
  \item{thousands}{Catch at age as number of fish in thousands}
  \item{millions}{Catch at age as number of fish in millions}
  \item{kg}{Catch at age as mass in kilogrammes}
  \item{T}{Catch at age as mass in tons}
  \item{kT}{Catch at age as mass in kilotonnes}
 }
}
