% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecaReport.R
\name{makeResultTableRECA}
\alias{makeResultTableRECA}
\title{Catch at Age result table}
\usage{
makeResultTableRECA(prediction, unit = "millions", plusGroup = NULL,
  alpha = 0.05)
}
\arguments{
\item{prediction}{as returned by \code{\link[Reca]{eca.predict}} or \code{\link[RstoxFDA]{runRECA}}.}

\item{unit}{unit of reported estimates. See details.}

\item{plusGroup}{Fish this age or older will be grouped in report}

\item{alpha}{value for percentiles.}
}
\value{
data.table() with columns:
\describe{
 \item{age}{Age statistics are reported for}
 \item{total}{Estimated total catch of age group}
 \item{unit}{Unit for total catch and standard deviation for estimated total}
 \item{sd}{Standard deviation for estimated total}
 \item{cv}{Coefficient of variation for estimated total}
 \item{lowerQuantile}{Lower quantile (100*alpha/2 percentile) for estimated total}
 \item{UpperQuantile}{Upper quantile: (100*(1-alpha/2) percentile) for estimated total}
 \item{alpha}{alpha for quantiles.}
}
}
\description{
Produces common reported statistics for Catch-At-Age estimation.
}
\details{
parameter 'unit' supports:
 \describe{
  \item{number}{Catch at age as number of fish}
  \item{thousands}{Catch at age as number of fish in thousands}
  \item{millions}{Catch at age as number of fish in millions}
  \item{kg}{Catch at age as mass in kilogrammes}
  \item{T}{Catch at age as mass in tons}
  \item{kT}{Catch at age as mass in kilotonnes}
 }
}
\examples{
 data(recaPrediction)
 makeResultTableRECA(recaPrediction, plusGroup=13)
}
