% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landingsAnalyses.R
\name{makeTripIds}
\alias{makeTripIds}
\title{Make trip IDs}
\usage{
makeTripIds(
  landings,
  vesselIdCol = "Radiokallesignal (seddel)",
  lastCatchCol = "Siste fangstdato"
)
}
\arguments{
\item{landings}{\code{\link[data.table]{data.table}} containing landings records}

\item{vesselIdCol}{character() that identifies a column in 'landings' that contain the vessel id (e.g. radio call signal) of the landing vessel. Default compatible with \code{\link[RstoxData]{readLssFile}}.}

\item{lastCatchCol}{character() that identifies a Date column in 'landings' that contain the date of last catch for each record. Default compatible with \code{\link[RstoxData]{readLssFile}}.}
}
\value{
\code{\link[data.table]{data.table}} with columns 'vesselId', 'time' and 'tripId'
}
\description{
Extracts trip IDs from landings.
}
\examples{
 \dontrun{
   #make trip ids for landings
   lssfile <- "" #set appropriately
   landings <- RstoxData::readLssFile(lssfile)
   tripIds <- makeTripIds(landings)
 }
}
