% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metier4table-datadoc.R
\docType{data}
\name{metier4table}
\alias{metier4table}
\title{MetierTable lvl 4 example}
\format{
\code{\link[RstoxFDA]{MetierTable}} with column 'gearcode' identifying gear codes used in Norwegian fisheries data (derived from NS 9400)
}
\usage{
data(metier4table)
}
\description{
Example of metier table for assigning fishing activity to Metier level 4.
This is not a universal conversion table, but an example of a table made for a particular purpose.
Some gear assignments are done with assumptions, and considering the exact metier codes acceptable by data recipients.
}
\examples{
data(metier4table)
data(activityCensus)
annotated <- appendMetier(activityCensus, metier4table, "gearNS", metierColName = "metier4")
table(annotated$gearFAO, annotated$metier4)
}
\seealso{
Other gear coding functions: 
\code{\link{AddGearGroupStoxBiotic}()},
\code{\link{AddGearGroupStoxLanding}()},
\code{\link{GearGroupFdirTable}},
\code{\link{SetShortGearBiotic}()},
\code{\link{appendMetier}()},
\code{\link{metier5table}},
\code{\link{metier6table}},
\code{\link{readMetierTable}()}
}
\concept{gear coding functions}
\keyword{datasets}
