% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAFOareas-datadoc.R
\docType{data}
\name{NAFOareas}
\alias{NAFOareas}
\title{NAFO areas}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with location names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}}.
}
\usage{
data(NAFOareas)
}
\description{
Definition for FAO Northwest Atlantic (Major Fishing Area 21) area coding system. NAFO areas as used in Norwegian fisheries reporting.
A coding convention for reporting these areas as an extention of the main areas defined by the Norwegian Directorate of Fisheries
identifies the areas in the columns 'StratumName' and 'homr'. International convetion is identified in the column 'nafo_names'.
}
\details{
Polygons are defined for all Divisions of Major Fishing Area 21, except Subarea 21.0, Subarea 21.5 and subarea 21.6.
A single polygon is defined for Subarea 21.0 and 21.5, and two for subarea 21.6.
For subarea 21.6, one polygon comprise Division 21.6.A-C, and one 21.6.D-H

For Division 21.3P and 21.4V a polygon is defined for each subdivision.

Polygons are defined in WGS84 coordinates (unprojected).
}
\examples{
 # plot NAFO areas alongside main areas
 data(mainareaFdir2017)
 data(NAFOareas)
 sp::plot(NAFOareas)
 sp::plot(mainareaFdir2017, add=TRUE)

 # conversion table Norwegian and international convention
 NAFOareas@data[,c("homr", "nafo_names")]
}
\concept{area code polygons}
\keyword{datasets}
