% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ParameterConvergenceData}
\alias{ParameterConvergenceData}
\title{Convergence Report for simulated parameters (ParameterConvergenceData)}
\description{
Convergence Report for multi-chained simulated parameters, 
 such as MCMC simulations with Reca.
 'chains' in this respect refers to statistically independent simulations.
 
 list with one members 'ConvergenceReport',
 which is a \code{\link[data.table]{data.table}} containing the following columns:
 \describe{
  \item{Parameter}{Identifies the parameter that is summarized.}
  \item{InterVariance}{The Mean-Squared-Deviation the means of the parameter in each chain, to the mean across all chains}
  \item{IntraVariance}{The mean of the within-chain variances of the parameter}
  \item{GelmanRubinR}{Gelman-Rubins R}
 }
}
\details{
Gelman-Rubins R is described by Gelman and Rubin (Statistical Science, 1992):
 DOI: https://doi.org/10.1214/ss/1177011136
}
\seealso{
\code{\link[RstoxFDA]{RunRecaModels}} for running Reca-analysis

\code{\link[RstoxFDA]{ReportParameterConvergence}} 
 for creating ParameterConvergenceData.
}
