% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ParameterizationSummaryData}
\alias{ParameterizationSummaryData}
\title{Summary statistics for simulated parameters (ParameterizationSummaryData)}
\description{
Summary statistics for (potentially multi-chained) simulated parameters, 
 such as MCMC simulations with Reca.
 'chains' in this respect refers to statistically independent simulations.
 
 list with two members 'ParameterSummary' and 'RunParameters',
 which both all \code{\link[data.table]{data.table}}s. 'ParameterSummary'
 contains parameter statistics for simulations and have the following columns:
 \describe{
  \item{Parameter}{Identifies the parameter that is summarized.}
  \item{Mean}{Mean of the parameter}
  \item{Variance}{Variance of the parameter}
  \item{chainId}{Identifies the parameterization chain.}
 }
 
 RunParameters summarizes global control parameters for each chain and has columns:
 \describe{
  \item{chainId}{Identifies the parameterization chain.}
  \item{Iterations}{The number of iterations for parameter simulation}
 }
}
\seealso{
\code{\link[RstoxFDA]{RunRecaModels}} for running Reca-analysis

\code{\link[RstoxFDA]{ReportRecaParameterStatistics}} 
 for creating ParameterizationSummaryData from Reca-simulations.
}
