% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalysisFunctions.R
\name{ParameterizeRecaModels}
\alias{ParameterizeRecaModels}
\title{Parameterize Reca.}
\usage{
ParameterizeRecaModels(
  RecaData,
  Nsamples = integer(),
  Burnin = integer(),
  Thin = integer(),
  ResultDirectory = character(),
  Lgamodel = c("log-linear", "non-linear"),
  Delta.age = numeric(),
  Seed = numeric()
)
}
\arguments{
\item{RecaData}{\code{\link[RstoxFDA]{RecaData}} as returned from \code{\link[RstoxFDA]{PrepareRecaEstimate}}}

\item{Nsamples}{number of MCMC samples that will be made available for \code{\link[Reca]{eca.predict}}. See documentation for \code{\link[Reca]{eca.estimate}},}

\item{Burnin}{number of MCMC samples run and discarded by \code{\link[Reca]{eca.estimate}} before any samples are saved. See documentation for \code{\link[Reca]{eca.estimate}}.}

\item{Thin}{controls how many iterations are run between each samples saved. Defaults to 0. This may be set to account for autocorrelation introduced by Metropolis-Hastings simulation. see documentation for \code{\link[Reca]{eca.estimate}}}

\item{ResultDirectory}{a directory where Reca may store temp-files \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}.}

\item{Lgamodel}{The length age relationship to use for length-age fits (options: "log-linear", "non-linear": Schnute-Richards model). See documentation for \code{\link[Reca]{eca.estimate}}.}

\item{Delta.age}{see documentation for \code{\link[Reca]{eca.estimate}}. Defaults to 0.001.}

\item{Seed}{see documentation for \code{\link[Reca]{eca.estimate}}. Defaults to random seed.}
}
\value{
\code{\link[RstoxFDA]{RecaParameterData}} results from Reca Model Parameterization.
}
\description{
Runs estimation of parameters for Reca models. Invokes \code{\link[Reca]{eca.estimate}}.
}
\details{
\code{\link[Reca]{eca.estimate}} performs Markov-chain Monte Carlo (MCMC) simulations to determine maximum likelihood of parameters for the given samples.
 This is computationally intensive and run time may be noticable. For a given model configuration running time is mainly determined by the parameters 'Nsample', 'Burnin' and 'Thin'.
}
\section{temporary files}{

 Various report functions may use output of this function with the function \code{\link[Reca]{eca.predict}} which samples the posterior distributions of parameters.
 Communication between \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}} is managed by writing and reading files, 
 and a directory for storing intermediate calculations must be provided with the parameter 'ResultDirectory'.
 For multi-chain analysis, a different directory should be provided for each chain.
}

\seealso{
\code{\link[RstoxFDA]{PrepareRecaEstimate}} for model configuration, and data preparation for this function, and
 \code{\link[RstoxFDA]{RunRecaModels}} for obtaining predictions / estimates from the Reca-models.
}
