% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{parseLSS}
\alias{parseLSS}
\title{Parses landings (sales notes)}
\usage{
parseLSS(file)
}
\arguments{
\item{file}{file with LSS landings}
}
\value{
data.table with LSS landings
}
\description{
Parses sales notes data from the Norwegian Directorate of Fisheries on the LSS format
}
