% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecaReport.R
\name{plotAgeTraces}
\alias{plotAgeTraces}
\title{Traceplot}
\usage{
plotAgeTraces(
  prediction,
  unit = "millions",
  plusGroup = NULL,
  nclust = 4,
  iter.max = 20,
  nstart = 10,
  agecolors = NULL,
  lowerquant = 0.05,
  upperquant = 0.95,
  catlimit = 8,
  title = ""
)
}
\arguments{
\item{prediction}{as returned by \code{\link[Reca]{eca.predict}} or \code{\link[RstoxFDA]{runRECA}}.}

\item{unit}{unit of traced estimates. See details.}

\item{plusGroup}{Fish this age or older will be grouped in one trace.}

\item{nclust}{the number of plots to distribute the ages and plus group on}

\item{iter.max}{maximal number of iterations for k-means clustering deciding which ages are ploted in same plot.}

\item{nstart}{the number of random sets chosen for the k-means clustering}

\item{agecolors}{named vector matching ages to colors, if null a default color scheme is used}

\item{lowerquant}{lower quantile in each age group to plot as points}

\item{upperquant}{upper quantile in each age group to plot as points}

\item{catlimit}{the upper limit for number of ages in a plot using categorical coloring. Plots with more than this number of ages will use a gradient coloring scheme}

\item{title}{main title for plot}
}
\description{
Plots a paneled plot with traces for age groups. That is the value for total catch at age in each iteration of \code{\link[Reca]{eca.predict}}
}
\details{
The number of iterations of \code{\link[Reca]{eca.predict}}
 is determined by the parameter 'nSamples' to \code{\link[Reca]{eca.estimate}} or \code{\link[RstoxFDA]{runRECA}}
 and the parameter 'caa.burnin' (different from parameter 'burnin') to \code{\link[Reca]{eca.predict}} or \code{\link[RstoxFDA]{runRECA}}

 parameter 'unit' supports:
 \describe{
  \item{number}{Catch at age as number of fish}
  \item{thousands}{Catch at age as number of fish in thousands}
  \item{millions}{Catch at age as number of fish in millions}
  \item{kg}{Catch at age as mass in kilogrammes}
  \item{T}{Catch at age as mass in tons}
  \item{kT}{Catch at age as mass in kilotonnes}
 }

 In order to trade off readability of plots with the number of panels needed. the traces are clustered with a simple clustering algorithm
 and grouped in the same plots accordingly. Adjust clustering parameters, to get fewer or more plots.
}
\examples{
 data(recaPrediction)
 plotAgeTraces(recaPrediction, plusGroup=13, nclust = 6)
}
