% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plotArea}
\alias{plotArea}
\title{Plot area definitions on a map}
\usage{
plotArea(
  data = NULL,
  latCol = NULL,
  lonCol = NULL,
  groupCol = NULL,
  areaDef,
  areaNameCol = "polygonName",
  areaLabels = is.null(data),
  xlim = NULL,
  ylim = NULL,
  areaLabelSize = 2,
  pointcol = "darkred",
  pointshape = 23,
  pointsize = 1,
  title = "",
  projection = 102014
)
}
\arguments{
\item{data}{data.frame with any points to be plotted}

\item{latCol}{character() identifing column in 'data' that specify latitudes (WGS84)}

\item{lonCol}{character() identifing column in 'data' that specify longitudes (WGS84)}

\item{groupCol}{character() identifying column in 'daat' that specify grouping of points}

\item{areaDef}{\code{\link[sp]{SpatialPolygonsDataFrame}}}

\item{areaNameCol}{identifies column in 'areaDef' with label names for the areas}

\item{areaLabels}{logical whether to plot area labels}

\item{xlim}{x axis limits in degrees longitude}

\item{ylim}{y axis limits in degrees latitude}

\item{areaLabelSize}{size for any area labels}

\item{pointcol}{color of any points to be plotted.}

\item{pointshape}{ggplot2 shape for any points to be plotted}

\item{pointsize}{size for any points to be plotted}

\item{title}{plot title}

\item{projection}{proj4string or EPSG code specifying the desired projection, see \code{\link[sf]{st_crs}}}
}
\description{
Plot area definitions on a map
}
\examples{
 # plot ICES areas with default projection
 data(ICESareas)
 plotArea(areaDef=ICESareas)

 data(mainareaFdir2018)
 data(NAFOareas)
 # plot mainarea and NAFO areas combined in mercator projection
 plotArea(title="Main area + NAFO",
          areaDef=rbind(mainareaFdir2018[,c("polygonName")],
                  NAFOareas[,c("polygonName")]),
          projection="+proj=merc +datum=WGS84")
}
