% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plotBubbleMap}
\alias{plotBubbleMap}
\title{Plots bubble plot on map}
\usage{
plotBubbleMap(
  data,
  areaCol,
  quantityCol,
  areaDef,
  areaNameCol = "StratumName",
  legendTitle = quantityCol,
  areaLabels = T,
  xlim = NULL,
  ylim = NULL,
  areaLabelSize = 2,
  bubbleColor = "darkred",
  bubbleSize = 10,
  bubbleShape = 21,
  title = "",
  projection = NULL
)
}
\arguments{
\item{data}{data.frame with quantities to be plotted}

\item{areaCol}{character() identifing column in 'data' that specify area codes, must correspond to 'areaNameCol'}

\item{quantityCol}{character() identifing column in 'data' that specify quantities to be plotted}

\item{areaDef}{\code{\link[sp]{SpatialPolygonsDataFrame}}}

\item{areaNameCol}{identifies column in 'areaDef' with label names for the areas, must correspond to 'areaCol'}

\item{legendTitle}{title for the legend (explains what the quantities are)}

\item{areaLabels}{logical whether to plot area labels}

\item{xlim}{x axis limits in degrees longitude}

\item{ylim}{y axis limits in degrees latitude}

\item{areaLabelSize}{size for any area labels}

\item{bubbleColor}{color of the bubbles}

\item{bubbleSize}{size for the bubbles}

\item{bubbleShape}{shape of the "bubbles". Default to circle (21). Use e.g. 22 for squares (consult list of pch values)}

\item{title}{plot title}

\item{projection}{proj4string or EPSG code specifying the desired projection, see \code{\link[sf]{st_crs}}. Defaults to mercator projection}
}
\description{
Plots scalar nonegative quantities associated with an area code, as bubbles on a map.
}
\details{
The quantities (quantityCol) is aggregated (summed, ignoring NAs) over combinations of area and group ('areaCol' and 'groupCol') (if given)
 Bubble areas are proportional to the quantities.
}
\examples{
 data(landings)
 data(ICESareas)
 plotBubbleMap(landings, "Area", "LiveWeightKG",
       areaDef = ICESareas, areaNameCol = "Area_Full",
       bubbleSize = 20, title="Landings on ICES areas")
}
\seealso{
Other landings functions: 
\code{\link{ReadLandingFDA}()},
\code{\link{ReportFdaLandings}()},
\code{\link{ReportFdaSampling}()},
\code{\link{addLogbookColumns}()},
\code{\link{appendTripIdLandings}()},
\code{\link{imputeCatchesLandings}()},
\code{\link{logbookAdjustment}()},
\code{\link{makeTripIds}()},
\code{\link{sourceLogbookColumns}()},
\code{\link{tabulateFisheries}()}
}
\concept{landings functions}
