% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecaReport.R
\name{plotCatchAtAge}
\alias{plotCatchAtAge}
\title{Catch at Age plot}
\usage{
plotCatchAtAge(prediction, unit = "millions", plusGroup = NULL,
  credibility = 0.95, title = NULL)
}
\arguments{
\item{prediction}{as returned by \code{\link[Reca]{eca.predict}} or \code{\link[RstoxFDA]{runRECA}}.}

\item{unit}{unit of reported estimates. See details.}

\item{plusGroup}{Fish this age or older will be grouped in plot}

\item{credibility}{The desired credibility for credibility intervals.}

\item{title}{Title for plot}
}
\description{
Plots estimated total catch at age for each age group.
 Constructs equal tailed credibility intervals (The probability mass above and below the interval is approximatelty the same).
}
\details{
parameter 'unit' supports:
 \describe{
  \item{number}{Catch at age as number of fish}
  \item{thousands}{Catch at age as number of fish in thousands}
  \item{millions}{Catch at age as number of fish in millions}
  \item{kg}{Catch at age as mass in kilogrammes}
  \item{T}{Catch at age as mass in tons}
  \item{kT}{Catch at age as mass in kilotonnes}
 }
}
\examples{
 data(recaPrediction)
 plotCatchAtAge(recaPrediction)
}
