% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxPlots.R
\name{PlotCatchAtAgeTotals}
\alias{PlotCatchAtAgeTotals}
\title{Plot Catch At Age}
\usage{
PlotCatchAtAgeTotals(ReportFdaCatchAtAgeData)
}
\arguments{
\item{ReportFdaCatchAtAgeData}{\code{\link[RstoxFDA]{ReportFdaCatchAtAgeData}} with catch at age estimates to plot}
}
\value{
\code{\link[RstoxFDA]{PlotCatchAtAgeTotalsData}}
}
\description{
Plots total catch in each age group as a barplot with error bars.
}
\details{
Error bars correspond to the columns 'High' and 'Low' in 'ReportFdaCatchAtAgeData'
 If 'ReportFdaCatchAtAgeData' has grouping variables a bar will be plotted for each group and age-group,
 and bars will be grouped by age group
}
\seealso{
Provide data for this plot with \code{\link[RstoxData]{ReportFdaCatchAtAge}}
}
\concept{StoX-functions}
