% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{PlotCatchAtAgeTotalsData}
\alias{PlotCatchAtAgeTotalsData}
\title{Total catch by age plot (PlotCatchAtAgeTotalsData)}
\description{
a ggplot object that renders estimated total catch by age.
}
\concept{Data types}
