% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxPlots.R
\name{PlotFisheriesOverviewSpatial}
\alias{PlotFisheriesOverviewSpatial}
\title{Plot spatial distribution of fisheries}
\usage{
PlotFisheriesOverviewSpatial(
  ReportFdaLandingData,
  StratumPolygon,
  AreaLabels = F
)
}
\arguments{
\item{ReportFdaLandingData}{\code{\link[RstoxFDA]{ReportFdaLandingData}}}

\item{StratumPolygon}{\code{\link[RstoxBase]{StratumPolygon}}}

\item{AreaLabels}{if TRUE, labels with area codes are plotted on map.}
}
\value{
\code{\link[RstoxFDA]{PlotFisheriesOverviewSpatialData}}
}
\description{
Plots catch density of landings on polygons.
}
\details{
'ReportFdaLandingData' must be configured with only one grouping variable, which must match
 the area names in StratumPolygon$StratumName
}
\seealso{
Provide data for this plot with \code{\link[RstoxFDA]{ReportFdaLandings}}
}
\concept{StoX-Reca functions}
\concept{StoX-functions}
\concept{landings functions}
