% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{PlotFisheriesOverviewSpatialData}
\alias{PlotFisheriesOverviewSpatialData}
\title{Spatial landings overview data (PlotFisheriesOverviewSpatialData)}
\description{
a ggplot object that renders catch density of landings on polygons.
}
\concept{Data types}
