% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxPlots.R
\name{PlotFisheriesOverviewTemporal}
\alias{PlotFisheriesOverviewTemporal}
\title{Plot landings}
\usage{
PlotFisheriesOverviewTemporal(ReportFdaLandingData)
}
\arguments{
\item{ReportFdaLandingData}{\code{\link[RstoxFDA]{ReportFdaLandingData}}}
}
\value{
\code{\link[RstoxFDA]{PlotFisheriesOverviewTemporalData}}
}
\description{
Plots landings by date of catch and by grouping variables.
}
\details{
"CatchDate" must be among the grouping variables of 'ReportFdaLandingData'
}
\seealso{
Provide data for this plot with \code{\link[RstoxFDA]{ReportFdaLandings}}
}
\concept{StoX-Reca functions}
\concept{StoX-functions}
\concept{landings functions}
