% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{PlotFisheriesOverviewTemporalData}
\alias{PlotFisheriesOverviewTemporalData}
\title{Temporal landings overview data (PlotFisheriesOverviewTemporalData)}
\description{
a ggplot object that renders landings by date of catch and by grouping variables.
}
\concept{Data types}
