% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxPlots.R
\name{PlotMeanLengthAtAge}
\alias{PlotMeanLengthAtAge}
\title{Plot length at age}
\usage{
PlotMeanLengthAtAge(ReportFdaLengthAtAgeData)
}
\arguments{
\item{ReportFdaLengthAtAgeData}{\code{\link[RstoxFDA]{ReportFdaLengthAtAgeData}} with mean weight statistics from Reca simulations}
}
\value{
\code{\link[RstoxFDA]{PlotMeanLengthAtAgeData}}
}
\description{
Plot mean individual length for each age group, along with an interval representation for the error of the mean.
 The interval plotted is as configured in 'ReportFdaLengthAtAgeData' (the interval Low / High) and is represented by dashed lines.
 
 The interval does not represent the range of weights occurring in each age group, but the error of the estimate of mean length
 
 If any grouping variables are configured for argument 'ReportFdaLengthAtAgeData', groups will be plotted in different colors.
}
\seealso{
Provide data for this plot with \code{\link[RstoxData]{ReportFdaLengthAtAge}}
}
\concept{StoX-functions}
\concept{convergence-checks}
