% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{PlotMeanLengthAtAgeData}
\alias{PlotMeanLengthAtAgeData}
\title{Mean length at age plot (PlotMeanLengthAtAgeData)}
\description{
a ggplot object that renders estimated mean length by age.
}
\concept{Data types}
