% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxPlots.R
\name{PlotMeanWeightAtAge}
\alias{PlotMeanWeightAtAge}
\title{Plot weight at age}
\usage{
PlotMeanWeightAtAge(ReportFdaWeightAtAgeData)
}
\arguments{
\item{ReportFdaWeightAtAgeData}{\code{\link[RstoxFDA]{ReportFdaWeightAtAgeData}} with mean weight statistics from Reca simulations}
}
\value{
\code{\link[RstoxFDA]{PlotMeanWeightAtAgeData}}
}
\description{
Plot mean individual weight for each age group, along with an interval representation for the error of the mean.
 The interval plotted is as configured in 'ReportFdaWeightAtAgeData' (the interval Low / High) and is represented by dashed lines.
 
 The interval does not represent the range of length occurring in each age group, but the error of the estimate of mean weight
 
 If any grouping variables are configured for argument 'ReportFdaWeightAtAgeData', groups will be plotted in different colors.
}
\seealso{
Provide data for this plot with \code{\link[RstoxData]{ReportFdaWeightAtAge}}
}
\concept{StoX-functions}
\concept{convergence-checks}
