% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{PlotMeanWeightAtAgeData}
\alias{PlotMeanWeightAtAgeData}
\title{Mean weight at age plot (PlotMeanWeightAtAgeData)}
\description{
a ggplot object that renders estimated mean weight by age.
}
\concept{Data types}
