% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxPlots.R
\name{PlotPosteriorTraces}
\alias{PlotPosteriorTraces}
\title{Reca Traceplots}
\usage{
PlotPosteriorTraces(
  RecaCatchAtAge,
  Parameter = c("TotalCatch", "MeanLength", "MeanWeight"),
  PlusGroup = integer(),
  LengthInterval = numeric(),
  UseDefaultPlotSettings = TRUE,
  Nclust = integer(),
  Iter.max = integer(),
  Nstart = integer(),
  LowerQuant = numeric(),
  UpperQuant = numeric(),
  CatLimit = integer(),
  Legend = TRUE
)
}
\arguments{
\item{RecaCatchAtAge}{Results from MCMC simulations (\code{\link[RstoxFDA]{RecaCatchAtAge}}).}

\item{Parameter}{which parameter plot traceplots for "TotalCatch", "MeanLength", or "MeanWeight", Defaults to}

\item{PlusGroup}{If given, ages 'PlusGroup' or older are included in a plus group.}

\item{LengthInterval}{width of length bins in cm, for TotalCatch traceplots. If not provided, length inteval will be set to the maximum length group..}

\item{UseDefaultPlotSettings}{if TRUE default plot settings are used, rather than the values provided by Nclust, Iter.max, Nstart, LowerQuant, UpperQuant, CatLimit, and Legend}

\item{Nclust}{the number of plots to distribute the ages and plus group on. Defaults to 4.}

\item{Iter.max}{maximal number of iterations for k-means clustering deciding which ages are plotted in same plot. Defaults to 20.}

\item{Nstart}{the number of random sets chosen for the k-means clustering. Defaults to 10.}

\item{LowerQuant}{lower quantile in each age group to plot as points. Defaults to .}

\item{UpperQuant}{upper quantile in each age group to plot as points. Defaults to 0.95.}

\item{CatLimit}{the upper limit for number of ages in a plot using categorical coloring. Plots with more than this number of age greoups will use a gradient coloring scheme. Defaults to 30.}

\item{Legend}{If true legend will be plotted.}
}
\value{
\code{\link[RstoxFDA]{PlotPosteriorTracesData}}
}
\description{
Plots the posterior distribution of parameters estimated with Reca.
Useful to corroborate convergence of simulations.

The distribution is plotted in order of iterations, so that issues with autocorrelation in the simulation can be detected.
Autocorrelation issues may be addressed by adjusting the argument 'Thin' to \code{\link[RstoxFDA]{ParameterizeRecaModels}}
The upper and lower quantiles of the distributions are highlighted, so that rare spikes or multi-modalities can be detected.
Rare spikes and multi-modalities can be addressed by adjusting the argument 'Burnin' to \code{\link[RstoxFDA]{ParameterizeRecaModels}}

In order to provide an overview of many age-groups at once. Several panels are created and age-groups that
have closer mean values are plotted together. This is achieved with a k-means clustering (\code{\link[stats]{kmeans}}),
and some key parameters for the clustering algorithm is provided as options Nclust, Iter.max, Nstart,

Any grouping variables or length groups in 'RecaCatchAtAge' are incorporated into the age group definition.
This tends to crowd the plots, and may make them unreadable. While it is desirable to ensure that the parameters has converged
for all ages, grouping variables and length groups, it is often necessary to compromise. One may
\describe{
\item{increase 'LengthInterval'}{Default is to collapse length groups entirely}
\item{adjust 'PlusGroup'}{Reduces the total number of age groups}
\item{adjust 'CatLimit'}{which removes legends from the most crowded plots}
\item{remove grouping variables}{make additional estimates for the same parameterisation, with fewer grouping variables. See: \code{\link[RstoxFDA]{RunRecaModels}}}
}
If 'LengthInterval' specifies only one length-group. Length groups will be removed from plot legends.

Additional convergence checks can be set up using several parameterisation runs and
the functions \code{\link[RstoxFDA]{ReportRecaParameterStatistics}} and \code{\link[RstoxFDA]{ReportParameterConvergence}}.
That analysis checks the convergence of model parameters, rather than the estimated parameters,
and supports handling a large number of model parameters, and filter out indications of non-convergence.
}
\examples{
 RstoxFDA::PlotPosteriorTraces(RstoxFDA::RecaCatchAtAgeExample, PlusGroup = 13)
 RstoxFDA::PlotPosteriorTraces(RstoxFDA::RecaCatchAtAgeExample, PlusGroup = 13,
          Parameter = "MeanWeight")
}
\seealso{
Provide data for this plot with \code{\link[RstoxFDA]{RunRecaModels}}
}
\concept{StoX-functions}
\concept{convergence-checks}
