% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{PlotSamplingCoverageData}
\alias{PlotSamplingCoverageData}
\title{Sampling variables plot data (PlotSamplingCoverageData)}
\description{
a ggplot object that renders a a barplot of landings with color codes for how much each part of the fishery is sampled.
}
\concept{Data types}
