% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{PlotSamplingOverviewCellData}
\alias{PlotSamplingOverviewCellData}
\title{Sampling Overview cell-plot data (PlotSamplingOverviewCellData)}
\description{
a ggplot object that renders the a colored 'cell plot' overview of samples and landings.
}
\concept{Data types}
