% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxPlots.R
\name{PlotSamplingVariables}
\alias{PlotSamplingVariables}
\title{Plot sampling variables}
\usage{
PlotSamplingVariables(
  ReportFdaSamplingData,
  Quantity = c("Catches", "Vessels", "WeightMeasurements", "LengthMeasurements",
    "AgeReadings", "WeightOfSampledCatches"),
  Landings = FALSE
)
}
\arguments{
\item{ReportFdaSamplingData}{\code{\link[RstoxFDA]{ReportFdaSamplingData}} with sampling report to plot}

\item{Quantity}{the sample quantity to plot for each sampling variable: "Catches", "Vessels", "WeightMeasurements", "LengthMeasurements", "AgeReadings", or "WeightOfSampledCatches"}

\item{Landings}{if TRUE total landings in each part of the fishery ('GroupingVariables' in ReportFdaSamplingData) is plotted on a secondary axis.}
}
\value{
\code{\link[RstoxFDA]{PlotSamplingVariablesData}}
}
\description{
Plots a barplot for a sample quantity, 
 illustrating the number of samples that has the sample variables in 'ReportFdaSamplingData'.
 If the fishery is partitioned by grouping variables a stacked barplot is produced.
 Total landings may be plotted on a secondary axis.
}
\concept{StoX-Reca functions}
\concept{StoX-functions}
\concept{landings functions}
