% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{PlotSamplingVariablesData}
\alias{PlotSamplingVariablesData}
\title{Sampling variables plot data (PlotSamplingVariablesData)}
\description{
a ggplot object that renders a stacked barplot of sampling variables for each part of the fishery, with total landings on a secondary axis.
}
\concept{Data types}
