% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locodes2020-datadoc.R
\docType{data}
\name{portcodes2020}
\alias{portcodes2020}
\title{UN Location codes for ports}
\format{
\code{\link[RstoxFDA]{LocodeTable}}.
}
\usage{
data(portcodes2020)
}
\description{
Table of United Nations Codes for Trade and Transport Locations (UN/LOCODE) that has a seaport.
Updated ca 2020.
}
\details{
There is a considerable lag between the national adaptation of new locodes and the updates to the UN/LOCODE lists.
The Norwegian institution that coordinates new needs for LOCODES is the Norwegian Coastal Administration (Kystverket).
}
\examples{
data(portcodes2020)
nocodes <- portcodes2020[portcodes2020$Country=="NO" & !is.na(portcodes2020$latitude),]
RstoxFDA::plotArea(nocodes, "latitude", "longitude", areaDef=RstoxFDA::mainareaFdir2017)
}
\concept{area code polygons}
\keyword{datasets}
