% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxFunctions.R
\name{PrepareReca}
\alias{PrepareReca}
\title{Prepare data for Reca.}
\usage{
PrepareReca(StoxBioticData, StoxLandingData, fixedEffects, randomEffects,
  carEffect = NULL, CarNeighbours = NULL, AgeErrorMatrix = NULL,
  stockSplitting = F, ClassificationErrorMatrix = NULL,
  minAge = NULL, maxAge = NULL, maxLength = NULL,
  lengthResolution = NULL, temporalResolution = c("Quarter", "Month"),
  hatchDate = 1)
}
\arguments{
\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} data with samples from fisheries
and approriate columns appended for identifying corresponding landings.}

\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} data with landings from fisheries
and approriate columns appended for identifying corresponding samples}

\item{fixedEffects}{character() vector identifying column names that should be treated as fixed effects}

\item{randomEffects}{character() vector identifying column names that should be treated as fixed effects}

\item{carEffect}{character(), optional, identifying the column name that should be treated as CAR-effect
(conditional autoregressive effect)}

\item{CarNeighbours}{\code{\link[RstoxFDA]{CarNeighbours}}, mandatory if 'carEffect' is given.
Identifies which values of the carEffect are to be considered as neighbours.}

\item{AgeErrorMatrix}{\code{\link[RstoxFDA]{AgeErrorMatrix}}, optional, specifies the probabilities of misreading ages.
If not provided age errors will not be modelled.}

\item{stockSplitting}{logical(), default FALSE, whether to run estimates for separate stocks in the data (coastal cod-analysis)}

\item{ClassificationErrorMatrix}{\code{\link[RstoxFDA]{ClassificationErrorMatrix}}, optional,
specifies the probability of misclassifying stock for an individual Used in conjunction with 'stockSplitting'. If not provided classification errors will not be modelled.}

\item{minAge}{integer(), optional, must match dimensions of any 'AgeErrorMatrix'.
If not provided it will be derived from data.}

\item{maxAge}{integer(), optional, must match dimensions of any 'AgeErrorMatrix'.
If not provided it will be derived from data.}

\item{maxLength}{numeric(), optional, maximal fish length in data in cm.
If not provided it will be derived from data.}

\item{lengthResolution}{numeric(), optional, resolution for length measurements in cm.
If not provided modal value from data is used.}

\item{temporalResolution}{character(), default "Quarter", code for temporal resolution in landings: "Month" or "Quarter".
Regulates temporal resolution for calculating fractional ages of fish.
Not to be confused with any temporal covariate.}

\item{hatchDay}{integer(), defaults to 1 representing Jan 1st.
encoding the day of the year when fish is consider to transition from one age to the next.}
}
\value{
\code{\link[RstoxFDA]{RecaData}}
}
\description{
Prepare data for Reca.
}
\details{
StoX-function.
 Performs data checks and data conversions,
 and stores some data-related parameters in preparation for running.
 \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}
 via \code{\link[RstoxFDA]{RunReca}}.
}
