% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\docType{data}
\name{processPropertyFormats}
\alias{processPropertyFormats}
\title{Define the process property formats for inclusion in stox UI}
\format{
An object of class \code{list} of length 9.
}
\usage{
processPropertyFormats
}
\description{
Define the process property formats for inclusion in stox UI
}
\keyword{datasets}
