% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{PSUCoInclusionProbabilities}
\alias{PSUCoInclusionProbabilities}
\title{PSU Co-Inclusion probabilities}
\description{
Co-Inclusion probabilites for selection of Primary Sampling Units
}
\details{
Encodes the co-inclusion probabilities for the selection of Primary Sampling Units in multi-stage sampling, used in analytical design based estimation.
 Information is encoded in three tables.
 
 The SampleTable encodes information about the sample of sampling units:
 \describe{
  \item{Stratum}{Mandatory, chr: Identifies the stratum the sample is taken from. Treat unstratified sample as single-stratum sampling (provide only one stratum.}
  \item{N}{Optional, num: The total number of PSUs in Stratum (total available for selection, not total selected)}
  \item{n}{Optional, num: The number of PSUs selected from the Stratum}
  \item{SelectionMethod}{Mandatory, chr: 'Poission', 'FSWR' or 'FSWOR'. The manner of selection for use in bootstrap or inference of inclusionProbabilities, selectionProbabilites, co-inclusion probabilities or co-selection probabilities.}
  \item{FrameDescription}{Optional, chr: Free text field describing the sampling frame.}
 }
 
 The CoSelectionTable encodes information abut the co-inclusion of sampling units for sampling:
 \describe{
  \item{Stratum}{Mandatory, chr: Identifies the stratum the PSU is taken from.}
  \item{SamplingUnitId}{Mandatoryl, chr: Identifes co-selected PSU.}
  \item{SamplingUnitId2}{Mandatory, chr: Identifes co-selected PSU.}
  \item{CoInclusionProbability}{Mandatory, num: The co-inclusion probability of the PSU}
 }
 
 The StratificationVariables table encodes information about which columns in the sampleTable are stratification variables (if any):
 \describe{
  \item{Stratum}{Mandatory, chr: Identifies the stratum. In addition the Stratum is identified by the combination of all other columns on this table.}
  \item{...}{Mandatory if present (may not contain NAs), chr: Additional columns in the sampleTable that are stratification variables.}
 }

Optional columns may be NA.

The selection methods available for 'SelectionMethod' are explained here:
\describe{
 \item{Poission}{Poission sampling. Selection is performed randomly without replacement, and each selection is performed individually. Sample size is not fixed, and 'n' represents the expected sample size.}
 \item{FSWR}{Fixed sample size with replacement. A random selection of a fixed sample size 'n' is chosen with replacement}
 \item{FSWOR}{Fixed sample size with replacement. A random selection of a fixed sample size 'n' is chosen without replacement. Order of selection should be specified in the 'selectionTable'}
}
}
\concept{Analytical estimation}
\concept{Data types}
