% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{PSUSamplingParametersData}
\alias{PSUSamplingParametersData}
\title{PSU Sampling Design Parameters}
\description{
Sampling parameters for selection of Primary Sampling Units
}
\details{
Encodes information about the selection of Primary Sampling Units in multi-stage sampling, used in analytical design based estimation.
 Information is encoded in three tables.
 
 The SampleTable encodes information about the sample of sampling units:
 \describe{
  \item{Stratum}{Mandatory, chr: Identifies the stratum the sample is taken from. Treat unstratified sample as single-stratum sampling (provide only one stratum.}
  \item{N}{Optional, num: The total number of PSUs in Stratum (total available for selection, not total selected)}
  \item{n}{Optional, num: The number of PSUs selected from the Stratum}
  \item{SelectionMethod}{Mandatory, chr: 'Poission', 'FSWR' or 'FSWOR'. The manner of selection for use in bootstrap or inference of inclusionProbabilities, selectionProbabilites, co-inclusion probabilities or co-selection probabilities.}
  \item{FrameDescription}{Optional, chr: Free text field describing the sampling frame.}
 }
 
 The SelectionTable encodes information abut the selection of sampling units for sampling:
 \describe{
  \item{Stratum}{Mandatory, chr: Identifies the stratum the PSU is taken from.}
  \item{Order}{Optional, num: Identifes the order of seleciton. May be necessary for inference when selections are not independent (e.g. FSWOR)}
  \item{SamplingUnitId}{Optional, chr: Identifes PSU. NA encodes non-response}
  \item{InclusionProbability}{Optional, num: The inclusion probability of the PSU}
  \item{HTsamplingWeight}{Optional, num: The normalized Horvitz-Thompson sampling weight of the PSU}
  \item{SelectionProbability}{Optional, num: The selection probability of the PSU}
  \item{HHsamplingWeight}{Optional, num: The normalized Hansen-Hurwitz sampling weight of the PSU}
  \item{SelectionDescription}{Optional, chr: Free text description of the PSU.}
 }
 
 The StratificationVariables table encodes information about which columns in the sampleTable are stratification variables (if any):
 \describe{
  \item{Stratum}{Mandatory, chr: Identifies the stratum. In addition the Stratum is identified by the combination of all other columns on this table.}
  \item{...}{Mandatory if present (may not contain NAs), chr: Additional columns in the sampleTable that are stratification variables.}
 }
 

Optional columns may be NA.

The selection methods available for 'SelectionMethod' are explained here:
\describe{
 \item{Poission}{Poission sampling. Selection is performed randomly without replacement, and each selection is performed individually. Sample size is not fixed, and 'n' represents the expected sample size.}
 \item{FSWR}{Fixed sample size with replacement. A random selection of a fixed sample size 'n' is chosen with replacement}
 \item{FSWOR}{Fixed sample size without replacement. A random selection of a fixed sample size 'n' is chosen without replacement. Order of selection could be specified in the 'selectionTable'}
}

\describe{
  \item{The SelectionProbability is defined as:}{The probability of selecting the sampling unit when it was selected from the population.} 
  \item{The HHsamplingWeight:}{The normalized sampling weight, or the fraction of the stratum represented by the sampled unit when estimating with the Hansen-Hurwitz strategy: 1 / (SelectionProbability*Q) , where Q is the sum of the reciprocal of the SelectionProbabilites for the sampled units. For equal probability sampling with replacement, this is simply 1/n, where n i sample size.}
  \item{The InclusionProbability is defined as:}{The probability of the sampling unit being included in the sample.}
  \item{The HTsamplingWeight:}{The normalized sampling weight, or the fraction of the stratum represented by the sample when estimating with the Horvitz-Thompson strategy: 1 / (InclusionProbability*P), where P is the sum of the reciprocal of the InclusionProbabilites for the sampled units. For equal probability sampling without replacement, this is simply 1/n, where n is sample size.}
}
}
\concept{Analytical estimation}
\concept{Data types}
