% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readFdirLandingsArchive}
\alias{readFdirLandingsArchive}
\title{Reads landings archive}
\usage{
readFdirLandingsArchive(filename, encoding = "Latin-1")
}
\arguments{
\item{filename}{file to read the archive from}

\item{encoding}{encoding of the file identified by filename, must be accepted by \code{\link[data.table]{fread}}.}
}
\value{
\code{\link[RstoxFDA]{LandingsArchiveData}}
}
\description{
Reads aggregated sales notes from archive format deliver by FDIR to IMR. E.g. sluttseddel_1978_2004_medVerdi.csv
}
\concept{IO functions}
