% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readLstFile}
\alias{readLstFile}
\title{Parses logbooks (lst)}
\usage{
readLstFile(filename, encoding = "Latin-1")
}
\arguments{
\item{filename}{file to read the logbook records from}

\item{encoding}{encoding of the file identified by filename, must be accepted by \code{\link[data.table]{fread}}.}
}
\value{
LstLogbookData
}
\description{
Parses logbooks from tabular format (.lst) delivered by Directorate of Fisheries (FDIR).
 This format is not matched with WMS records and contains less detail than the format read by
 \code{\link[RstoxData]{readErsFile}}
}
\details{
A variant of .lst does not provide headers. These are read with assumed standard order of headers,
 and a warning is issued.
}
\seealso{
Other IO functions: 
\code{\link{ReadLandingFDA}()},
\code{\link{readFdirLandingsArchive}()}

Other logbook functions: 
\code{\link{addLogbookColumns}()},
\code{\link{appendTripIdLandings}()},
\code{\link{appendTripIdLogbooks}()},
\code{\link{calculateLogbookPartitionByTrip}()},
\code{\link{convertToErsData}()},
\code{\link{imputeCatchesLandings}()},
\code{\link{logbookAdjustment}()},
\code{\link{makeTripIds}()},
\code{\link{sourceLogbookColumns}()}
}
\concept{IO functions}
\concept{logbook functions}
