% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metierAnnotation.R
\name{readMetierTable}
\alias{readMetierTable}
\title{Read metier table}
\usage{
readMetierTable(filename, encoding = "UTF8")
}
\arguments{
\item{filename}{character() path to file that contains metier definitions. See details for format.}

\item{encoding}{The character encoding of the file identified by 'filename'}
}
\value{
\code{\link[RstoxFDA]{MetierTable}} containing metier definitons.
}
\description{
Reads a table of metier definitions.
}
\details{
The file identified by 'filename' must be a tab-separated file, and must provided headers which must match column names in \code{\link[RstoxFDA]{MetierTable}}

 gearnotation:
 The file format allows a shorthand notation for metiers that are defined the same way for different unmeshed gearcodes or targets.
 These may be written on one line, with the different gear codes separated by commas or different targets separated by commas.

 Optional columns may be omitted. They will be interpreted as NA.
 Comments may be provided on lines with a leading '#'.
 Logical values ('meshedGear' and 'meshedSelectivityDevice') should be encoded with 'T' for true and 'F' for false.
}
