% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{readStockSplittingParamteres}
\alias{readStockSplittingParamteres}
\title{Read Stock splitting parameters from file
 @noRd}
\usage{
readStockSplittingParamteres(resourceFilePath, encoding)
}
\description{
Read Stock splitting parameters from file
 @noRd
}
