% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxFunctions.R
\name{readUnifiedDefinition}
\alias{readUnifiedDefinition}
\title{Read unified categorical definition}
\usage{
readUnifiedDefinition(filepath, encoding = "ascii", formats = NULL)
}
\arguments{
\item{filepath}{path to resource file}

\item{encoding}{encoding of resource file}

\item{formats}{the formats to extract definitions for, if NULL all formats will be extracted.}
}
\value{
list() with one member for each format, each a list mapping codes in that format to the unified variable.
}
\description{
From a resource file, read a definition for a categorical variable that can unify coding systems
 between formats such as NMDlanding and NMDbiotic.
}
\details{
When different data formats encode similar information in with different coding systems
 a unified value can be defined and set in correspondance with one or more codes in the different formats.
 this is done in order to convert coding systems to the unified coding system.

 Definitions are stored with in a tab separated file with headers. Columns defined as:
 \describe{
 \item{Column 1: 'CovariateID' or 'UnifiedValue'}{Unified value (key)}
 \item{Column 2: 'Source' or 'Format'}{The format (key)}
 \item{Column 3: 'Definition'}{A comma separated list of values, any of which will be defined as the unified value in the format.}
 }
}
