% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{RecaCatchAtAge}
\alias{RecaCatchAtAge}
\title{Reca Results (RecaCatchAtAge)}
\description{
Reca Results (RecaCatchAtAge)
}
\details{
a list of data tables:
\describe{
 \item{CatchAtAge}{Tabulating the predicted catch at age by length group for each Reca iteration (MCMC sample)}
 \item{MeanLength}{Tabulating the mean length in cm by age for each Reca iteration (MCMC sample)}
 \item{MeanWeight}{Tabulating the mean weight in g by age for each Reca iteration (MCMC sample)}
}
}
