% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecaCatchAtAgeExample-datadoc.R
\docType{data}
\name{RecaCatchAtAgeExample}
\alias{RecaCatchAtAgeExample}
\title{Saithe landings 2021.}
\format{
\code{\link[RstoxFDA]{RecaCatchAtAge}}
}
\usage{
data(RecaCatchAtAgeExample)
}
\description{
Example of data formatted as \code{\link[RstoxFDA]{RecaCatchAtAgeExample}}.
Data is obtained by running Reca with the data in \code{\link[RstoxFDA]{StoxBioticDataExample}}, and \code{\link[RstoxFDA]{StoxLandingDataExample}}.
}
\examples{
 RstoxFDA::ReportRecaCatchAtAge(RstoxFDA::RecaCatchAtAgeExample, PlusGroup = 13)
}
\concept{StoX-Reca functions}
\keyword{datasets}
