% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecaWrap.R
\name{rEcaDataReport}
\alias{rEcaDataReport}
\title{Data report for R-ECA preparation}
\usage{
rEcaDataReport(samples, landings, covariates)
}
\arguments{
\item{samples}{data.table() with samples (as in \code{\link[RstoxFDA]{prepRECA}}), each row corresponding to one sampled fish. Contains columns:
\describe{
 \item{catchId}{Column identifying the catch that the sample was taken from. Typically a haul or a landing.}
 \item{sampleId}{Column identifying the sample. If only one sample is taken for each catch. This can be set equal to catchId}
 \item{Age}{integer() Age of fish.}
 \item{Length}{numeric() Length of fish. Must be complete (no NAs)}
 \item{Weight}{numeric() Weight of fish.}
 \item{...}{Additional columns which may be used as covariates.}
}}

\item{landings}{data.table() with total landings (as in \code{\link[RstoxFDA]{prepRECA}}), each row corresponding to one cell. Contains columns:
\describe{
 \item{LiveWeightKG}{numeric(). Total landings (Live/Round weight in Kg) for the cell}
 \item{...}{Additional columns which may be used as covariates. These will define each cell.}
}}

\item{covariates}{character() vector of columns to consider for covariates.}
}
\value{
data.table() with columns
\describe{
 \item{<Covariates in landings>}{one column for each. Defines the cells.}
 \item{LiveWeightKG}{The total weight (kg) in the cell.}
 \item{LiveWeightCumFraction}{The fraction of landings in this cell AND all the cells with higher total weight than this cell.}
 \item{<Count of covariates not in landings>}{Count of unique values for covariate. one column for each. Column name is covariate name (from samples) prefixed with N}
 \item{Ncatch}{The number of unique catches sampled in the cell.}
 \item{Nsample}{The number of unique catch-samples in the cell.}
 \item{Nage}{The number of age readings in the cell.}
 \item{Nweight}{The number of fish weight measurements in the cell.}
 \item{Nlength}{The number of fish length measurements in the cell.}
}
}
\description{
Generates overview of samples to inform on sample availability in potential cell definitions.
 Informs on which columns maybe be used as fixed effect covariates, and how grouping of covariates is best done for random effects.

 The covariates in landings define the cells.
 For each cell the covariates defining the cell is reported, before the total landed weight, along with the number of unique occurances of covariates not in landings (including catchId and sampleId).
 Lastly the number of fish measurements for Age, Weight and Length is reported.
}
\examples{
 data(catchsamples)
 catchsamples$catchId <- catchsamples$LEid
 catchsamples$sampleId <- catchsamples$SAid
 catchsamples$date <- catchsamples$LEdate
 catchsamples$Metier5 <- catchsamples$LEmetier5

 data(landings)
 landings$LiveWeightKG <- landings$OfficialLandingsWeight
 landings$Metier5 <- landings$FishingActivityCategoryEuropeanLvl5

 rEcaDataReport(catchsamples, landings, c("Metier5", "VDencrCode"))
}
