% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recaResult-datadoc.R
\docType{data}
\name{recaPrediction}
\alias{recaPrediction}
\title{Example result from Reca}
\format{
see element prediction in return from \code{\link[RstoxFDA]{runRECA}}.
}
\usage{
data(recaPrediction)
}
\description{
Example is produced by \code{\link[RstoxFDA]{runRECA}}
}
\examples{
data(recaPrediction)
plotCatchAtAge(recaPrediction)
}
\seealso{
Other Reca functions: 
\code{\link{catchsamples}},
\code{\link{convertRecaData}()},
\code{\link{getCovariateMap}()},
\code{\link{getLandings}()},
\code{\link{landings}},
\code{\link{makeAgeTracesRECA}()},
\code{\link{makeResultTableRECA}()},
\code{\link{plotAgeTraces}()},
\code{\link{plotCatchAtAge}()},
\code{\link{prepRECA}()},
\code{\link{rEcaDataReport}()},
\code{\link{recaDataExample}},
\code{\link{runRECA}()},
\code{\link{sanitizeRecaInput}()},
\code{\link{tabulateFisheries}()}
}
\concept{Reca functions}
\keyword{datasets}
