% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalyticalReportFunction.R
\name{ReportAnalyticalCatchAtLength}
\alias{ReportAnalyticalCatchAtLength}
\title{Report catch at length}
\usage{
ReportAnalyticalCatchAtLength(
  AnalyticalPopulationEstimateData,
  LengthGroupVariable = character(),
  IntervalWidth = numeric(),
  Decimals = integer(),
  Unit = RstoxData::getUnitOptions("cardinality", conversionRange = c(1, 1e+12))
)
}
\arguments{
\item{AnalyticalPopulationEstimateData}{Results from analytical estimates (\code{\link[RstoxFDA]{AnalyticalPopulationEstimateData}}). A variable identifying length groups (argument: 'LengthGroupVariable') must be among the domain variables. This must be formatted as done by \code{\link[RstoxFDA]{AddLengthGroupStoxBiotic}}.}

\item{LengthGroupVariable}{Name of domain variable in 'AnalyticalPopulationEstimateData' that identifies length group.}

\item{IntervalWidth}{The width of the reported confidence interval. A value of 0.9 gives 90 per cent confidence intervals.}

\item{Decimals}{integer specifying the number of decimals to report for 'CatchAtLength', 'SD', 'Low' and 'High'.}

\item{Unit}{unit for 'CatchAtLength', 'SD', 'Low' and 'High'}
}
\value{
\code{\link[RstoxFDA]{ReportFdaCatchAtLengthData}}
}
\description{
Tabulates summary statistics for analytical catch at length estimate.
 Summary statistics are obtained as analytical domain estimates, including a length-group
 domain, obtained by annotating sample data with the function \code{\link[RstoxFDA]{AddLengthGroupStoxBiotic}}.
 An estimate of the standard deviation of their sampling distribution (the standard error) is also provided.
 Confidence intervals are calculated from a Gaussian approximation to the sampling distribution.
 
 If AnalyticalPopulationEstimateData contains estimates for domains that include more than just length group, such as
 area, gear, stock, etc., summary statistics will be presented similarly.
 
 Rounding of numbers according to the argument 'Decimals' is done with \code{\link[base]{round}},
 so that negative numbers specify rounding to powers of ten, and rounding of the digit 5 is towards the even digit.

 The units considered valid for catch at length in numbers are those listed for quantity 'cardinaltiy' in \code{\link[RstoxData]{StoxUnits}}
}
\seealso{
\code{\link[RstoxFDA]{AddLengthGroupStoxBiotic}} for annotating length groups, \code{\link[RstoxFDA]{AnalyticalPopulationEstimate}} and \code{\link[RstoxFDA]{AnalyticalRatioEstimate}} for obtaining analytical estimates. See \code{\link[RstoxFDA]{ReportAnalyticalCatchAtAge}} for reporting catch at age.
}
\concept{StoX-Reca functions}
\concept{StoX-functions}
