% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalyticalReportFunction.R
\name{ReportAnalyticalLengthAtAge}
\alias{ReportAnalyticalLengthAtAge}
\title{Report mean length at age}
\usage{
ReportAnalyticalLengthAtAge(
  AnalyticalPopulationEstimateData,
  PlusGroup = integer(),
  IntervalWidth = numeric(),
  Decimals = integer(),
  Unit = RstoxData::getUnitOptions("length", conversionRange = c(0.001, 1))
)
}
\arguments{
\item{AnalyticalPopulationEstimateData}{Results from analytical estimates (\code{\link[RstoxFDA]{AnalyticalPopulationEstimateData}}). The StoxBiotic variable 'IndividualAge' must be among the DomainVariables, and estimates must be available for the StoxBiotic-variable 'IndividualTotalLength'.}

\item{PlusGroup}{If given, ages 'PlusGroup' or older are included in a plus group.}

\item{IntervalWidth}{The width of the reported confidence interval. A value of 0.9 gives 90 per cent confidence intervals.}

\item{Decimals}{integer specifying the number of decimals to report for 'MeanIndividualLength', 'SD', 'Low' and 'High'.}

\item{Unit}{unit for 'MeanIndividualLength', 'SD', 'Low' and 'High'}
}
\value{
\code{\link[RstoxFDA]{ReportFdaLengthAtAgeData}}
}
\description{
Tabulates summary statistics for analytical mean total length at age estimates.
 Summary statistics are obtained as analytical domain estimates,
 along with an estimate of the standard deviation of their sampling distribution (the standard error).
 Confidence intervals are calculated from a Gaussian approximation to the sampling distribution.
 
 If AnalyticalPopulationEstimateData contains estimates for domains that include more than just age, such as
 area, gear, stock, etc., summary statistics will be presented similarly.
 
 Rounding of numbers according to the argument 'Decimals' is done with \code{\link[base]{round}},
 so that negative numbers specify rounding to powers of ten, and rounding of the digit 5 is towards the even digit.

 Covariances for means of a variable between domains are not always defined, and the variances (and hence 'SD', 'Low' and 'High') 
 for means in the plusgroup is approximated with an assumption of independence.

 The units considered valid for catch at age in numbers are those listed for quantity 'length' in \code{\link[RstoxData]{StoxUnits}}
}
\seealso{
\code{\link[RstoxFDA]{AnalyticalPopulationEstimate}} and \code{\link[RstoxFDA]{AnalyticalRatioEstimate}} for obtaining analytical estimates.
}
\concept{StoX-Reca functions}
\concept{StoX-functions}
