% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportFdaByAgeData}
\alias{ReportFdaByAgeData}
\title{Age group statistics (ReportFdaByAgeData)}
\description{
Results from catch at age estimations. The results may be presented
 decomposed on combinations of aggregation variables, such as gear, area, stock etc.
 
 ReportFdaByAgeData is a \code{\link[data.table]{data.table}} which may have the following columns:
 \describe{
  \item{AgeGroup}{character. The age group the estimate is reported for. May be age or plus group}
  \item{Age}{integer. The lower age the estimate is reported for. May be an age or lower limit of plus group (inclusive)}
  \item{LengthGroup}{character. The length group the estimate is reported for.}
  \item{Length}{numeric. The upper length of the length group.}
  \item{<Statistic>}{A reported statistic}
  \item{SD}{Standard deviation for the reported statistic.}
  \item{Low}{The lower limit of the estimated interval for the reported statistic.}
  \item{High}{The higher limit of the estimated interval for the reported statistic.}
  \item{...}{Any aggregation variables. The names of these are listed in 'GroupingVariables'}
 }

 Units are configurable, and can be inspected by ~\code{\link[RstoxData]{getUnit}}
}
\seealso{
Other Data types: 
\code{\link{AgeErrorMatrix}},
\code{\link{AreaPosition}},
\code{\link{CarNeighbours}},
\code{\link{KommunePolygons}},
\code{\link{LandingData}},
\code{\link{LandingsArchiveData}},
\code{\link{LengthConversionTable}},
\code{\link{LocodeTable}},
\code{\link{LstLogbookData}},
\code{\link{MetierTable}},
\code{\link{ParameterConvergenceData}},
\code{\link{ParameterizationSummaryData}},
\code{\link{RecaCatchAtAge}},
\code{\link{RecaData}},
\code{\link{RecaParameterData}},
\code{\link{RecaResult}},
\code{\link{ReportFdaCatchAtAgeCovarianceData}},
\code{\link{ReportFdaCatchAtAgeData}},
\code{\link{ReportFdaCatchAtLengthAndAgeData}},
\code{\link{ReportFdaCatchAtLengthData}},
\code{\link{ReportFdaLandingData}},
\code{\link{ReportFdaLengthAtAgeData}},
\code{\link{ReportFdaSamplingData}},
\code{\link{ReportFdaSopData}},
\code{\link{ReportFdaSummaryData}},
\code{\link{ReportFdaWeightAtAgeData}},
\code{\link{StockSplittingParameters}},
\code{\link{TemporalDefinition}},
\code{\link{TripPartition}},
\code{\link{WeightConversionTable}},
\code{\link{is.AgeErrorMatrix}()},
\code{\link{is.AreaPosition}()},
\code{\link{is.CarNeighbours}()},
\code{\link{is.LengthConversionTable}()},
\code{\link{is.ParameterConvergenceData}()},
\code{\link{is.ParameterizationSummaryData}()},
\code{\link{is.RecaCatchAtAge}()},
\code{\link{is.RecaData}()},
\code{\link{is.RecaParameterData}()},
\code{\link{is.RecaResult}()},
\code{\link{is.ReportFdaByAgeData}()},
\code{\link{is.ReportFdaSOP}()},
\code{\link{is.ReportFdaSamplingData}()},
\code{\link{is.StockSplittingParameters}()},
\code{\link{is.TemporalDefinition}()},
\code{\link{is.Translation}()},
\code{\link{is.WeightConversionTable}()}
}
\concept{Data types}
