% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportFdaCatchAtAgeCovarianceData}
\alias{ReportFdaCatchAtAgeCovarianceData}
\title{Fisheries dependent Catch At Age Covariance Report (ReportFdaCatchAtAgeCovarianceData)}
\description{
Covariances from catch at age estimations. Covariances are presented
 for age groups, together with any additional grouping (aggregation variables), 
 such as gear, area, stock etc.
 
 list with two members 'CovarianceNbyAge' and 'Variables'.
 'CovarianceNbyAge' is a \code{\link[data.table]{data.table}} which may have the following columns:
 \describe{
    \item{VariableId1}{Identifies the one of the variables the covariance is calculated for}
    \item{VariableId2}{Identifies the other one of the variables the covariance is calculated for}
    \item{Covariance}{The covariance of catch at age in numbers between groups identified by 'Variable1' and 'Variable2'.}
 }
 \describe{
  \item{VariableId}{Identifier for variable that covariances are provided for.}
  \item{AgeGroup}{character. The age group for the variable. May be age or plus group}
  \item{Age}{integer. The lower age group for the variable. May be an age or lower limit of plus group (inclusive)}
  \item{...}{Any aggregation variables.}
 }

 Units are configurable, and can be inspected by ~\code{\link[RstoxData]{getUnit}}
}
\seealso{
Other Data types: 
\code{\link{AgeErrorMatrix}},
\code{\link{AreaPosition}},
\code{\link{CarNeighbours}},
\code{\link{KommunePolygons}},
\code{\link{LandingData}},
\code{\link{LandingsArchiveData}},
\code{\link{LengthConversionTable}},
\code{\link{LocodeTable}},
\code{\link{LstLogbookData}},
\code{\link{MetierTable}},
\code{\link{ParameterConvergenceData}},
\code{\link{ParameterizationSummaryData}},
\code{\link{RecaCatchAtAge}},
\code{\link{RecaData}},
\code{\link{RecaParameterData}},
\code{\link{RecaResult}},
\code{\link{ReportFdaByAgeData}},
\code{\link{ReportFdaCatchAtAgeData}},
\code{\link{ReportFdaCatchAtLengthAndAgeData}},
\code{\link{ReportFdaCatchAtLengthData}},
\code{\link{ReportFdaLandingData}},
\code{\link{ReportFdaLengthAtAgeData}},
\code{\link{ReportFdaSamplingData}},
\code{\link{ReportFdaSopData}},
\code{\link{ReportFdaSummaryData}},
\code{\link{ReportFdaWeightAtAgeData}},
\code{\link{StockSplittingParameters}},
\code{\link{TemporalDefinition}},
\code{\link{TripPartition}},
\code{\link{WeightConversionTable}},
\code{\link{is.AgeErrorMatrix}()},
\code{\link{is.AreaPosition}()},
\code{\link{is.CarNeighbours}()},
\code{\link{is.LengthConversionTable}()},
\code{\link{is.ParameterConvergenceData}()},
\code{\link{is.ParameterizationSummaryData}()},
\code{\link{is.RecaCatchAtAge}()},
\code{\link{is.RecaData}()},
\code{\link{is.RecaParameterData}()},
\code{\link{is.RecaResult}()},
\code{\link{is.ReportFdaByAgeData}()},
\code{\link{is.ReportFdaSOP}()},
\code{\link{is.ReportFdaSamplingData}()},
\code{\link{is.StockSplittingParameters}()},
\code{\link{is.TemporalDefinition}()},
\code{\link{is.Translation}()},
\code{\link{is.WeightConversionTable}()}
}
\concept{Data types}
