% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportFdaCatchAtAgeData}
\alias{ReportFdaCatchAtAgeData}
\title{Fisheries dependent Catch At Age Report (ReportFdaCatchAtAgeData)}
\description{
A list with two members: 'NbyAge' and 'GroupingVariables'.
 
 \describe{
  \item{NbyAge}{A \code{\link[RstoxFDA]{ReportFdaByAgeData}} table with reported <Statistic> being 'CatchAtAge': the total catch at age in numbers.}
  \item{GroupingVariables}{Any specified Grouping variables.}
 }
 
 Units are configurable, and can be inspected by ~\code{\link[RstoxData]{getUnit}}
 
 'GroupingVariables' is a \code{\link[data.table]{data.table}} with a column containing the names of any aggregation variables.
}
\concept{Data types}
