% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportFdaCatchAtLengthAndAgeData}
\alias{ReportFdaCatchAtLengthAndAgeData}
\title{Fisheries dependent Catch At Age Report (ReportFdaCatchAtLengthAndAgeData)}
\description{
A \code{\link[RstoxFDA]{ReportFdaByAgeData}} object with the reported <Statistic> being:
 
 \describe{
  \item{CacthAtAge}{The total catch at length in numbers.}
 }
 
 Units are configurable, and can be inspected by ~\code{\link[RstoxData]{getUnit}}
}
