% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportFdaData}
\alias{ReportFdaData}
\title{Fisheries dependent analysis report (ReportFdaData)}
\description{
Results from catch estimations. The results may be presented
 decomposed on combinations of aggregation variables, such as gear, area, stock etc.
 
 list with two members 'FdaReport' and 'AggregationVariables'.
 'FdaReport' is a \code{\link[data.table]{data.table}} with the columns:
 \describe{
  \item{AgeGroup}{character. The age group the estimate is reported for. May be age or plus group}
  \item{Age}{integer. The lower age the estimate is reported for. May be an age or lower limit of plus group (inclusive)}
  \item{<Statistic>}{A reported statistic}
  \item{SD}{Standard deviation for the reported statistic.}
  \item{CI.05}{The lower limit of a 90\% Credible Interval or Confidence interval.}
  \item{CI.95}{The higher limit of the 90\% Credible Interval or Confidence interval.}
  \item{...}{Any aggregation variables. The names of these are listed in 'AggregationVariables'}
 }
 'AggregationVariables' is a \code{\link[data.table]{data.table}} with a column containing the names of any aggregation variables.
}
