% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportFdaLandingData}
\alias{ReportFdaLandingData}
\title{Landings Report data (ReportFdaLandingData)}
\description{
list with tow members:
 \describe{
  \item{GroupingVariables}{a \code{\link[data.table]{data.table}} with the variables used for aggregation in 'FisheriesLandings' stored in the column 'GroupingVariables'}
  \item{FisheriesLandings}{a \code{\link[data.table]{data.table}} described below.}
 }
 
 FisheriesLandings is a report of landings for partitions of a fishery.
 The report is a \code{\link[data.table]{data.table}} with columns:
 \describe{
  \item{...}{A column for each of the provided Aggregation variables}
  \item{LandedRoundWeight}{Total landings in kg}
 }
}
