% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportFdaLengthAtAgeData}
\alias{ReportFdaLengthAtAgeData}
\title{Fisheries dependent Length At Age Report (ReportFdaLengthAtAgeData)}
\description{
Results from Reca catch at age estimations. A \code{\link[RstoxFDA]{ReportFdaByAgeData}} object
 with the reported <Statistic> being:
 
 \describe{
  \item{MeanIndividualLength}{The mean length at age.}
 }
 
 Units are configurable, and can be inspected by ~\code{\link[RstoxData]{getUnit}}
 
 Note that the summary statistics are reported for summaries of mean lengths, 
 so that e.g. SD report the standard deviation of the means,
 and does not characterize the length distribution of fish.
}
