% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R, R/StoxReportFunctions.R
\name{ReportFdaSOP}
\alias{ReportFdaSOP}
\title{Sum of Products report (ReportFdaSOP)}
\usage{
ReportFdaSOP(
  ReportFdaCatchAtAgeData,
  ReportFdaWeightAtAgeData,
  StoxLandingData,
  AggregationVariables = NULL
)
}
\arguments{
\item{ReportFdaCatchAtAgeData}{\code{\link[RstoxFDA]{ReportFdaCatchAtAgeData}} with estimates of total catch at age}

\item{ReportFdaWeightAtAgeData}{\code{\link[RstoxFDA]{ReportFdaWeightAtAgeData}} with estimates of mean weight at age for individual fish}

\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} data with landings from fisheries}

\item{AggregationVariables}{Columns of 'StoxLandingData' that partitions the landings into groups SOP tests should be reported for.}
}
\value{
\code{\link[RstoxFDA]{ReportFdaSOP}}
}
\description{
Sum of Products report (SOP-report), comparing the total landed weight of fish
 with the product of mean weight at age estimates and total number 
 at age estimates 
 
 list with two members 'SopReport' and 'AggregationVariables'.
 'SopReport' is a \code{\link[data.table]{data.table}} with the columns:
 \describe{
  \item{TotalWeightEstimated}{Total round weight (kg) estimated}
  \item{LandedWeight}{Landed round weight (kg) reported}
  \item{Difference}{The difference between estimated and reported landed weight}
  \item{RelativeDifference}{The difference between estimated and reported landed weight relative to reported weight}
  \item{...}{Any aggregation variables. The names of these are listed in 'AggregationVariables'}
 }
 'AggregationVariables' is a \code{\link[data.table]{data.table}} with a column containing the names of any aggregation variables.

Report sum-of-product test (SOP-test) for catch estimates

 Mean weight at age and estimated catch at age is used to compute total catches
 and the relative difference to reported landings are reported.
 
 The report will be generated for landings decomposed on the provided 'AggregationVariables'
 which must be available in both 'ReportFdaCatchAtAgeData' and 'ReportFdaWeightAtAgeData'
 and 'StoxLandingData'.
 
 'ReportFdaCatchAtAgeData' and 'ReportFdaWeightAtAgeData' must be decomposed on the same
 'AggregationVariables' and must be reported for the same age groups
}
\seealso{
\code{\link[RstoxFDA]{ReportRecaWeightAtAge}} and \code{\link[RstoxFDA]{ReportRecaCatchAtAge}} for some ways of preparing 'ReportFdaWeightAtAgeData' and 'ReportFdaCatchAtAgeData'.
 \code{\link[RstoxData]{StoxLanding}} and \code{\link[RstoxData]{FilterStoxLanding}} for ways of preparing 'StoxLandingData'.
}
