% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportFdaWeightAtAgeData}
\alias{ReportFdaWeightAtAgeData}
\title{Reca Weight At Age Report (ReportFdaWeightAtAgeData)}
\description{
Results from Reca catch at age estimations. A \code{\link[RstoxFDA]{ReportFdaByAgeData}} object
 with the reported <Statistic> being:
 
 \describe{
  \item{MeanIndividualWeight}{The mean weight at age}
 }
 
 Units are configurable, and can be inspected by ~\code{\link[RstoxData]{getUnit}}
 
 Note that the summary statistics are reported for summaries
 of mean weights, so that e.g. SD report the standard deviation of the means,
 and does not characterize the weight distribution of fish.
}
