% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxReportFunctions.R
\name{ReportRecaCatchAtAge}
\alias{ReportRecaCatchAtAge}
\title{Report catch at age}
\usage{
ReportRecaCatchAtAge(
  RecaCatchAtAge,
  PlusGroup = integer(),
  IntervalWidth = numeric(),
  Decimals = integer(),
  Unit = RstoxData::getUnitOptions("cardinality", conversionRange = c(1, 1e+12))
)
}
\arguments{
\item{RecaCatchAtAge}{Results from MCMC simulations (\code{\link[RstoxFDA]{RecaCatchAtAge}}).}

\item{PlusGroup}{If given, ages 'PlusGroup' or older are included in a plus group.}

\item{IntervalWidth}{The width of the reported credible interval. A value of 0.9 gives 90 per cent credible intervals. Defaults to 0.9}

\item{Decimals}{integer specifying the number of decimals to report for 'CatchAtAge', 'SD', 'Low' and 'High'. Defaults to 0.}

\item{Unit}{unit for 'CatchAtAge', 'SD', 'Low' and 'High'}
}
\value{
\code{\link[RstoxFDA]{ReportFdaCatchAtAgeData}}
}
\description{
Tabulates summary statistics for total catch (number) at age from MCMC simulations using Reca.
MCMC simulations are typically obtained with \code{\link[RstoxFDA]{RunRecaModels}}.
Summary statistics are obtained from the posterior distribution, and
the interval is reported as equal-tailed credible intervals.

If 'RecaCatchAtAge' contains estimate for a set of aggregation variables, such as
area, gear, stock, etc., summary statistics will be presented similarly.

Rounding of numbers according to the argument 'Decimals' is done with \code{\link[base]{round}},
so that negative numbers specify rounding to powers of ten, and rounding of the digit 5 is towards the even digit.

The units considered valid for catch at age in numbers are those listed for quantity 'cardinaltiy' in \code{\link[RstoxData]{StoxUnits}}
}
\examples{
  catchAtAgeReport <- RstoxFDA::ReportRecaCatchAtAge(RstoxFDA::RecaCatchAtAgeExample, 
        PlusGroup = 13)
  catchAtAgeReport$NbyAge
}
\seealso{
\code{\link[RstoxFDA]{RunRecaModels}} for running Reca-analysis and \code{\link[RstoxFDA]{ReportRecaCatchAtLength}} for reporting length composition.
}
\concept{StoX-Reca functions}
\concept{StoX-functions}
