% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxReportFunctions.R
\name{ReportRecaCatchAtAgeCovariance}
\alias{ReportRecaCatchAtAgeCovariance}
\title{Report covariance of catch at age}
\usage{
ReportRecaCatchAtAgeCovariance(
  RecaCatchAtAge,
  PlusGroup = integer(),
  Decimals = integer(),
  Unit = RstoxData::getUnitOptions("cardinality", conversionRange = c(1, 1e+12))
)
}
\arguments{
\item{RecaCatchAtAge}{Results from MCMC simulations (\code{\link[RstoxFDA]{RecaCatchAtAge}}).}

\item{PlusGroup}{If given, ages 'PlusGroup' or older are included in a plus group.}

\item{Decimals}{integer specifying the number of decimals to report for 'Covariance'. Defaults to zero.}

\item{Unit}{unit for 'CatchAtAge'. Covariance will be provided as the square of this unit.}
}
\value{
\code{\link[RstoxFDA]{ReportFdaCatchAtAgeCovarianceData}}
}
\description{
Tabulates the covariance between age groups of (number) at age from MCMC simulations using Reca.
MCMC simulations are typically obtained with \code{\link[RstoxFDA]{RunRecaModels}}.
Covariances are obtained from the posterior distribution.

If 'RecaCatchAtAge' contains estimate for a set of aggregation variables, such as
area, gear, stock, etc., covariances are calculated between age groups for each of these aggregation variables.

Rounding of numbers according to the argument 'Decimals' is done with \code{\link[base]{round}},
so that negative numbers specify rounding to powers of ten, and rounding of the digit 5 is towards the even digit.

The units considered valid for catch at age in numbers are those listed for quantity 'cardinaltiy' in \code{\link[RstoxData]{StoxUnits}}
}
\seealso{
\code{\link[RstoxFDA]{RunRecaModels}} for running Reca-analysis and \code{\link[RstoxFDA]{ReportRecaCatchAtLength}} for reporting length composition.
}
\concept{StoX-Reca functions}
\concept{StoX-functions}
