% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxReportFunctions.R
\name{ReportRecaCatchAtLengthAndAge}
\alias{ReportRecaCatchAtLengthAndAge}
\title{Report catch at length and age}
\usage{
ReportRecaCatchAtLengthAndAge(
  RecaCatchAtAge,
  PlusGroup = integer(),
  LengthInterval = numeric(),
  IntervalWidth = numeric(),
  Decimals = integer(),
  Unit = RstoxData::getUnitOptions("cardinality", conversionRange = c(1, 1e+12))
)
}
\arguments{
\item{RecaCatchAtAge}{Results from MCMC simulations (\code{\link[RstoxFDA]{RecaCatchAtAge}}).}

\item{PlusGroup}{If given, ages 'PlusGroup' or older are included in a plus group.}

\item{LengthInterval}{width of length bins in cm. If not provided, the interval in 'RecaCatchAtAge' will be used.}

\item{IntervalWidth}{The width of the reported credible interval. Defaults to 0.9 for 90 per cent credible intervals.}

\item{Decimals}{integer specifying the number of decimals to report for 'CatchAtAge', 'SD', 'Low' and 'High'. Defaults to zero.}

\item{Unit}{unit for 'CatchAtAge', 'SD', 'Low' and 'High'}
}
\value{
\code{\link[RstoxFDA]{ReportFdaCatchAtLengthAndAgeData}}
}
\description{
Tabulates summary statistics for total catch (number) at all age-length combinations from MCMC simulations using Reca.
MCMC simulations are typically obtained with \code{\link[RstoxFDA]{RunRecaModels}}.
Summary statistics are obtained from the posterior distribution, and
the interval is reported as equal-tailed credible intervals.

Different length groups than the ones reported in the argument 'RecaCatchAtAge'
may be specified with the argument 'IntervalWidth'. This will specify equi-intervalled
lengthgroups with the smallest lengthgroup starting at 0. If it does not align with the
length groups reported in 'RecaCatchAtAge' length group assignment is done to
the highest overlapping length group.

If 'RecaCatchAtAge' contains estimate for a set of aggregation variables, such as
area, gear, stock, etc., summary statistics will be presented similarly.

Rounding of numbers according to the argument 'Decimals' is done with \code{\link[base]{round}},
so that negative numbers specify rounding to powers of ten, and rounding of the digit 5 is towards the even digit.

The units considered valid for catch at length in numbers are those listed for quantity 'cardinaltiy' in \code{\link[RstoxData]{StoxUnits}}
}
\seealso{
\code{\link[RstoxFDA]{RunRecaModels}} for running Reca-analysis and \code{\link[RstoxFDA]{ReportRecaCatchAtAge}} for reporting age composition

Other Reca functions: 
\code{\link{AddStratumStoxLanding}()},
\code{\link{DefineAgeErrorMatrix}()},
\code{\link{DefineStockSplittingParameters}()},
\code{\link{ParameterizeRecaModels}()},
\code{\link{PrepareRecaEstimate}()},
\code{\link{ReportRecaCatchAtAgeCovariance}()},
\code{\link{ReportRecaCatchAtAge}()},
\code{\link{ReportRecaCatchAtLength}()},
\code{\link{ReportRecaCatchStatistics}()},
\code{\link{ReportRecaLengthAtAge}()},
\code{\link{ReportRecaParameterStatistics}()},
\code{\link{ReportRecaWeightAtAge}()},
\code{\link{RunRecaModels}()}
}
\concept{Reca functions}
