% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxReportFunctions.R
\name{ReportRecaCatchStatistics}
\alias{ReportRecaCatchStatistics}
\title{Report catch statistics}
\usage{
ReportRecaCatchStatistics(RecaCatchAtAge, IntervalWidth = numeric())
}
\arguments{
\item{RecaCatchAtAge}{Results from MCMC simulations (\code{\link[RstoxFDA]{RecaCatchAtAge}}).}

\item{IntervalWidth}{The width of the reported credible interval. Defaults to 0.9 for 90 per cent credible intervals.}
}
\value{
\code{\link[RstoxFDA]{ReportFdaSummaryData}}
}
\description{
Report summary statistics for landed catches from MCMC simulations using Reca.
 Summary statistics are obtained from the posterior distribution, and
 the interval is reported as equal-tailed credible intervals are reported.
 
 Weights are reported in kg, lengths in cm.
 
 If 'RecaCatchAtAge' contains estimate for a set of aggregation variables, such as
 area, gear, stock, etc., summary statistics will be presented similarly.
}
\seealso{
\code{\link[RstoxFDA]{RunRecaModels}} for running Reca-analysis
}
