% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportRecaData}
\alias{ReportRecaData}
\title{Reca report (ReportRecaData)}
\description{
Results from Reca catch estimations. The results may be presented
 decomposed on combinations of aggregation variables, such as gear, area, stock etc.
 
 list with two members 'RecaReport' and 'AggregationVariables'.
 'RecaReport' is a \code{\link[data.table]{data.table}} with the columns:
 \describe{
  \item{AgeGroup}{character. The age group the estimate is reported for. May be age or plus group}
  \item{Age}{integer. The lower age the estimate is reported for. May be an age or lower limit of plus group (inclusive)}
  \item{<Statistic>}{A reported statistic. Mean of the posterior distribution.}
  \item{SD}{Standard deviation for the catch at age posterior distribution.}
  \item{CI.05}{The lower limit of the 90\% Credible Interval. That is the 5\% quantile of the posterior distribution.}
  \item{CI.95}{The higher limit of the 90\% Credible Interval. That is the 95\% quantile of the posterior distribution.}
  \item{...}{Any aggregation variables. The names of these are listed in 'AggregationVariables'}
 }
 'AggregationVariables' is a \code{\link[data.table]{data.table}} with a column containing the names of any aggregation variables.
}
