% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportRecaLengthAtAgeData}
\alias{ReportRecaLengthAtAgeData}
\title{Reca Length At Age Report (ReportRecaLengthAtAgeData)}
\description{
Results from Reca catch at age estimations. A \code{\link[RstoxFDA]{ReportRecaData}} object
 with the reported <Statistic> being:
 
 \describe{
  \item{MeanIndividualLength}{The mean length at age in cm.}
 }
 
 Note that the summary statistics are reported for posterior distributions
 of mean lengths, so that e.g. SD report the standard deviation of the means,
 and does not charachterize the length distribution of fish.
}
