% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxReportFunctions.R
\name{ReportRecaParameterStatistics}
\alias{ReportRecaParameterStatistics}
\title{Report summary statistics for Reca paramters}
\usage{
ReportRecaParameterStatistics(
  RecaParameterData,
  ParameterizationSummaryData,
  AppendReport = FALSE
)
}
\arguments{
\item{RecaParameterData}{Simulated Reca parameters}

\item{ParameterizationSummaryData}{summary of Reca parameters that the results should be appended to. Optional.}

\item{AppendReport}{if true, the results are appended to another report provided by 'ParameterizationSummaryData'}
}
\value{
\code{\link[RstoxFDA]{ParameterizationSummaryData}}
}
\description{
Reports means and variances over iterations run for Reca parameterization, which may be used as input to convergence checks.
}
\details{
Multiple chains may be aggregated into one summary table, by repeatedly applying this
function with the aggregated result provided as the argument 'ParameterizationSummaryData'.
This requires that chains are different, and that they are run for the same number of iterations.

Parameters in the summary are identified with the following notation:
\<model name\>-\<covariate name\>-\<any covariate value/level\>:\<any Age group\> \<parameter type\>,
e.g: 'ProportionAtAgeModel-Area:47:Age:2 Intercept ' for
the intercept of age 2 in area 47 in the Proportion-At-Age model.
}
\seealso{
\code{\link[RstoxFDA]{ParameterizeRecaModels}} for model parameterisation
\code{\link[RstoxFDA]{ReportParameterConvergence}} for convergence checks.

Other StoX-Reca functions: 
\code{\link{AddStratumStoxLanding}()},
\code{\link{DefineAgeErrorMatrix}()},
\code{\link{DefineStockSplittingParameters}()},
\code{\link{ParameterizeRecaModels}()},
\code{\link{PrepareRecaEstimate}()},
\code{\link{ReportRecaCatchAtAgeCovariance}()},
\code{\link{ReportRecaCatchAtAge}()},
\code{\link{ReportRecaCatchAtLengthAndAge}()},
\code{\link{ReportRecaCatchAtLength}()},
\code{\link{ReportRecaCatchStatistics}()},
\code{\link{ReportRecaLengthAtAge}()},
\code{\link{ReportRecaWeightAtAge}()},
\code{\link{RunRecaModels}()}

Other StoX-functions: 
\code{\link{AddAreaPositionStoxLanding}()},
\code{\link{AddGearGroupStoxBiotic}()},
\code{\link{AddGearGroupStoxLanding}()},
\code{\link{AddPeriodStoxBiotic}()},
\code{\link{AddPeriodStoxLanding}()},
\code{\link{AddStratumStoxBiotic}()},
\code{\link{AddStratumStoxLanding}()},
\code{\link{ConvertLengthBiotic}()},
\code{\link{ConvertWeightBiotic}()},
\code{\link{DefineAgeErrorMatrix}()},
\code{\link{DefineAreaPosition}()},
\code{\link{DefineCarNeighbours}()},
\code{\link{DefineLengthConversionParameters}()},
\code{\link{DefinePeriod}()},
\code{\link{DefineStockSplittingParameters}()},
\code{\link{DefineWeightConversionFactor}()},
\code{\link{FilterAgeLengthOutliersStoxBiotic}()},
\code{\link{FilterWeightLengthOutliersStoxBiotic}()},
\code{\link{ListBioticDifference}()},
\code{\link{LoadFdaStratumPolygon}()},
\code{\link{ParameterizeRecaModels}()},
\code{\link{PrepareRecaEstimate}()},
\code{\link{ReadLandingFDA}()},
\code{\link{ReportFdaLandings}()},
\code{\link{ReportFdaSOP}()},
\code{\link{ReportFdaSampling}()},
\code{\link{ReportParameterConvergence}()},
\code{\link{ReportRecaCatchAtAgeCovariance}()},
\code{\link{ReportRecaCatchAtAge}()},
\code{\link{ReportRecaCatchAtLengthAndAge}()},
\code{\link{ReportRecaCatchAtLength}()},
\code{\link{ReportRecaCatchStatistics}()},
\code{\link{ReportRecaLengthAtAge}()},
\code{\link{ReportRecaWeightAtAge}()},
\code{\link{RunRecaModels}()},
\code{\link{SetAreaPositionsBiotic}()},
\code{\link{SetShortGearBiotic}()},
\code{\link{SetStartDateBiotic}()},
\code{\link{SetTimeBiotic}()}
}
\concept{StoX-Reca functions}
\concept{StoX-functions}
