% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxReportFunctions.R
\name{ReportRecaWeightAtAge}
\alias{ReportRecaWeightAtAge}
\title{Report weight at age}
\usage{
ReportRecaWeightAtAge(RecaCatchAtAge)
}
\arguments{
\item{RecaCatchAtAge}{Results from MCMC simulations (\code{\link[RstoxFDA]{RecaCatchAtAge}}).}
}
\value{
\code{\link[RstoxFDA]{ReportRecaWeightAtAgeData}}
}
\description{
Tabulates summary statistics for mean weights at age from MCMC simulations using Reca.
 MCMC simulations are typically obtained with \code{\link[RstoxFDA]{RunRecaModels}}.
 
 If 'RecaCatchAtAge' contains estimate for a set of aggregation variables, such as
 area, gear, stock, etc., summary statistics will be presented similarly.
}
