% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportRecaWeightAtAgeData}
\alias{ReportRecaWeightAtAgeData}
\title{Reca Weight At Age Report (ReportRecaWeightAtAgeData)}
\description{
Results from Reca catch at age estimations. A \code{\link[RstoxFDA]{ReportRecaData}} object
 with the reported <Statistic> being:
 
 \describe{
  \item{MeanIndividualWeight}{The mean weight at age in grammes}
 }
 
 Note that the summary statistics are reported for posterior distributions
 of mean weights, so that e.g. SD report the standard deviation of the means,
 and does not charachterize the weight distribution of fish.
}
