% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RstoxFDA-package.R, R/pkgname.R
\docType{package}
\name{RstoxFDA}
\alias{RstoxFDA}
\title{Fisheries Dependent Analysis with RstoX}
\description{
Fisheries dependent analysis, including running the Estimated Catch at Age model through the Reca package developed by the Norwegian Computing Center.

Fisheries dependent analysis, including running the Estimated Catch at Age model through the Reca package developed by the Norwegian Computing Center.
}
\details{
The estimated catch at age (ECA) model uses the correlation structure in fisheries dependent data to distribute age readings from cathes (samples) onto the total reported landings. The ECA model is described in Hirst, D., Aanes, S., Storvik, G., Huseby, R. B., & Tvete, I. F. (2004). Estimating catch at age from market sampling data by using a Bayesian hierarchical model. Journal of the Royal Statistical Society: Series C (Applied Statistics), 53(1), 1-14.

The estimated catch at age (ECA) model uses the correlation structure in fisheries dependent data to distribute age readings from cathes (samples) onto the total reported landings. The ECA model is described in Hirst, D., Aanes, S., Storvik, G., Huseby, R. B., & Tvete, I. F. (2004). Estimating catch at age from market sampling data by using a Bayesian hierarchical model. Journal of the Royal Statistical Society: Series C (Applied Statistics), 53(1), 1-14.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StoXProject/RstoxFDA}
  \item Report bugs at \url{https://github.com/StoXProject/RstoxFDA/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/StoXProject/RstoxFDA}
  \item Report bugs at \url{https://github.com/StoXProject/RstoxFDA/issues}
}

}
\author{
\strong{Maintainer}: Abe Solberg \email{abesolberg@hi.no}

Authors:
\itemize{
  \item Edvin Fuglebakk
  \item Arne Johannes Holmin \email{arnejh@hi.no}
}

}
