% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{SamplingDesignParametersData}
\alias{SamplingDesignParametersData}
\title{Sampling Design Parameters}
\description{
Sampling parameters for selection of a sampling unit
}
\details{
Encodes key information about the selection of a sampling unit, used in analytical design based estimation.
 The selection encoded may be of a obesvered parameter from an individual, or of some intermediate selection in
 hierarchical sampling (e.g. a haul, a vessel, etc.). Information is encoded in three tables.
 
 The sampleTable encodes information about the sample of sampling units:
 \describe{
  \item{Stratum}{Mandatory, chr: Identifies the stratum the sample is taken from. Treat unstratified sample as single-stratum sampling (provide only one stratum.}
  \item{N}{Optional, num: The total number of selection units in Stratum}
  \item{n}{Optional, num: The number of selection units selected from the Stratum}
  \item{SelectionMethod}{Mandatory, chr: 'Poission', 'FSWR' or 'FSWOR'. The manner of selection for use in bootstrap or inference of inclusionProbabilities, selectionProbabilites, co-inclusion probabilities or co-selection probabilities.}
  \item{Finite}{Mandatory, logical: Whether selection is from finite population. That is whether sampling probabilites describes relative or absolute statistical weights.}
  \item{FrameDescription}{Optional, chr: Free text field describing the sampling frame.}
  \item{...}{mandatory, chr: Additional columns are stratification variables. These are mandatory if present (NAs not allowed). They provided are for linking with other data. They must be specified in 'stratificationVariables' and their combination must uniquely identify the Stratum.}
 }
 
 The selectionTable encodes information abut the selection of sampling units for sampling:
 \describe{
  \item{Stratum}{Mandatory: Identifies the stratum the sampling unit is taken from.}
  \item{Order}{Optional: Identifes the order of seleciton. May be necessary for inference when selections are not independent (e.g. FSWOR)}
  \item{SamplingUnitId}{Optional: Identifes sampling unit. NA encodes non-response}
  \item{InclusionProbability}{Optional: The inclusion probability of the sampling unit}
  \item{SelectionProbability}{Optional: The selection probability of the sampling unit}
  \item{SelectionDescription}{Optional: Free text description of sampling unit.}
 }
 
 The stratificationVariables table encodes information about which columns in the sampleTable are stratification variables:
 \describe{
  \item{stratificationVariables}{columns in the sampleTable that are stratification variables.}
 }

Optional columns may be NA.

The selection methods available for 'SelectionMethod' are explained here:
\describe{
 \item{Poission}{Poission sampling. Selection is performed randomly without replacement, and each selection is performed individually. Sample size is not fixed, and 'n' represents the expected sample size.}
 \item{FSWR}{Fixed sample size with replacement. A random selection of a fixed sample size 'n' is chosen with replacement}
 \item{FSWOR}{Fixed sample size with replacement. A random selection of a fixed sample size 'n' is chosen without replacement. Order of selection should be specified in the 'selectionTable'}
}

The selectionProbability is defined as: The probability of selecting the sampling unit in a random draw from the population.
The inclusionProbability is defined as: The probability of the sampling unit being included in the sample.
}
\concept{Analytical estimation}
\concept{Data types}
