% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{SetTimeBiotic}
\alias{SetTimeBiotic}
\title{Set time Biotic}
\usage{
SetTimeBiotic(BioticData, Time = character(), OverWrite = F)
}
\arguments{
\item{BioticData}{\code{\link[RstoxData]{BioticData}} data for which time should be set}

\item{Time}{character encoding time: format as \code{\link[base]{strptime}}-string: \%H:\%M:\%S, e.g. 12:00:00}

\item{Overwrite}{if True any existing values in stationstarttime will be overwritten.}
}
\value{
BioticData \code{\link[RstoxData]{BioticData}}
}
\description{
Set the column 'stationstarttime' on the table 'fishstation' in \code{\link[RstoxData]{BioticData}} to a fixed time
}
\details{
'stationstarttime' is on the table 'fishstation' in data originating from NMDBiotic (http://www.imr.no/formats/nmdbiotic/).
 For bioticdata that does not conform to this, no modifications are done.
 Detting a fixed time to stationstarttime facilitates conversion to \code{\link[RstoxData]{StoxBioticData}} 
 using \code{\link[RstoxData]{StoxBiotic}}, and is applicable if hourly resolution of date and time is not necessary
 for subsequent analysis.
}
