% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landingsAnalyses.R
\name{sourceLogbookColumns}
\alias{sourceLogbookColumns}
\title{Fill in values from logbooks}
\usage{
sourceLogbookColumns(
  landings,
  logbooks,
  tripIdCol = "tripid",
  catchIdCol,
  speciesColLand = "Art FAO (kode)",
  speciesColLog = "FANGSTART_FAO"
)
}
\arguments{
\item{landings}{\code{\link[data.table]{data.table}} containing landings records, and columns identifying the trip and catch ('tripIdCol', 'catchIdCol').}

\item{logbooks}{\code{\link[data.table]{data.table}} containing logbook records, and columns identifying the trip and catch ('tripIdCol', 'catchIdCol').}

\item{tripIdCol}{character() that identifies the column in 'landings' and 'logbooks' that contain trip ids}

\item{catchIdCol}{character() that identifies the column in 'landings' and 'logbooks' that contain catch ids}

\item{speciesColLand}{character() that identifies the column in 'landings' that identify species landed}

\item{speciesColLog}{character() that identifies the column in 'logbooks' that identify species caught.}
}
\description{
Fills selected columns in imputed landings from corresponding logbook records
}
\details{
The following columns in 'landings' may be altered. All others are left untouched:
 \describe{
  \item{Hovedområde (kode)}{set to corresponding value in logbooks}
  \item{Hovedområde}{set to NA for landings with corresponding entries in logbooks}
  \item{Lokasjon (kode)}{set to corresponding value in logbooks}
  \item{Siste fangstdato}{set to corresponding value in logbooks}
  \item{Redskap}{set to corresponding value in logbooks}
 }

 'landings' should already imputed with information about individual catches.
 The columns 'tripIdCol' and 'catchIdCol' must be set on 'landings' and 'logbooks'. This may be achieved by \code{\link[RstoxFDA]{appendTripIdLogbooks}}, \code{\link[RstoxFDA]{appendTripIdLandings}} and \code{\link[RstoxFDA]{imputeCatchesLandings}}.
 
 The function sets values for gear, area codes, and date of last catch from logbooks.
 Landings that does not have corresponding logbook records are left untouched
}
\seealso{
\code{\link[RstoxFDA]{makeTripIds}}, \code{\link[RstoxFDA]{appendTripIdLandings}}, \code{\link[RstoxFDA]{appendTripIdLogbooks}}, and \code{\link[RstoxFDA]{imputeCatchesLandings}} for obtaining 'landings' and 'logbooks' with trip-ids and catch-ids.

Other logbook functions: 
\code{\link{addLogbookColumns}()},
\code{\link{appendTripIdLandings}()},
\code{\link{appendTripIdLogbooks}()},
\code{\link{calculateLogbookPartitionByTrip}()},
\code{\link{convertToErsData}()},
\code{\link{imputeCatchesLandings}()},
\code{\link{logbookAdjustment}()},
\code{\link{makeTripIds}()},
\code{\link{readLstFile}()}

Other landings functions: 
\code{\link{ReadLandingFDA}()},
\code{\link{ReportFdaLandings}()},
\code{\link{ReportFdaSampling}()},
\code{\link{addLogbookColumns}()},
\code{\link{appendTripIdLandings}()},
\code{\link{imputeCatchesLandings}()},
\code{\link{logbookAdjustment}()},
\code{\link{makeTripIds}()},
\code{\link{plotBubbleMap}()},
\code{\link{tabulateFisheries}()}
}
\concept{landings functions}
\concept{logbook functions}
