% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBioticExample-datadoc.R
\docType{data}
\name{StoxBioticDataExample}
\alias{StoxBioticDataExample}
\title{Data from Norwegian port sampling program.}
\format{
\code{\link[RstoxData]{StoxBioticData}}
}
\usage{
data(StoxBioticDataExample)
}
\description{
Example of data formatted as \code{\link[RstoxData]{StoxBioticData}}, with some
additional columns added with \code{\link[RstoxFDA]{AddGearGroupStoxBiotic}}, \code{\link[RstoxFDA]{AddStratumStoxBiotic}}, \code{\link[RstoxFDA]{AddPeriodStoxBiotic}}, and \code{\link[RstoxData]{AddToStoxBiotic}}.
The data contain saithe samples from norwegian port-sampling in 2021.
}
\examples{
 RstoxFDA::plotArea(RstoxFDA::StoxBioticDataExample$Station, 
      areaDef=RstoxFDA::mainareaFdir2018, 
      latCol = "Latitude", 
      lonCol = "Longitude", 
      areaLabels = TRUE)
}
\concept{StoX-Reca functions}
\keyword{datasets}
