% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxLandingExample-datadoc.R
\docType{data}
\name{StoxLandingDataExample}
\alias{StoxLandingDataExample}
\title{Saithe landings 2021.}
\format{
\code{\link[RstoxData]{StoxLandingData}}
}
\usage{
data(StoxLandingDataExample)
}
\description{
Example of data formatted as \code{\link[RstoxData]{StoxLandingData}}, with some
additional columns added with \code{\link[RstoxFDA]{AddGearGroupStoxLanding}}, \code{\link[RstoxFDA]{AddStratumStoxLanding}}, and \code{\link[RstoxFDA]{AddAreaPositionStoxLanding}}, and \code{\link[RstoxFDA]{AddStratumStoxLanding}}, and \code{\link[RstoxFDA]{AddPeriodStoxLanding}}.
The data contain saithe landings by Norwegian vessels in areas (\code{\link[RstoxFDA]{mainareaFdir2018}}) "00", "03", "04", "05","06", in 2021.
}
\examples{
 RstoxFDA::plotBubbleMap(RstoxFDA::StoxLandingDataExample$Landing, 
      "Area", "RoundWeight", RstoxFDA::mainareaFdir2018)
}
\concept{StoX-Reca functions}
\keyword{datasets}
