% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landingsAnalyses.R
\name{tabulateFisheries}
\alias{tabulateFisheries}
\title{Tabulate fisheries}
\usage{
tabulateFisheries(
  data,
  weightCol = "LiveWeightKG",
  cellCols = c("Metier5", "quarter", "Area"),
  complete = F
)
}
\arguments{
\item{data}{data.frame containg fisheries data}

\item{weightCol}{character() column in 'data' that contains catch weights}

\item{cellCols}{character() vector of columns in 'data' defining cells}

\item{complete}{logical() whether all combinations of the columns in cellCols should be tabulated, even if they contain no catch.}
}
\value{
\code{\link[data.table]{data.table}} with the cells specified in cellCols tabulated by decreasing weight and with the columns 'weight', 'frac' and 'cumFrac' containing the weight in each cells and the fraction and cumulative fraction of total weight in that cell.
}
\description{
Tabulates fisheries based on custom cell definitions
}
\details{
A fishery will be decomposed into cells, within which total weight will be reported.
}
\examples{
 data(landings)
 tabulateFisheries(landings)
}
\seealso{
Other landings functions: 
\code{\link{ReadLandingFDA}()},
\code{\link{ReportFdaLandings}()},
\code{\link{ReportFdaSampling}()},
\code{\link{addLogbookColumns}()},
\code{\link{appendTripIdLandings}()},
\code{\link{imputeCatchesLandings}()},
\code{\link{landings}},
\code{\link{logbookAdjustment}()},
\code{\link{makeTripIds}()},
\code{\link{plotBubbleMap}()},
\code{\link{sourceLogbookColumns}()}

Other Reca functions: 
\code{\link{catchsamples}},
\code{\link{convertRecaData}()},
\code{\link{getCovariateMap}()},
\code{\link{getLandings}()},
\code{\link{landings}},
\code{\link{makeAgeTracesRECA}()},
\code{\link{makeResultTableRECA}()},
\code{\link{plotAgeTraces}()},
\code{\link{plotCatchAtAge}()},
\code{\link{prepRECA}()},
\code{\link{rEcaDataReport}()},
\code{\link{recaDataExample}},
\code{\link{recaPrediction}},
\code{\link{runRECA}()},
\code{\link{sanitizeRecaInput}()}
}
\concept{Reca functions}
\concept{landings functions}
