% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{UnifiedVariableDefinition}
\alias{UnifiedVariableDefinition}
\title{Unified Variable Definition (UnifiedVariableDefinition)}
\description{
Table (\code{\link[data.table]{data.table}}) defining a unified variable for different data formats
}
\details{
\describe{
  \item{UnifiedVariable}{Unified code}
  \item{Source}{Format for which the unified variable has corresponding codes}
  \item{Definition}{The codes defining the unified code in the 'source'. Comma-separated list of codes.}
 }
}
