% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AddAreaPositionStoxLanding}
\alias{AddAreaPositionStoxLanding}
\title{Append position to landings data}
\usage{
AddAreaPositionStoxLanding(
  StoxLandingData,
  AreaPosition,
  LocationVariable = c("None", "Location", "Coastal")
)
}
\arguments{
\item{StoxLandingData}{landing data, see \code{\link[RstoxData]{StoxLandingData}}}

\item{AreaPosition}{coordinates for Area and Location codes, see \code{\link[RstoxFDA]{AreaPosition}}}

\item{LocationVariable}{Specify which column in 'StoxLandingsData' should are represented by 'Location' in 'AreaPosition'. See details.}
}
\value{
\code{\link[RstoxData]{StoxLandingData}} with columns for latitude and longitude appended.
}
\description{
StoX function
Appends a position to landings data, based on Area and Location codes.
}
\details{
Positions are appended in the new columns 'Latitude' and 'Longitude'
When 'LocationVariable' is specified as 'None' Area is looked up from 'AreaPosition', using the row where 'Location' is missing.
When 'LocationVariable' is specified as 'Location', 'Area' and 'Location' in 'StoxLandingData' is looked up against 'Area' and 'Location' in 'AreaPosition'.
When 'LocationVariable' is specified as 'Coastal', 'Area' and 'Costal' in 'StoxLandingData' is looked up against 'Area' and 'Location' in 'AreaPosition'.
}
