% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AddGearGroupStoxLanding}
\alias{AddGearGroupStoxLanding}
\title{Add Gear group to StoxLandingData}
\usage{
AddGearGroupStoxLanding(StoxLandingData, Translation)
}
\arguments{
\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} data which will be annotated.}

\item{Translation}{Translation table (\code{\link[RstoxData]{Translation}}). See details.}
}
\value{
\code{\link[RstoxData]{StoxLandingData}} with column 'GearGroup' appended.
}
\description{
Adds a column to StoxLandingData with gear groups
}
\details{
Gear groups are defined as translations from a gear code to a gear group.
Translation need to be defined for the VariableName 'Gear' (see \code{\link[RstoxData]{Translation}}).
The provided Translation should map values ('Gear' in Translation) for the variable 'Gear' in 'StoxBioticData' to a gear group ('NewValue' in Translation).
The gear group will be added to 'StoxLandingData' as the column 'GearGroup'

For comparison or co-analysis with sample data, a \code{\link[RstoxData]{Translation}} should be defined for \code{\link[RstoxData]{StoxBioticData}}
with the same gear groups.
}
\seealso{
\code{\link[RstoxData]{DefineTranslation}} for configuring translation to gear groups,
\code{\link[RstoxFDA]{AddGearGroupStoxBiotic}} for similar function for sample data,
\code{\link[RstoxFDA]{PrepareRecaEstimate}} for use of 'GearGroup' as an effect in Reca-estimation,
and \code{\link[RstoxFDA]{ReportFdaSampling}} for use of 'GearGroup' as an aggregation variable when comparing sampling with landed volume.

Other gear coding functions: 
\code{\link{AddGearGroupStoxBiotic}()},
\code{\link{GearGroupFdirTable}},
\code{\link{SetShortGearBiotic}()},
\code{\link{appendMetier}()},
\code{\link{metier4table}},
\code{\link{metier5table}},
\code{\link{metier6table}},
\code{\link{readMetierTable}()}

Other StoX-functions: 
\code{\link{AddAreaPositionStoxLanding}()},
\code{\link{AddGearGroupStoxBiotic}()},
\code{\link{AddPeriodStoxBiotic}()},
\code{\link{AddPeriodStoxLanding}()},
\code{\link{AddStratumStoxBiotic}()},
\code{\link{AddStratumStoxLanding}()},
\code{\link{ConvertLengthBiotic}()},
\code{\link{ConvertWeightBiotic}()},
\code{\link{DefineAgeErrorMatrix}()},
\code{\link{DefineAreaPosition}()},
\code{\link{DefineCarNeighbours}()},
\code{\link{DefineLengthConversionParameters}()},
\code{\link{DefinePeriod}()},
\code{\link{DefineStockSplittingParameters}()},
\code{\link{DefineWeightConversionFactor}()},
\code{\link{FilterAgeLengthOutliersStoxBiotic}()},
\code{\link{FilterWeightLengthOutliersStoxBiotic}()},
\code{\link{ListBioticDifference}()},
\code{\link{LoadFdaStratumPolygon}()},
\code{\link{ParameterizeRecaModels}()},
\code{\link{PrepareRecaEstimate}()},
\code{\link{ReadLandingFDA}()},
\code{\link{ReportFdaLandings}()},
\code{\link{ReportFdaSOP}()},
\code{\link{ReportFdaSampling}()},
\code{\link{ReportParameterConvergence}()},
\code{\link{ReportRecaCatchAtAgeCovariance}()},
\code{\link{ReportRecaCatchAtAge}()},
\code{\link{ReportRecaCatchAtLengthAndAge}()},
\code{\link{ReportRecaCatchAtLength}()},
\code{\link{ReportRecaCatchStatistics}()},
\code{\link{ReportRecaLengthAtAge}()},
\code{\link{ReportRecaParameterStatistics}()},
\code{\link{ReportRecaWeightAtAge}()},
\code{\link{RunRecaModels}()},
\code{\link{SetAreaPositionsBiotic}()},
\code{\link{SetShortGearBiotic}()},
\code{\link{SetStartDateBiotic}()},
\code{\link{SetTimeBiotic}()}
}
\concept{StoX-functions}
\concept{gear coding functions}
