% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AddGearGroupStoxLanding}
\alias{AddGearGroupStoxLanding}
\title{Add Gear group to StoxLandingData}
\usage{
AddGearGroupStoxLanding(StoxLandingData, Translation)
}
\arguments{
\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} data which will be annotated.}

\item{Translation}{Translation table (\code{\link[RstoxData]{Translation}}). See details.}
}
\value{
\code{\link[RstoxData]{StoxLandingData}} with column 'GearGroup' appended.
}
\description{
Adds a column to StoxLandingData with gear groups
}
\details{
Gear groups are defined as translations from a gear code to a gear group.
 The provided Translation should map values ('Value' in Translation) for the variable 'Gear' in 'StoxBioticData' ('VariableName' in Translation) to a gear group ('NewValue' in Translation).
 The gear group will be added to 'StoxLandingData' as the column 'GearGroup'
 
 For comparison or co-analysis with sample data, a \code{\link[RstoxData]{Translation}} should be defined for \code{\link[RstoxData]{StoxBioticData}}
 with the same gear groups.
}
\seealso{
\code{\link[RstoxData]{DefineTranslation}} for configuring translation to gear groups, 
 \code{\link[RstoxFDA]{AddGearGroupStoxBiotic}} for similar function for sample data, 
 \code{\link[RstoxFDA]{PrepareRecaEstimate}} for use of 'GearGroup' as an effect in Reca-estimation,
 and \code{\link[RstoxFDA]{ReportFdaSampling}} for use of 'GearGroup' as an aggregation variable when comparing sampling with landed volume.
}
