% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AddPeriodStoxLanding}
\alias{AddPeriodStoxLanding}
\title{Add Period to StoxLandingData}
\usage{
AddPeriodStoxLanding(StoxLandingData, TemporalDefinition)
}
\arguments{
\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} data which will be annotated.}

\item{TemporalDefinition}{\code{\link[RstoxFDA]{TemporalDefinition}} definiton of temporal category.}
}
\value{
StoxLandingData with column appended. See \code{\link[RstoxData]{StoxLandingData}}.
}
\description{
Add a column 'Period' to StoxLandingData with a temporal category, such as 'quarter'.
}
\details{
Temporal definitions (\code{\link[RstoxFDA]{TemporalDefinition}}) may be produced by
 \code{\link[RstoxFDA]{DefinePeriod}}
}
