% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AddStratumStoxBiotic}
\alias{AddStratumStoxBiotic}
\title{Adds Stratum to StoxBioticData}
\usage{
AddStratumStoxBiotic(StoxBioticData, StratumPolygon)
}
\arguments{
\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} data which will be annotated. Needs postions appended. See details.}

\item{StratumPolygon}{Definition of spatial strata. See \code{\link[RstoxBase]{StratumPolygon}}}
}
\value{
StoxBioticData with column appended to data.table 'Station'. See \code{\link[RstoxData]{StoxBioticData}}.
}
\description{
Adds a column 'Stratum' to StoxBioticData with the spatial strata each row belongs to.
}
\details{
The strata are added to the new column 'Stratum'.
}
\seealso{
\code{\link[RstoxBase]{DefineStratumPolygon}} for configuring stratum definitions.
\code{\link[RstoxFDA]{AddStratumStoxLanding}} for similar function for landing data,
\code{\link[RstoxFDA]{PrepareRecaEstimate}} for use of 'Stratum' as an effect in Reca-estimation,
\code{\link[RstoxFDA]{DefineCarNeighbours}} for obtaining a neighbour-definition for using 'Stratum' as CAR-effect in Reca-estimation.
and \code{\link[RstoxFDA]{ReportFdaSampling}} for use of 'Stratum' as an aggregation variable when comparing sampling with landed volume.
}
