% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AddStratumStoxBiotic}
\alias{AddStratumStoxBiotic}
\title{Adds Stratum to StoxBioticData}
\usage{
AddStratumStoxBiotic(StoxBioticData, StratumPolygon)
}
\arguments{
\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} data which will be annotated. Needs postions appended. See details.}

\item{StratumPolygon}{Definition of spatial strata. See \code{\link[RstoxBase]{StratumPolygon}}}
}
\value{
StoxBioticData with column appended to data.table 'Station'. See \code{\link[RstoxData]{StoxBioticData}}.
}
\description{
Adds a column 'Stratum' to StoxBioticData with the spatial strata each row belongs to.
}
\details{
The strata are added to the new column 'Stratum'.
}
\seealso{
\code{\link[RstoxBase]{DefineStratumPolygon}} for configuring stratum definitions.
\code{\link[RstoxFDA]{AddStratumStoxLanding}} for similar function for landing data,
\code{\link[RstoxFDA]{PrepareRecaEstimate}} for use of 'Stratum' as an effect in Reca-estimation,
\code{\link[RstoxFDA]{DefineCarNeighbours}} for obtaining a neighbour-definition for using 'Stratum' as CAR-effect in Reca-estimation.
and \code{\link[RstoxFDA]{ReportFdaSampling}} for use of 'Stratum' as an aggregation variable when comparing sampling with landed volume.

Other spatial coding functions: 
\code{\link{AddAreaPositionStoxLanding}()},
\code{\link{AddStratumStoxLanding}()},
\code{\link{DefineAreaPosition}()},
\code{\link{DefineCarNeighbours}()},
\code{\link{LoadFdaStratumPolygon}()},
\code{\link{SetAreaPositionsBiotic}()},
\code{\link{appendAreaCode}()},
\code{\link{appendPosition}()},
\code{\link{areaCodeConversionTable}()},
\code{\link{mergePolygons}()},
\code{\link{plotAreaComparison}()},
\code{\link{plotArea}()},
\code{\link{writeSpDataFrameAsWKT}()}

Other StoX-functions: 
\code{\link{AddAreaPositionStoxLanding}()},
\code{\link{AddGearGroupStoxBiotic}()},
\code{\link{AddGearGroupStoxLanding}()},
\code{\link{AddPeriodStoxBiotic}()},
\code{\link{AddPeriodStoxLanding}()},
\code{\link{AddStratumStoxLanding}()},
\code{\link{ConvertLengthBiotic}()},
\code{\link{ConvertWeightBiotic}()},
\code{\link{DefineAgeErrorMatrix}()},
\code{\link{DefineAreaPosition}()},
\code{\link{DefineCarNeighbours}()},
\code{\link{DefineLengthConversionParameters}()},
\code{\link{DefinePeriod}()},
\code{\link{DefineStockSplittingParameters}()},
\code{\link{DefineWeightConversionFactor}()},
\code{\link{FilterAgeLengthOutliersStoxBiotic}()},
\code{\link{FilterWeightLengthOutliersStoxBiotic}()},
\code{\link{ListBioticDifference}()},
\code{\link{LoadFdaStratumPolygon}()},
\code{\link{ParameterizeRecaModels}()},
\code{\link{PrepareRecaEstimate}()},
\code{\link{ReadLandingFDA}()},
\code{\link{ReportFdaLandings}()},
\code{\link{ReportFdaSOP}()},
\code{\link{ReportFdaSampling}()},
\code{\link{ReportParameterConvergence}()},
\code{\link{ReportRecaCatchAtAgeCovariance}()},
\code{\link{ReportRecaCatchAtAge}()},
\code{\link{ReportRecaCatchAtLengthAndAge}()},
\code{\link{ReportRecaCatchAtLength}()},
\code{\link{ReportRecaCatchStatistics}()},
\code{\link{ReportRecaLengthAtAge}()},
\code{\link{ReportRecaParameterStatistics}()},
\code{\link{ReportRecaWeightAtAge}()},
\code{\link{RunRecaModels}()},
\code{\link{SetAreaPositionsBiotic}()},
\code{\link{SetShortGearBiotic}()},
\code{\link{SetStartDateBiotic}()},
\code{\link{SetTimeBiotic}()}
}
\concept{StoX-functions}
\concept{spatial coding functions}
