% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{AddStratumStoxLanding}
\alias{AddStratumStoxLanding}
\title{Adds Stratum to StoxLandingData}
\usage{
AddStratumStoxLanding(StoxLandingData, StratumPolygon)
}
\arguments{
\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} data which will be annotated. Needs postions appended. See details.}

\item{StratumPolygon}{Definition of spatial strata. See \code{\link[RstoxBase]{StratumPolygon}}}

\item{columnName}{Name of the column that will be appended. Defaults to 'Stratum'.}
}
\value{
StoxLandingData with column appended. See \code{\link[RstoxData]{StoxLandingData}}.
}
\description{
StoX function
 Adds a column to StoxLandingData with the spatial strata each row belongs to.
}
\details{
The strata are added to the new column 'Stratum'.
 \code{\link[RstoxData]{StoxLandingData}} does not contain columns for positions,
 these need to be added as columns 'Latitude' and 'Longitude' before calling this function.
 \code{\link[RstoxFDA]{AddAreaPositionStoxLanding}} may be used to append positions, based on area codes.
}
\seealso{
\code{\link[RstoxFDA]{AddAreaPositionStoxLanding}} for appending positions to \code{\link[RstoxData]{StoxLandingData}}.
}
