% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxFunctions.R
\name{AgeErrorMatrix}
\alias{AgeErrorMatrix}
\title{Age Error Matrix (AgeErrorMatrix)}
\description{
Table (\code{\link[data.table]{data.table}})
defining probabilities of misreading age.
}
\details{
\describe{
  \item{columns 1..n}{numeric() [0,1]. Probability of reading read age, given that true age is as column name.}
  \item{ReadAge}{The read age.}
 }

 Columns sum to 1.
}
