% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalyticalBaselineFunctions.R
\name{AnalyticalPopulationEstimate}
\alias{AnalyticalPopulationEstimate}
\title{Analytical estimate of population parameters}
\usage{
AnalyticalPopulationEstimate(
  PSUSamplingParametersData,
  AnalyticalPSUEstimateData,
  MeanOfMeans = F
)
}
\arguments{
\item{PSUSamplingParametersData}{\code{\link[RstoxFDA]{PSUSamplingParametersData}} with sampling parameters for a sample of Primary Samplig Units.}

\item{AnalyticalPSUEstimateData}{\code{\link[RstoxFDA]{AnalyticalPSUEstimateData}} with estimates for each of the Primary Sampling Units in PSUSamplingParametersData}

\item{MeanOfMeans}{logical. Determines which estimators are used for frequencies and means. See details.}
}
\value{
\code{\link[RstoxFDA]{AnalyticalPopulationEstimateData}} with estimated population parameters by stratum and domain.
}
\description{
Provides analytical estimates of population parameters and their co-variances from a multi-stage sampling design.
Estimates and co-variance estimates are provided for abundance, frequencies, totals, and means, by domain and strata.
}
\details{
Estimates are provided with the Hansen-Hurwitz estimator, given estimates for each Primary Sampling Unit (PSU),
and the design parameters for the PSU selection. See \code{\link[RstoxFDA]{AnalyticalPSUEstimate}} for how to obtain estimates for each PSU.

Estimation of totals and abundance require selection probabilites to be known. In cases where only sampling weights are known,
the option 'MeanOfMeans' provides estimation of means and frequencies, but not abundance and totals. These will be set to NA in such cases.

Means may be unknown for a certain domain for two reasons. They estimate of abundance and frequencies for the domain may be zero.
In this case unknown means are provded as NaNs, as they result from dividing by zero.
If either PSU means sampling parameters are unkown, or a the strata is not sampled, means will be provided as NA.

The stratification incorporates both stratified estimates for each PSU and any stratified selection of PSUs.
E.g. each PSU may provide estimates by length strata, and at the same time the selection of PSUs may be stratified by area.
In that case this function would return estimates for each combination of length stratum and area.
This requires all PSUs to provide estimates for the same strata, and the functions halts with error if that is not the case.
See \code{\link[RstoxFDA]{LiftStrata}} for a way to infer PSU-estimates for strata that have zero abundance.
If simpler stratification is desired, see \code{\link[RstoxFDA]{CollapseStrata}}.

The domains will be the combination of the domain the PSU belongs to (PSU-domains), and
the domains defined in estimation for later sampling stages and encoded in 'AnalyticalPSUEstimateData'.
Consider the arguments to \code{\link[RstoxFDA]{AnalyticalPSUEstimate}} if other
later stage domains are desired. For variance estimation, it will sometimes be necessary
to distinguish the domain size of the PSU-domains. In these cases we will refer to the
PSU-domain of a domain. For example, catches may be PSUs and a domain will be defined as 'age 5 fish caught with
purse seine'. The gear (Purse seine) is a property of the catch and is naturally specified as a
PSU-domain, so the PSU-domain of "age 5 fish caught with purse seine" is "purse seine". The domain will be
the total number of PSUs in the population that caught age 5 fish with purse seine, while the domain size of
the PSU domain will be simply all purse seine catches in the population. Hence domain sizes are always smaller
than their corresponding PSU-domain sizes.

In general unbiased estimates rely on known selection probabilites, and domain definitions that coincides
with stratification. When only sampling weights are known, or the domain definitions are not aligned
with the stratification, ratio estimates are provided for which unbiasedness is not guaranteed.

Only between-PSU variance is accounted for, ignoring the variance contribution from the later sampling stages.
This provides and unbiased estimate of the co-variances when PSUs are selected completely independently.
This is the case for Poission sampling or sampling with replacement. It is also approximately true for
regular sampling without replacement when the sampling intensity is low (Only a small fraction of the population PSUs are sampled).

For the quantities abundance, frequency of individuals, and the total and mean of variables, an estimate \eqn{\hat{x}} and an estimate of sampling co-variances \eqn{\widehat{CoVar}(\hat{x},\hat{y})}
is provided. Variances may be extracted as \eqn{\widehat{Var}(\hat{x})=\widehat{CoVar}(\hat{x},\hat{x})}.
Some other useful error statistics may be derived, such as standard error estimates as \eqn{\widehat{SE}(\hat{x})=\sqrt{\widehat{Var}(\hat{x})}} and the coefficient of variation as \eqn{CV(\hat{x})=\frac{\widehat{SE}(\hat{x})}{\hat{x}}}.
All estimates are normally distributed for large enough sample sizes (asymptotically). When sample size is large enough,
confidence intervals for estimates can be deduced from a normal distribution with \eqn{\mu=\hat{x}} and \eqn{\sigma=\hat{SE}({\hat{x}})}.

Abundances, frequencies, totals, and means are estimated with the formulas below. A vocabulary of notation is provided after the equations.
Estimates of PSU domain sizes are provided without variance-estimations. These are also documented in the vocabulary below.

\describe{
\item{Abundance:}{
The estimate of the total number of individuals in domain \eqn{d} and stratum \eqn{s}:
\deqn{\hat{N}^{(s,d)} = \frac{1}{n^{(s)}} \sum_{i=1}^{n} \hat{D}^{(s,d)}\frac{\hat{N}^{(s,d)}_{i}}{p_{i}}I^{(s,d)}_{i}}
with co-variance:
\deqn{\widehat{CoVar}(\hat{N}^{(s,d_{1})}, \hat{N}^{(s,d_{2})}) = \frac{1}{\hat{P}^{(s,d_{1})}\hat{P}^{(s,d_{2})}}\frac{1}{n^{(s)}(n^{(s)}-1)} \sum_{i=1}^{n} 
  I^{(s,d_{1})}_{i} I^{(s,d_{2})}_{i} (\hat{D}^{(s,d)}\frac{\hat{N}^{(s,d_{1})}_{i}}{p_{i}}I^{(s,d_{1})}_{i} - \hat{N}^{(s,d_{1})}) (\hat{D}^{(s,d)}\frac{\hat{N}^{(s,d_{2})}_{i}}{p_{i}}I^{(s,d_{2})}_{i} - \hat{N}^{(s,d_{2})})}

Note that unless \eqn{\hat{P}^{(s,d_{1})}=\hat{P}^{(s,d_{2})}} the covariance is zero.

In the general case \eqn{\hat{D}^{(s,d)}} and \eqn{\hat{P}^{(s,d)}} is estimated by a ratio estimator, and the error those estimates are ignored. When \eqn{d} covers all of \eqn{s}, \eqn{\hat{D}^{(s,d)}=1} and \eqn{\hat{P}^{(s,d)}=1} is known.
In this case both expressions can be shown to be unbiased. In addition the estimate may depend on any ratio estimation for \eqn{\hat{N}^{(s,d)}}. See \code{\link[RstoxFDA]{AnalyticalPSUEstimate}}.

These quantities can only be calculated when \eqn{p_{i}} is provided, and will otherwise be NA.}

\item{Frequency:}{
The estimate of the fraction of individuals in stratum \eqn{s} that are in domain \eqn{d}, when MeanOfMeans is false:
\deqn{ \hat{f}^{(s,d)} = \frac{\hat{N}^{(s,d)}}{\hat{N}^{(s)}} }
with co-variance:
\deqn{ \widehat{CoVar}(\hat{f}^{(s,d_{1})}, \hat{f}^{(s,d_{2})}) = \frac{1}{(\hat{N}^{(s)})^{2}}\widehat{CoVar}(\hat{N}^{(s,d_{1})}, \hat{N}^{(s,d_{2})})}

These are ratio estimates depending on the ratio to the estimate value \eqn{\hat{N}^{(s)}}, and the error in this estimate is ignored.
In addition, the estimate may depend on a ratio estimate for \eqn{\hat{N}^{(s,d)}} and \eqn{\hat{N}^{(s,d)}_{i}}, as explained for 'Abundance'
and in \code{\link[RstoxFDA]{AnalyticalPSUEstimate}}.
}

\item{Frequency, Mean of Means:}{
The estimate of the fraction of individuals in stratum \eqn{s} that are in domain \eqn{d}, when MeanOfMeans is true:
\deqn{ \hat{f}^{(s,d)} =  \sum_{i=1}^{n}\frac{w_{i}}{\hat{D}^{(s,d)}}\hat{f}^{(s,d)}_{i}I^{(s,d)}_{i}}
with co-variance:
\deqn{\widehat{CoVar}(\hat{f}^{(s,d_{1})}, \hat{f}^{(s,d_{2})}) = \frac{1}{\hat{P}^{(s,d_{1})}\hat{P}^{(s,d_{2})}}\frac{1}{n^{(s)}(n^{(s)}-1)} \sum_{i=1}^{n} 
  I^{(s,d_{1})}_{i} I^{(s,d_{2})}_{i} (\hat{f}_{i}I^{(s,d_{1})}_{i} - \hat{f}^{(s,d_{1})}) (\hat{f}_{i}I^{(s,d_{2})}_{i} - \hat{f}^{(s,d_{2})})}.}

Note that unless \eqn{\hat{P}^{(s,d_{1})}=\hat{P}^{(s,d_{2})}} the covariance is zero.

These are ratio estimates depending on the ratio ratio estimation of \eqn{\hat{f}^{(s,d)}}, \eqn{\hat{f}^{(s,d)}_{i}}, \eqn{\hat{D}^{(s,d)}}, and \eqn{\hat{P}^{(s,d)}}, and the error in these estimates are ignored.
See comments above (\eqn{\hat{f}^{(s,d)}}) and in \code{\link[RstoxFDA]{AnalyticalPSUEstimate}} (\eqn{\hat{f}^{(s,d)}_{i}}).

\item{Total:}{
The estimate of the total value of some variable \eqn{v} in domain \eqn{d} and stratum \eqn{s}.
\deqn{\hat{t}^{(s,d,v)}=\frac{1}{n^{(s,d)}}{\sum_{i=1}^{n}}\hat{D}^{(s,d)}\frac{\hat{t}^{(s,d,v)}_{i}}{p_{i}}I^{(s,d)}_{i}}
with co-variance:
\deqn{\widehat{CoVar}(\hat{t}^{(s,d_{1},v_{1})}, \hat{t}^{(s,d_{2},v_{2})}) = \frac{1}{\hat{P}^{(s,d_{1})}\hat{P}^{(s,d_{2})}}\frac{1}{n^{(s)}(n^{(s)}-1)} \sum_{i=1}^{n} 
  I^{(s,d_{1})}_{i} I^{(s,d_{2})}_{i}(\hat{D}^{(s,d)}\frac{\hat{t}^{(s,d_{1},v_{1})}_{i}}{p_{i}}I^{(s,d_{1})}_{i} - \hat{t}^{(s,d_{1},v_{1})}) (\hat{D}^{(s,d)}\frac{\hat{t}^{(s,d_{2},v_{2})}_{i}}{p_{i}}I^{(s,d_{2})}_{i} - \hat{t}^{(s,d_{2},v_{2})})}.

Note that unless \eqn{\hat{P}^{(s,d_{1})}=\hat{P}^{(s,d_{2})}} the covariance is zero.

In the general case \eqn{\hat{D}^{(s,d)}} and \eqn{\hat{P}^{(s,d)}} is estimated by a ratio estimator, and the error in these estimates are ignored. When \eqn{d} covers all of \eqn{s}, \eqn{\hat{D}^{(s,d)}=1} and \eqn{\hat{P}^{(s,d)}=1}is known. In this case both expressions can be shown to be unbiased. These quantities can only be calculated when \eqn{p_{i}} is provided, and will otherwise be NA.}

\item{Mean:}{
The estimate of the mean value of some variable \eqn{v} in domain \eqn{d} and stratum \eqn{s}, when MeanOfMeans is false:
\deqn{\hat{\mu}^{(s,d,v)} = \frac{\hat{t}^{(s,d,v)}}{\hat{N}^{(s,d)}}}
with co-variance:
\deqn{\widehat{CoVar}(\hat{\mu}^{(s,d_{1},v_{1})}, \hat{\mu}^{(s,d_{2},v_{2})}) = \frac{1}{\hat{N}^{(s,d_{1})} \hat{N}^{(s,d_{2})}}\widehat{CoVar}(\hat{t}^{(s,d_{1},v_{1})}, \hat{t}^{(s,d_{2},v_{2})})}
These are ratio estimates depending on the ratio to the estimated value \eqn{\hat{N}^{(s,d)}}, and the error in this estimate is ignored.
In addition, the estimate may depend on a ratio estimate for \eqn{\hat{t}^{(s,d,v)}} and \eqn{\hat{t}^{(s,d,v)}_{i}}, as explained for 'Abundance'
and in \code{\link[RstoxFDA]{AnalyticalPSUEstimate}}.
}

\item{Mean, Mean of Means:}{
The estimate of the mean value of some variable in domain \eqn{d} and stratum \eqn{s}, when MeanOfMeans is true:
\deqn{\hat{\mu}^{(s,d,v)}=\sum_{i=1}^{n}\frac{w_{i}}{\hat{d}^{(s,d)}}\hat{\mu}_{i}I^{(s,d,v)}_{i}H(\hat{N}^{(s,d)}_{i})}
with co-variance:
\deqn{\widehat{CoVar}(\hat{\mu}^{(s,d_{1},v_{1})}, \hat{\mu}^{(s,d_{2},v_{2})}) = \frac{1}{(\hat{d}^{(s,d_{1} \cap d_{2})})^{2}}\frac{1}{n^{(s)}(n^{(s)}-1)} \sum_{i=1}^{n} 
  I^{(s,d_{1})}_{i}I^{(s,d_{2})}_{i} H(\hat{f}^{(s,d_{1})}_{i})H(\hat{f}^{(s,d_{2})}_{i})( \hat{\mu}^{(s,d_{1},v_{1})}_{i} - \hat{\mu}^{(s,d_{1},v_{1})}) (\hat{\mu}^{(s,d_{2},v_{2})}_{i} - \hat{\mu}^{(s,d_{2},v_{2})})}}
These are ratio estimates depending on the ratio ratio estimation of \eqn{\hat{d}^{(s,d)}}, \eqn{\hat{d}^{(s,d_{1} \cap d_{2})}} and \eqn{\hat{\mu}^{(s,d,v)}_{i}}, and the error in these estimates are ignored.
}

Vocabulary for notation used above:
\describe{
\item{\eqn{H(x)}}{A step function which is 1 when \eqn{x>0}, otherwise it is zero.}
\item{\eqn{I^{(s)}_{i}}}{The indicator function for stratum \eqn{s}. Is 1 when \eqn{i} is in stratum \eqn{s}, otherwise it is zero.}
\item{\eqn{I^{(s,d)}_{i}}}{The indicator function for domain \eqn{d} and stratum \eqn{s}. Is 1 when \eqn{i} is in stratum \eqn{s} and domain \eqn{d}, otherwise it is zero.}
\item{\eqn{J^{(s,d)}_{i}}}{The indicator function for domain \eqn{d} and stratum \eqn{s}. Is 1 when \eqn{i} is in stratum \eqn{s} and the PSU-domain of domain \eqn{d}, otherwise it is zero.}
\item{\eqn{n}}{Sample size, the number of PSUs sampled.}
\item{\eqn{n^{(s)}}}{Stratum sample size, the number of PSUs sampled in stratum \eqn{s}: \eqn{n_{s}=\sum_{i=1}^{n}I^{(s)}_{i}}.}
\item{\eqn{n^{(s,d)}}}{Domain sample size, the number of PSUs sampled in domain{d} and stratum \eqn{s}: \eqn{n^{(s,d)}=\sum_{i=1}^{n}I^{(s,d)}_{i}}. 'Samples' in \code{\link[RstoxFDA]{AnalyticalPopulationEstimateData}}.}
\item{\eqn{p_{i}}}{The selection probability of PSU \eqn{i}. 'SelectionProbability' in \code{\link[RstoxFDA]{PSUSamplingParametersData}}.}
\item{\eqn{w_{i}}}{The normalized Hansen-Hurwitz sampling weight: \eqn{w_{i}=\frac{1}{p_{i}Q_{i}}}, \eqn{Q_{i}=\sum_{j=1}^{n}\frac{I^{(s(i))}_{j}}{p_{j}}}, where \eqn{s(i)} denote the strata of sample \eqn{i}. 'HHsamplingWeight' in \code{\link[RstoxFDA]{PSUSamplingParametersData}}.}
\item{\eqn{\hat{D}^{(s,d)}}}{The estimated relative domain size (fraction of PSUs) of domain \eqn{d} in stratum \eqn{s}: \eqn{\hat{D}^{(s,d)}=\sum_{i=1}^{n}w_{i}I^{(s,d)}_{i}}.}
\item{\eqn{\hat{P}^{(s,d)}}}{The estimated relative domain size (fraction of PSUs) of the PSU-domain of domain \eqn{d} in stratum \eqn{s}: \eqn{\hat{P}^{(s,d)}=\sum_{i=1}^{n}w_{i}J^{(s,d)}_{i}}.  'PSURelativeDomainSize' in \code{\link[RstoxFDA]{AnalyticalPopulationEstimateData}}.}
\item{\eqn{\hat{M}^{(s,d)}}}{The estimated domain size (number of PSUs) of the PSU-domain of domain \eqn{d} in stratum \eqn{s}: \eqn{\hat{M}^{(s,d)}=\sum_{i=1}^{n}\frac{1}{p_{i}}J^{(s,d)}_{i}}.  'PSURelativeDomainSize' in \code{\link[RstoxFDA]{AnalyticalPopulationEstimateData}}.}
\item{\eqn{\hat{d}^{(s,d)}}}{The estimated relative domain size (fraction of PSUs) that has observations (positive abundance or frequency) in domain \eqn{d} in stratum \eqn{s}: \eqn{\hat{d}^{(s,d)}=\sum_{i=1}^{n}w_{i}I^{(s,d)}_{i}H(\hat{f}^{(s,d)}_{i})}.}
\item{\eqn{\hat{d}^{(s,d_{1} \cap d_{2})}}}{The estimated relative domain size (total number of PSUs) that has observations (positive abundance or frequency) in both domain \eqn{d_{1}} and \eqn{d_{2}} in stratum \eqn{s}: \eqn{\hat{d}^{(s,d)}=\sum_{i=1}^{n}w_{i}I^{(s,d_{1})}_{i}I^{(s,d_{2})}_{i}H(\hat{f}^{(s,d_{1})}_{i})H(\hat{f}^{(s,d_{2})}_{j})}.}
\item{\eqn{\hat{N}^{(s)}}}{The estimated abundance in stratum \eqn{s}: \eqn{\hat{N}^{(s)}=\frac{1}{n_{s}}\sum_{i=1}^{n}\frac{\hat{N}_{i}}{p_{i}}I^{(s)}_{i}}.}
\item{\eqn{\hat{N}^{(s)}_{i}}}{The estimated total abundance in stratum \eqn{s} at PSU \eqn{i}. 'Abundance' in \code{\link[RstoxFDA]{AnalyticalPSUEstimateData}}.}
\item{\eqn{\hat{N}^{(s,d)}_{i}}}{The estimated abundance in domain \eqn{d} and stratum \eqn{s} at PSU \eqn{i}. 'Abundance' in \code{\link[RstoxFDA]{AnalyticalPSUEstimateData}}.}
\item{\eqn{\hat{f}^{(s,d)}_{i}}}{The estimated frequency in domain \eqn{d} for stratum \eqn{s} at PSU \eqn{i}. 'Frequency' in \code{\link[RstoxFDA]{AnalyticalPSUEstimateData}}.}
\item{\eqn{\hat{t}^{(s,d,v)}_{i}}}{The estimated total of some variable \eqn{v} in domain \eqn{d} and stratum \eqn{s} at PSU \eqn{i}. 'Total' in \code{\link[RstoxFDA]{AnalyticalPSUEstimateData}}.}
\item{\eqn{\hat{\mu}^{(s,d,v)}_{i}}}{The estimated mean of some variable \eqn{v} in domain \eqn{d} and stratum \eqn{s} at PSU \eqn{i}. 'Mean' in \code{\link[RstoxFDA]{AnalyticalPSUEstimateData}}.}
}
}
\examples{
 PSUsamplingParameters <- RstoxFDA::AssignPSUSamplingParameters(
                                       RstoxFDA::CatchLotterySamplingExample, 
                                       RstoxFDA::CatchLotteryExample, 
                                       "serialnumber", "Haul", "MissingAtRandom")
 individualSamplingParameters <-  RstoxFDA:::DefineIndividualSamplingParameters(NULL, 
                                       RstoxFDA::CatchLotteryExample, "SRS", c("IndividualAge"))
                                       
 psuEst <- RstoxFDA:::AnalyticalPSUEstimate(RstoxFDA::CatchLotteryExample, 
                                       individualSamplingParameters, 
                                       c("IndividualRoundWeight"), c("IndividualAge"))
 popEst <- RstoxFDA:::AnalyticalPopulationEstimate(PSUsamplingParameters, psuEst)
 
 #tabulate abundance
 abundance <- popEst$Abundance[,list(Abundance=Abundance), by=c("Stratum", "Domain")]
 #add SE and CV
 abundance <- merge(abundance, popEst$AbundanceCovariance[Domain1==Domain2,
                   list(SE=sqrt(AbundanceCovariance)), 
                   by=list(Stratum=Stratum, Domain=Domain1)])
 abundance$CV <- abundance$SE/abundance$Abundance
 
 #need to order as numeric. Domain is always a chr:
 abundance <- abundance[order(as.numeric(abundance$Domain)),]
 abundance
}
\concept{Analytical estimation}
