% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{AnalyticalPopulationEstimateData}
\alias{AnalyticalPopulationEstimateData}
\title{Analytical Population Estimate Data}
\description{
Analytical estimates for a population
 
 Abundance
 \describe{
   \item{Stratum}{}
   \item{Domain}{}
   \item{Abundance}{}
   \item{Frequency}{}
 }
 
 Variables
 \describe{
   \item{Stratum}{}
   \item{Domain}{}
   \item{Variable}{}
   \item{Total}{}
   \item{Mean}{}
 }
 
 AbundanceCovariance
 \describe{
   \item{Stratum}{}
   \item{Domain1}{}
   \item{Domain2}{}
   \item{AbundanceCovariance}{}
   \item{FrequencyCovariance}{}
 }
 
 VariableCovariance
 \describe{
   \item{Stratum}{}
   \item{Domain1}{}
   \item{Domain2}{}
   \item{Variable1}{}
   \item{Variable2}{}
   \item{TotalCovariance}{}
   \item{MeanCovariance}{}
 }
 
 StratificationVariables
 \describe{
  \item{Stratum}{}
  \item{StratificationVariables}{}
 }
 
 DomainVariables
 \describe{
  \item{Domain}{}
  \item{DomainVariables}{}
 }
}
\concept{Analytical estimation}
\concept{Data types}
