% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{AnalyticalPopulationEstimateData}
\alias{AnalyticalPopulationEstimateData}
\title{Analytical Population Estimate Data}
\description{
Analytical estimates for a population
 
 Abundance
 \describe{
   \item{Stratum}{Stratum that abundance is provided for.}
   \item{Domain}{The domain that the abundance is provided for.}
   \item{Abundance}{The estimated number of individuals in the domain and stratum.}
   \item{Frequency}{The estimated fraction of individuals in the domain, relative ot the total number in stratum.}
 }
 
 Variables
 \describe{
   \item{Stratum}{The stratum that estimates are provided for.}
   \item{Domain}{The domain that estimates are provided for.}
   \item{Variable}{The variable (measurment) that estimates are provided for.}
   \item{Total}{The estimated total value of the variable in domain and stratum.}
   \item{Mean}{The estimated mean value fo the variable in domain and stratum.}
 }
 
 AbundanceCovariance
 \describe{
   \item{Stratum}{The stratum that covariances are provided for.}
   \item{Domain1}{A domain that covariances are provided for. Unique, given Stratum}
   \item{Domain2}{A domain that covariances are provided for. Unique, given Stratum and Domain 1.}
   \item{AbundanceCovariance}{The estimated covariance of abundance between Domain1 and Domain2.}
   \item{FrequencyCovariance}{The estimated covariance of frequency between Domain1 and Domain2.}
 }
 
 VariableCovariance
 \describe{
   \item{Stratum}{The stratum that covariances are provied for.}
   \item{Domain1}{A domain that covariances are provided for. Unique, given Stratum}
   \item{Domain2}{A domain that covariances are provided for. Unique, given Stratum and Domain 1.}
   \item{Variable1}{A variable that covariances are provided for. Unique, given Stratum and domains.}
   \item{Variable2}{A variable that covariances are provided for. Unique, given Stratum, domains and Variable1.}
   \item{TotalCovariance}{The estimated covariance of total value of Variable1 in Domain1 and Variable2 in Domain2}
   \item{MeanCovariance}{The estimated covariance of the mean value of Variable1 in Domain1 and Variable2 in Domain2}
 }
 
 StratificationVariables
 \describe{
  \item{Stratum}{A stratum, as identified in other tables.}
  \item{StratificationVariables}{Columns that relate the stratum to data records.}
 }
 
 DomainVariables
 \describe{
  \item{Domain}{A domain, as identified in other tables.}
  \item{DomainVariables}{Columns that relate the domain to data records.}
 }
}
\concept{Analytical estimation}
\concept{Data types}
