% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalyticalBaselineFunctions.R
\name{AnalyticalRatioEstimate}
\alias{AnalyticalRatioEstimate}
\title{Ratio estimate of abundance}
\usage{
AnalyticalRatioEstimate(
  AnalyticalPopulationEstimateData,
  StoxLandingData,
  WeightVariable = character()
)
}
\arguments{
\item{AnalyticalPopulationEstimateData}{\code{\link[RstoxFDA]{AnalyticalPopulationEstimateData}} with estimates of mean or total weights and frequencies or abundance in domains}

\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} with census data on total weight in each stratum}

\item{WeightVariable}{character() name of variable in 'AnalyticalPopulationEstimateData' that represent weight of individuals.}
}
\value{
\code{\link[RstoxFDA]{AnalyticalPopulationEstimateData}} with ratio estimates of abundance
}
\description{
Performs ratio estimate of abundance, based on either the ratio of abundance in domains to total weight in a stratum, or the frequency and mean weights of each domain.
}
\concept{Analytical estimation}
