% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.R
\name{appendAreaCode}
\alias{appendAreaCode}
\title{Append area code}
\usage{
appendAreaCode(table, areaPolygons, latName, lonName, colName)
}
\arguments{
\item{table}{data.table to be annotated.}

\item{areaPolygons}{sp::SpatialPolygons with strata names in ID slot}

\item{latName}{name of WGS84 lat column}

\item{lonName}{name of WGS84 lon column}

\item{colName}{name of column to be appended}
}
\description{
Append area code
}
\details{
Appends a column with an area code to a table, based on positions.
}
