% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.R
\name{appendAreaCode}
\alias{appendAreaCode}
\title{Append area code}
\usage{
appendAreaCode(
  table,
  areaPolygons,
  latName,
  lonName,
  colName,
  StratumName = "StratumName"
)
}
\arguments{
\item{table}{data.table to be annotated.}

\item{areaPolygons}{\code{\link[sp]{SpatialPolygonsDataFrame}}}

\item{latName}{name of WGS84 lat column in 'table'}

\item{lonName}{name of WGS84 lon column in 'table}

\item{colName}{name of column to be appended to 'table'}

\item{StratumName}{name of column in 'areaPolygons' that identify the area name}
}
\value{
'table' with the area appended in the column 'colName'
}
\description{
Append area code
}
\details{
Appends a column with an area code to a table, based on positions.
}
