% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.R
\name{appendPosition}
\alias{appendPosition}
\title{Append positions}
\usage{
appendPosition(
  table,
  areaPolygons,
  areaName,
  latColName,
  lonColName,
  StratumName = "StratumName"
)
}
\arguments{
\item{table}{data.table to be annotated.}

\item{areaPolygons}{\code{\link[sp]{SpatialPolygonsDataFrame}}}

\item{areaName}{name of column that identifies the area in 'table'}

\item{latColName}{name of the latitdue column to be appended to 'table'}

\item{lonColName}{name of the longitude column to be appended to 'table'}

\item{StratumName}{name of the column in 'areaPolygons' that identifies the area.}
}
\value{
'table' with the positions appended in the columns 'latColName' and 'lonColName'.
}
\description{
Appends columns with positions to a data table, based on an area code.

 Coordinates are retrieved from a \code{\link[sp]{SpatialPolygonsDataFrame}} and not calculated
 the exact defintion of the coordinates depend on how the polygons were constructed.

 Datum and projection is not enforced, but a warning is issued if 'areaPolygons' does not pass some
 checks to verify that it is not a planar projection.
}
\seealso{
Other spatial coding functions: 
\code{\link{AddAreaPositionStoxLanding}()},
\code{\link{AddStratumStoxBiotic}()},
\code{\link{AddStratumStoxLanding}()},
\code{\link{DefineAreaPosition}()},
\code{\link{DefineCarNeighbours}()},
\code{\link{LoadFdaStratumPolygon}()},
\code{\link{SetAreaPositionsBiotic}()},
\code{\link{appendAreaCode}()},
\code{\link{mergePolygons}()},
\code{\link{plotArea}()}
}
\concept{spatial coding functions}
