% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.R
\name{areaCodeConversionTable}
\alias{areaCodeConversionTable}
\title{Area code conversion table}
\usage{
areaCodeConversionTable(
  areaDef1,
  areaDef2,
  areaName1 = "StratumName",
  areaName2 = areaName1,
  method = c("overlap", "centroids")
)
}
\arguments{
\item{areaDef1}{\code{\link[sp]{SpatialPolygonsDataFrame}} defining area codes}

\item{areaDef2}{\code{\link[sp]{SpatialPolygonsDataFrame}} defining area codes}

\item{areaName1}{column in areaDef1 identifying the name of areas}

\item{areaName2}{column in areaDef2 identifying the name of areas}

\item{method}{method for mapping area codes. See details.}
}
\value{
a list mapping area codes in areaDef1 to those in areaDef2
}
\description{
Make conversion table from area code-definitions.
 
 The conversion table can be constructed with the following methods (argument: 'method'):
 \describe{
 \item{overlap}{Each area in areaDef1 is mapped to the area in areaDef2 with the largest overlap.}
 \item{centroids}{Eac area in areaDef1 is mapped to the area in areaDef2 which contains its centroid.}
 }
 Centroids may be preferred for performance reasons, but depending on the shape of areas, 
 the centroid may not always be a good representation.
 
 Areas in areaDef1 that has no corresponding area in areaDef2 will be omitted from results.
}
\examples{
 # map fdir areas to ICES areas by overlap
 fdir.ICES.map <- areaCodeConversionTable(RstoxFDA::mainareaFdir2018, RstoxFDA::ICESareas)
 
 # map fdir locations to ICES statistical rectangles, by centroids
 loc.rectangles.map <- areaCodeConversionTable(RstoxFDA::locationsFdir2018, 
        RstoxFDA::ICESrectangles, method="centroids")
 
 # map selected statistical rectangles to ICES areas by overlap
 data(catchsamples)
 selectedRects <- RstoxFDA::ICESrectangles[
             RstoxFDA::ICESrectangles$StratumName \%in\% catchsamples$LEstatRect,]
 catchsamples$LEarea <- RstoxFDA::convertCodes(catchsamples$LEstatRect, 
             areaCodeConversionTable(selectedRects, 
             RstoxFDA::ICESareas))
}
\seealso{
\code{\link[RstoxFDA]{plotAreaComparison}} for visual inspection of how different area definitions correspond.

Other spatial coding functions: 
\code{\link{AddAreaPositionStoxLanding}()},
\code{\link{AddStratumStoxBiotic}()},
\code{\link{AddStratumStoxLanding}()},
\code{\link{DefineAreaPosition}()},
\code{\link{DefineCarNeighbours}()},
\code{\link{LoadFdaStratumPolygon}()},
\code{\link{SetAreaPositionsBiotic}()},
\code{\link{appendAreaCode}()},
\code{\link{appendPosition}()},
\code{\link{mergePolygons}()},
\code{\link{plotAreaComparison}()},
\code{\link{plotArea}()},
\code{\link{writeSpDataFrameAsWKT}()}
}
\concept{spatial coding functions}
