% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metierAnnotation.R
\name{assignMetier}
\alias{assignMetier}
\title{Annotate metier}
\usage{
assignMetier(
  data,
  metiertable,
  gearColumn,
  targetColumn = NULL,
  meshSizeColumn = NULL,
  selectivityDeviceColumn = NULL,
  selectivityDeviceMeshSizeColumn = NULL,
  areaColumn = NULL,
  metierColName = "metier"
)
}
\arguments{
\item{data}{\code{\link[data.table]{data.table}} with data to be annotated}

\item{metiertable}{\code{\link[RstoxFDA]{MetierTable}} with metier definitions. Accepts all NAs for columns not used.}

\item{gearColumn}{character() identifies the column in 'data' that encodes gear. Gear definition must match metiertable$gearcode}

\item{targetColumn}{character(), optional, identifies the column in 'data' that encodes target species. Definition must match metiertable$target}

\item{meshSizeColumn}{integer(), optional, identifies the column in 'data' that encodes the mesh size of the gear.}

\item{selectivityDeviceColumn}{character(), optional, identifies the column in 'data' that encodes selectivitydevices mounted on gear. Definition must mathc metiertable$selectivityDevice}

\item{selectivityDeviceMeshSizeColumn}{integer(), optional, identifies the column in 'data' that encodes the mesh size of any mounted selectivity device.}

\item{areaColumn}{character(), optional, identifies the column in 'data' that encodes areas for which different metierdefinitions are applicable.}

\item{metierColName}{character() name of the column that should be appended to 'data'}

\item{strict}{logical(), whether strict annotation should be applied (halting on missing definitions).}
}
\value{
\code{\link[data.table]{data.table}} 'data' with the column 'metierColName' appended (character).
}
\description{
annotates metier to data table
}
\details{
The parameter 'areaColumn' may be NULL which specifies that metiers are not dependent on area for the annotation.
 If areaColumn is given, NA values in 'metiertable' specifies that metiers apply to all areas, so that area codes need only be provided for exceptions to this rule.
}
