% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatchLotteryExample-datadoc.R
\docType{data}
\name{CatchLotteryExample}
\alias{CatchLotteryExample}
\title{Data from the Norwegian catch lottery sampling program.}
\format{
\code{\link[RstoxData]{StoxBioticData}}
}
\usage{
data(CatchLotteryExample)
}
\description{
Example of data formatted as \code{\link[RstoxData]{StoxBioticData}}, with an added column 'serialnumber' at the Haul-table and an added column 'CountryVessel' at the Station-table
which identifies the selection in the catch sampling lottery that this data was recorded for (see 'SamplingUnitId' in \code{\link[RstoxFDA]{CatchLotterySamplingExample}})
}
\details{
Hauls are primary sampling units, selected by Poission sampling with selection probabilities proportional to the catch size.
Corresponding sampling parameters are provided in \code{\link[RstoxFDA]{CatchLotterySamplingExample}}
The data contain North Sea herring samples from catch lottery sampling in 2022.
}
\examples{
 RstoxFDA::plotArea(RstoxFDA::CatchLotteryExample$Station, 
      areaDef=RstoxFDA::mainareaFdir2018, 
      latCol = "Latitude", 
      lonCol = "Longitude", 
      areaLabels = TRUE)
}
\concept{Analytical estimation}
\keyword{datasets}
