% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalyticalBaselineFunctions.R
\name{ComputePSUSamplingParameters}
\alias{ComputePSUSamplingParameters}
\title{Compute PSU Sampling Design Parameters}
\usage{
ComputePSUSamplingParameters(
  StoxBioticData,
  DefinitionMethod = c("AdHocStoxBiotic"),
  SamplingUnitId = character(),
  StratificationColumns = character()
)
}
\arguments{
\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} Sample data to construct design parameters from}

\item{DefinitionMethod}{'AdHocStoxBiotic'}

\item{SamplingUnitId}{name of column in 'StoxBioticData' that identifies the Primary Sampling Unit the design is constructed for.}

\item{StratificationColumns}{name of any column (at the same table as 'SamplingUnitId') that are to be used to define Strata for sampling.}
}
\value{
\code{\link[RstoxFDA]{PSUSamplingParametersData}}
}
\description{
Compute sampling parameters for Primary Sampling Units in multi-stage sampling.
}
\details{
Computes Sampling Design Parameters from data, given some assumptions specified by the 'DefinitionMethod'.

If 'DefinitionMethod' is 'AdHocStoxBiotic', equal probability sampling with fixed sample size will be assumed, as well as
selection with replacement and complete response.
This is a reasonable approximation if within-strata sampling is approximately simple random selections,
the sample intensitiy is low (only a small fraction of the population is sampled),
and non-response is believed to be random.
}
\examples{
 # parameters for simpler random haul-selection, stratified by GearGroup
 PSUparams <- ComputePSUSamplingParameters(RstoxFDA::StoxBioticDataExample, 
  "AdHocStoxBiotic", 
  "Haul", 
  "GearGroup")

}
\concept{Analytical estimation}
\concept{StoX-functions}
