% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.R
\name{convertCodes}
\alias{convertCodes}
\title{Convert codes.}
\usage{
convertCodes(code, conversionTable)
}
\arguments{
\item{code}{character() with original codes}

\item{conversionTable}{list() mapping code to converted code.}
}
\value{
character() with converted codes
}
\description{
Apply conversion table, perform approriate checks and return result.
}
\details{
Will stop with error if any codes can not be converted, or if any entries are NA.
 Require all codes (original and converted) to be character().
}
\examples{
 gearConversion <- list()
 gearConversion["TBS"] <- "OTB"
 gearConversion["TBN"] <- "OTB"
 gearConversion["OTB"] <- "OTB"
 convertCodes(c("TBS", "TBN", "OTB"), gearConversion)
}
