% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{DefineAgeErrorMatrix}
\alias{DefineAgeErrorMatrix}
\title{Define Age Error Matrix}
\usage{
DefineAgeErrorMatrix(
  processData,
  DefinitionMethod = c("ResourceFile"),
  FileName = character(),
  UseProcessData = F
)
}
\arguments{
\item{processData}{\code{\link[RstoxFDA]{AgeErrorMatrix}} as returned from this function}

\item{DefinitionMethod}{'ResourceFile'. See details.}

\item{FileName}{path to resource file}

\item{UseProcessData}{If TRUE, bypasses execution of function and returns existing 'processData'}
}
\value{
Age Error Matrix, see: \code{\link[RstoxFDA]{AgeErrorMatrix}}.
}
\description{
Defines probabilities for misreading ages.
}
\details{
Defines probabilities for misreading ages, that is assumed probabilites for determining
a fish to be of age 'x' given that it actually is age 'y'.

Definitions are read from a tab separated file with headers and row names in first column.
All row and column names should be integers representing ages.
The matrix encodes the probability of observing an age (rows), given true age (columns),
and columns must sum to 1.
}
\seealso{
\code{\link[RstoxFDA]{PrepareRecaEstimate}} for use of age-error matrices in Reca-estimation

Other StoX-Reca functions: 
\code{\link{AddStratumStoxLanding}()},
\code{\link{DefineStockSplittingParameters}()},
\code{\link{ParameterizeRecaModels}()},
\code{\link{PrepareRecaEstimate}()},
\code{\link{ReportRecaCatchAtAgeCovariance}()},
\code{\link{ReportRecaCatchAtAge}()},
\code{\link{ReportRecaCatchAtLengthAndAge}()},
\code{\link{ReportRecaCatchAtLength}()},
\code{\link{ReportRecaCatchStatistics}()},
\code{\link{ReportRecaLengthAtAge}()},
\code{\link{ReportRecaParameterStatistics}()},
\code{\link{ReportRecaWeightAtAge}()},
\code{\link{RunRecaModels}()}
}
\concept{StoX-Reca functions}
