% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{DefineAreaPosition}
\alias{DefineAreaPosition}
\title{Define Area Code Positions}
\usage{
DefineAreaPosition(
  processData,
  DefinitionMethod = c("ResourceFile", "StratumPolygon"),
  FileName = character(),
  StratumPolygon,
  UseProcessData = F
)
}
\arguments{
\item{processData}{\code{\link[RstoxFDA]{AreaPosition}} as returned from this function.}

\item{DefinitionMethod}{'ResourceFile' or 'StratumPolygon', see details.}

\item{FileName}{path to resource file}

\item{StratumPolygon}{\code{\link[RstoxBase]{StratumPolygon}} to extract area positions from.}

\item{UseProcessData}{If TRUE, bypasses execution of function and returns existing 'processData'}
}
\value{
\code{\link[RstoxFDA]{AreaPosition}}.
}
\description{
Define positions for areas of a spatial coding system.
}
\details{
Defines an association between area codes and positions that represent that area code.
For example an area code could be associated with the centre of mass of the area, or
some other point within the area. This may be useful for providing approximate
coordinates when locations are identified by area codes. The soruce of the
area-position association may be a table or a appropriately formatted
polygon-definition (\code{\link[RstoxBase]{StratumPolygon}}).

For DefinitionMethod 'ResourceFile':
Definitions are read from a tab separated, UTF-8 encoded file with headers. Columns defined as:
\describe{
\item{Column 1: 'Area'}{Area code (key)}
\item{Column 2: 'Location'}{optional subdivision of area. If provided, positions for missing locations should be encoded as well.}
\item{Column 3: 'Latitude'}{WGS84 Latitude, decimal degrees}
\item{Column 4: 'Longitude'}{WGS84 Longitude, decimal degrees}
}

For DefinitionMethod 'StratumPolygon':
Definitions are extracted from a \code{\link[RstoxBase]{StratumPolygon}}:
'Area' in \code{\link[RstoxFDA]{AreaPosition}} is derived from the column 'StratumName' in \code{\link[RstoxBase]{StratumPolygon}}.
'Location' in \code{\link[RstoxFDA]{AreaPosition}} is encoded as missing.
'Latitude' and 'Longitude' in \code{\link[RstoxFDA]{AreaPosition}} are the coordinates set for each polygon in \code{\link[RstoxBase]{StratumPolygon}}.
}
\seealso{
\code{\link[RstoxFDA]{SetAreaPositionsBiotic}} and \code{\link[RstoxFDA]{AddAreaPositionStoxLanding}} for adding positions to data.

Other spatial coding functions: 
\code{\link{AddAreaPositionStoxLanding}()},
\code{\link{AddStratumStoxBiotic}()},
\code{\link{AddStratumStoxLanding}()},
\code{\link{DefineCarNeighbours}()},
\code{\link{LoadFdaStratumPolygon}()},
\code{\link{SetAreaPositionsBiotic}()},
\code{\link{appendAreaCode}()},
\code{\link{appendPosition}()},
\code{\link{areaCodeConversionTable}()},
\code{\link{mergePolygons}()},
\code{\link{plotAreaComparison}()},
\code{\link{plotArea}()},
\code{\link{writeSpDataFrameAsWKT}()}

Other StoX-functions: 
\code{\link{AddAreaPositionStoxLanding}()},
\code{\link{AddGearGroupStoxBiotic}()},
\code{\link{AddGearGroupStoxLanding}()},
\code{\link{AddPeriodStoxBiotic}()},
\code{\link{AddPeriodStoxLanding}()},
\code{\link{AddStratumStoxBiotic}()},
\code{\link{AddStratumStoxLanding}()},
\code{\link{ConvertLengthBiotic}()},
\code{\link{ConvertWeightBiotic}()},
\code{\link{DefineAgeErrorMatrix}()},
\code{\link{DefineCarNeighbours}()},
\code{\link{DefineLengthConversionParameters}()},
\code{\link{DefinePeriod}()},
\code{\link{DefineStockSplittingParameters}()},
\code{\link{DefineWeightConversionFactor}()},
\code{\link{FilterAgeLengthOutliersStoxBiotic}()},
\code{\link{FilterWeightLengthOutliersStoxBiotic}()},
\code{\link{ListBioticDifference}()},
\code{\link{LoadFdaStratumPolygon}()},
\code{\link{ParameterizeRecaModels}()},
\code{\link{PrepareRecaEstimate}()},
\code{\link{ReadLandingFDA}()},
\code{\link{ReportFdaLandings}()},
\code{\link{ReportFdaSOP}()},
\code{\link{ReportFdaSampling}()},
\code{\link{ReportParameterConvergence}()},
\code{\link{ReportRecaCatchAtAgeCovariance}()},
\code{\link{ReportRecaCatchAtAge}()},
\code{\link{ReportRecaCatchAtLengthAndAge}()},
\code{\link{ReportRecaCatchAtLength}()},
\code{\link{ReportRecaCatchStatistics}()},
\code{\link{ReportRecaLengthAtAge}()},
\code{\link{ReportRecaParameterStatistics}()},
\code{\link{ReportRecaWeightAtAge}()},
\code{\link{RunRecaModels}()},
\code{\link{SetAreaPositionsBiotic}()},
\code{\link{SetShortGearBiotic}()},
\code{\link{SetStartDateBiotic}()},
\code{\link{SetTimeBiotic}()}
}
\concept{StoX-functions}
\concept{spatial coding functions}
