% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{DefineCarNeighbours}
\alias{DefineCarNeighbours}
\title{Define CAR neighbours}
\usage{
DefineCarNeighbours(processData, resourceFilePath, encoding = "UTF-8",
  useProcessData = F)
}
\arguments{
\item{processData}{data.table() as returned from this function}

\item{resourceFilePath}{path to resource file}

\item{encoding}{encoding of resource file}

\item{useProcessData}{logical() Bypasses execution of function, and simply returns argument 'processData'}
}
\value{
Area Neighbour Definition, see: \code{\link[RstoxFDA]{CarNeighbours}}.
}
\description{
StoX function.
 Define which spatial strata are to be considered neighbours,
 when used as a CAR-variable (Conditional AutoRegressive variable).
}
\details{
Definitions are read from a tab separated file with headers. Columns defined as:
 \describe{
 \item{Column 1: 'CarValue'}{Value for the CAR-variable (key)}
 \item{Column 2: 'Neigbhours'}{Comma-separated list of neighbours (each should occur in Column 1)}
 }
 The neighbour definition must be symmetric.
 If a is among the neighbours of b, b must also be among the neighbours of a.
}
