% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalyticalBaselineFunctions.R
\name{DefineIndividualSamplingParameters}
\alias{DefineIndividualSamplingParameters}
\title{Define Sub-Sampling Parameters for Individuals}
\usage{
DefineIndividualSamplingParameters(
  processData,
  StoxBioticData,
  DefinitionMethod = c("SRS", "Stratified", "LengthStratified"),
  Parameters = c(),
  LengthInterval = numeric(),
  StratificationColumns = character(),
  UseProcessData = FALSE
)
}
\arguments{
\item{processData}{\code{\link[RstoxFDA]{IndividualSamplingParametersData}} as returned from this function.}

\item{StoxBioticData}{Data to define individual sampling parameters for}

\item{DefinitionMethod}{Method to infer sampling parameters, 'SRS', 'Stratified' or 'LengthStratified'. See details.}

\item{Parameters}{Measurements / observations of individuals included in the design specification. Must all be column on the Individual-table of StoxBioticData.}

\item{LengthInterval}{width of length strata in cm. Specifies left closed intervals used for Length Stratified selection (DefinitionMethod 'Stratified'). A value of 5 indicates that observation are selected stratified on length groups [0 cm,5 cm>, [5 cm, 10 cm>, and so on.}

\item{StratificationColumns}{names of columns in the Individual table of StoxBioticData that identify strata for Stratified selection (DefinitionMethod 'Stratified').}

\item{UseProcessData}{If TRUE, bypasses execution of function and returns existing 'processData'}
}
\value{
\code{\link[RstoxFDA]{IndividualSamplingParametersData}}
}
\description{
Define approximate sampling design parameters for a sub-sample of individuals. Design parameters are inferred from data provided in ~\code{\link[RstoxData]{StoxBioticData}},
and specify how a set of individuals recorded on the Individual table were selected for observation/measurement from a sample recorded on the Sample table.
}
\details{
Sampling parameters are approximate inferred, assuming that all selected individuals are recorded, and based on some user-controllable assumptions about the selection process,
specified by the appropriate 'DefinitionMethod'. Individuals with a non-missing value for any of the parameters in 'Parameters' are treated as selected for observation.

The available DefinitionMethods are:
\describe{
\item{SRS}{Simple Random Selection. Individuals are selected for measurment by simple random selection without replacement}
\item{Stratified}{Stratified Selection. Individuals are selected for measurement by stratified random selection without replacement. Strata are specified as the combination of columns provided in 'StratificationColumns'. The number of fish in each stratum is estimated by the total in sample and the proportion of measured fish in each stratum.}
\item{LengthStratified}{Length stratified selection. Individuals are selected for measurement by stratified random selection without replacement. Strata are length groups, specified by the left closed intervals starting with [0,'LengthInterval'>.}
}
}
\concept{Analytical estimation}
\concept{StoX-functions}
