% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{DefinePeriod}
\alias{DefinePeriod}
\title{Define Periods}
\usage{
DefinePeriod(
  processData,
  TemporalCategory = c("Quarter", "Month", "Custom"),
  CustomPeriods = character(),
  UseProcessData = F
)
}
\arguments{
\item{processData}{\code{\link[RstoxFDA]{TemporalDefinition}} as returned from this function.}

\item{TemporalCategory}{type of temporal category: 'Quarter', 'Month' or 'Custom', defaults to 'Quarter'}

\item{CustomPeriods}{provide if temporalCategory is 'Custom', vector of strings formatted as DD-MM or DD-MM-YYYY, giving the start date of each temporal category. See 'details'}

\item{UseProcessData}{If TRUE, bypasses execution of function and returns existing 'processData'}
}
\value{
Temporal Categories, see: \code{\link[RstoxFDA]{TemporalDefinition}}.
}
\description{
Define periods, temporal categories such as 'Quarter', 'Month' or some custom defintion.
}
\details{
Periods define an association between dates and a categorical variable definition,
such as 'Month', 'Quarter' or some custom definition.

Periods may be seasonal, which means that dates are associated with the categories based on
day and month only. This distinction does not matter if all data is from the same year.
The 'TemporalCategory'-options 'Quarter' and 'Month' produce seasonal definitions, so
that Jan 1st 2015 and Jan 1st 2016 belon to the same category (e.g. "Q1" if 'TemporalCategory'
is 'Quarter').

In order to make custom seasonal definitions, use the TemporalCategory'-option 'Custom'
without providing years in 'CustomPeriods': "DD-MM". Periods defined in this way will automatically
"wrap around" so that dates in January may grouped together with dates in December.

In order to make non-seasonal definitions for Quarter or Month, provide the 'CustomPeriods'
for all years of interest on the format "DD-MM-YYYY". If years are provided the periods do not "wrap around"
and categories are automatically extended to the entire year, if necesesarry.
That is a category starting with 1. Jan is added if not present, and likewise a category ending with 31. Dec.
}
\examples{
 # Define seasonal periods quarter
 DefinePeriod(TemporalCategory = "Quarter")
 
 # Define two periods "15th March - 15th Sep" and "15th Sep - 15th March"
 DefinePeriod(TemporalCategory = "Custom", CustomPeriods = c("15-09", "15-03"))

 # Add years to speficiation to define four periods "15th March 2015 - 31st Dec 2015" ...
 DefinePeriod(TemporalCategory = "Custom", CustomPeriods = c("15-09-2105", "15-03-2015"))

}
\seealso{
\code{\link[RstoxFDA]{AddPeriodStoxBiotic}} and \code{\link[RstoxFDA]{AddPeriodStoxLanding}} for adding Periods to data.

Other temporal coding functions: 
\code{\link{AddPeriodStoxBiotic}()},
\code{\link{AddPeriodStoxLanding}()},
\code{\link{SetStartDateBiotic}()},
\code{\link{SetTimeBiotic}()},
\code{\link{categoriseDate}()}

Other StoX-functions: 
\code{\link{AddAreaPositionStoxLanding}()},
\code{\link{AddGearGroupStoxBiotic}()},
\code{\link{AddGearGroupStoxLanding}()},
\code{\link{AddPeriodStoxBiotic}()},
\code{\link{AddPeriodStoxLanding}()},
\code{\link{AddStratumStoxBiotic}()},
\code{\link{AddStratumStoxLanding}()},
\code{\link{ConvertLengthBiotic}()},
\code{\link{ConvertWeightBiotic}()},
\code{\link{DefineAgeErrorMatrix}()},
\code{\link{DefineAreaPosition}()},
\code{\link{DefineCarNeighbours}()},
\code{\link{DefineLengthConversionParameters}()},
\code{\link{DefineStockSplittingParameters}()},
\code{\link{DefineWeightConversionFactor}()},
\code{\link{FilterAgeLengthOutliersStoxBiotic}()},
\code{\link{FilterWeightLengthOutliersStoxBiotic}()},
\code{\link{ListBioticDifference}()},
\code{\link{LoadFdaStratumPolygon}()},
\code{\link{ParameterizeRecaModels}()},
\code{\link{PrepareRecaEstimate}()},
\code{\link{ReadLandingFDA}()},
\code{\link{ReportFdaLandings}()},
\code{\link{ReportFdaSOP}()},
\code{\link{ReportFdaSampling}()},
\code{\link{ReportParameterConvergence}()},
\code{\link{ReportRecaCatchAtAgeCovariance}()},
\code{\link{ReportRecaCatchAtAge}()},
\code{\link{ReportRecaCatchAtLengthAndAge}()},
\code{\link{ReportRecaCatchAtLength}()},
\code{\link{ReportRecaCatchStatistics}()},
\code{\link{ReportRecaLengthAtAge}()},
\code{\link{ReportRecaParameterStatistics}()},
\code{\link{ReportRecaWeightAtAge}()},
\code{\link{RunRecaModels}()},
\code{\link{SetAreaPositionsBiotic}()},
\code{\link{SetShortGearBiotic}()},
\code{\link{SetStartDateBiotic}()},
\code{\link{SetTimeBiotic}()}
}
\concept{StoX-functions}
\concept{temporal coding functions}
