% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalyticalBaselineFunctions.R
\name{DefineSamplingHierarchy}
\alias{DefineSamplingHierarchy}
\title{Extend IndividualSamplingParametersData to reflect selection from PSU by specifying intermediate selection.
Not yet implemented.}
\usage{
DefineSamplingHierarchy(
  StoxBioticData,
  IndividualSamplingParametersData,
  Hierarchy = character(),
  Stratification = character(),
  StrataSizes = character(),
  SelectionMetod = character()
)
}
\arguments{
\item{StoxBioticData}{data records}

\item{IndividualSamplingParametersData}{sampling parameters for lower level selection}

\item{Hierarchy}{character vector describing the hierarchy}

\item{Stratification}{character vector describing the stratification at each level in the hierarchy}

\item{StrataSizes}{identifies column where strata sizes are provided}

\item{SelectionMethod}{sampling method for each level in the hierarchy}
}
\value{
\code{\link[RstoxFDA]{IndividualSamplingParametersData}}
}
\description{
Extend IndividualSamplingParametersData to reflect selection from PSU by specifying intermediate selection.
Not yet implemented.
}
\concept{Analytical estimation}
