% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{DefineStockSplittingParamteres}
\alias{DefineStockSplittingParamteres}
\title{Define Classification Error (for stock splitting)}
\usage{
DefineStockSplittingParamteres(
  processData,
  DefinitionMethod = c("ResourceFile", "FunctionParameters"),
  FileName = character(),
  UseProcessData = F
)
}
\arguments{
\item{processData}{data.table() as returned from this function}

\item{FileName}{path to resource file}

\item{UseProcessData}{logical() Bypasses execution of function, and simply returns argument 'processData'}
}
\value{
\code{\link[RstoxFDA]{StockSplittingParamteres}}.
}
\description{
Defines parameters for the stock-splitting analysis in Reca, including
 parameters for misclassifying when determining stock membership of a specimen.
}
\details{
For DefinitionMethod 'ResourceFile', definitions are read from a UTF-8 encoded tab separated file with headers and one row
 with headers corresponding to column names in \code{\link[RstoxFDA]{StockSplittingParamteres}}.
 see \code{\link[RstoxFDA]{StockSplittingParamteres}} for further explanation on the coding system.
}
