% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDIRfactors-datadoc.R
\docType{data}
\name{FDIR.factors.VIII.2022}
\alias{FDIR.factors.VIII.2022}
\title{Weigh conversion Factors (FDIR VIII, 2022)}
\format{
\code{\link[RstoxFDA]{WeightConversionTable}}.
}
\usage{
data(FDIR.factors.VIII.2022)
}
\description{
Conversion factors for converting gutted weight to round (live) weight.
This dataset contains a selection of conversion factors from the official conversion factors
published by the Norwegian Directorate of fisheries (Document "Norske Omregningsfaktorer" version VIII, valid from 2022-01-01). 
Codes for converting from gutted fish without head, round cut (GWOH),
and from gutted fish with head (GWH) are provided for selected species.
}
\details{
ProductTypes and species codes are encoded as in NMDBiotic v3. That is code 3 for GWH and 4 for GWOH,
IMR-species codes (usually identical to ITIS tsn number)
}
\seealso{
Other parameter conversion functions: 
\code{\link{ConvertLengthBiotic}()},
\code{\link{ConvertWeightBiotic}()},
\code{\link{DefineLengthConversionParameters}()},
\code{\link{DefineWeightConversionFactor}()},
\code{\link{convertCodes}()}
}
\concept{parameter conversion functions}
\keyword{datasets}
