% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICESdivision-datadoc.R
\docType{data}
\name{ICESdivision}
\alias{ICESdivision}
\title{ICES areas}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with area names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}}.
}
\usage{
data(ICESdivision)
}
\description{
Definition ICES Divisions as they have been defined from 2018 inclusive.
}
\details{
Polygons are derived from shapefiles provided by ICES web-portals, and has been edited with some simplifications.
Notably detailed coast-lines have been removed, in favor of drawing area borders on land-mass.

Polygons are defined in WGS84 coordinates (unprojected).
}
\examples{
 # plot ICES areas
 data(ICESdivision)
 plotArea(areaDef=ICESdivision)
}
\keyword{datasets}
