% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICESunit-datadoc.R
\docType{data}
\name{ICESunit}
\alias{ICESunit}
\title{ICES units (area codes)}
\format{
\code{\link[sf]{sf}} with area names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}} (v.2).
}
\usage{
data(ICESunit)
}
\description{
Definition ICES Units (area codes).
}
\details{
Polygons are derived from shapefiles provided by ICES web-portals, and has been edited with some simplifications.
Notably detailed coast-lines have been removed, in favor of drawing area borders on land-mass.

Polygons are defined in WGS84 coordinates (unprojected).
}
\examples{
 # plot ICES areas
 data(ICESunit)
 plotArea(areaDef=ICESunit)
}
\concept{area code polygons}
\keyword{datasets}
