% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.AgeErrorMatrix}
\alias{is.AgeErrorMatrix}
\title{Check if argument is AgeErrorMatrix}
\usage{
is.AgeErrorMatrix(AgeErrorMatrix)
}
\arguments{
\item{AgeErrorMatrix}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conformed to specification for \code{\link[RstoxFDA]{AgeErrorMatrix}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{AgeErrorMatrix}}
}
