% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{is.RecaParameterData}
\alias{is.RecaParameterData}
\title{Check if argument is RecaParameterData}
\usage{
is.RecaParameterData(RecaParameterData)
}
\arguments{
\item{RecaParameterData}{argument to be checked for data conformity}
}
\value{
logical, TRUE if argument conforms to specification for \code{\link[RstoxFDA]{RecaParameterData}}
}
\description{
Checks if argument conforms to specification for \code{\link[RstoxFDA]{RecaParameterData}}
}
