% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{KommunePolygons}
\alias{KommunePolygons}
\title{Kommune polygons}
\description{
\code{\link[sp]{SpatialPolygonsDataFrame}} with area names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}}.
 
 Polygons are defined in WGS84 coordinates (unprojected).

 The data contains the following columns with rows for each area:
  \describe{
   \item{id}{Offical code for the muncipality (kommunenummer)}
   \item{name}{Official name for the muncipality}
   \item{StratumName}{Equal to id}
  }
}
\concept{Data types}
