% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings-datadoc.R
\docType{data}
\name{landings}
\alias{landings}
\title{Landings of Haddock}
\format{RDB CL-table version 1.3 (\href{https://www.ices.dk/marine-data/Documents/RDB/RDB Exchange Format.pdf}{RDBES exchange format v 1.13}). Some data types may differ from specification. Notably species, which is formatted as character().}
\usage{
data(landings)
}
\description{
Aggregated landings of Haddock by Norwegian vessels in 2018 as reported in salesnotes.
Gear-codes are converted from national standard and should be considered approximate.
}
\examples{
data(landings)
sum(landings$OfficialLandingsWeight)
}
\keyword{datasets}
