% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{LengthConversionTable}
\alias{LengthConversionTable}
\title{Length Conversion Table (LengthConversionTable)}
\description{
Length conversion parameters realting different length measurements.
}
\details{
Length conversion factors relating different length measurements, such as 'standard length' and 'fork length'
 based on a linear regression fit between length measurements:
 L1 = alpha + beta \* L2,
 where L1 and L2 are different length measurements
 and 'alpha' and 'beta' are species-specific coefficients.
 
 \code{\link[data.table]{data.table}} with columns:
 \describe{
 \item{'Description'}{character: Free-text description of the product}
 \item{'Species'}{character: Identifier for the species that the conversion applies to}
 \item{'MeasurmentType'}{character: Identifier for the type of length measurement for the independent variable (L2 above)}
 \item{'Alpha'}{numeric: scalar value representing the intercept (in cm) of a linear regression fit between length measurements.}
 \item{'Beta'}{numeric: scalar value representing the slope of a linear regression fit between length measurements.}
 }
}
\seealso{
Other Data types: 
\code{\link{AgeErrorMatrix}},
\code{\link{AreaPosition}},
\code{\link{CarNeighbours}},
\code{\link{KommunePolygons}},
\code{\link{LandingData}},
\code{\link{LandingsArchiveData}},
\code{\link{LocodeTable}},
\code{\link{LstLogbookData}},
\code{\link{MetierTable}},
\code{\link{ParameterConvergenceData}},
\code{\link{ParameterizationSummaryData}},
\code{\link{RecaCatchAtAge}},
\code{\link{RecaData}},
\code{\link{RecaParameterData}},
\code{\link{RecaResult}},
\code{\link{ReportFdaByAgeData}},
\code{\link{ReportFdaCatchAtAgeCovarianceData}},
\code{\link{ReportFdaCatchAtAgeData}},
\code{\link{ReportFdaCatchAtLengthAndAgeData}},
\code{\link{ReportFdaCatchAtLengthData}},
\code{\link{ReportFdaLandingData}},
\code{\link{ReportFdaLengthAtAgeData}},
\code{\link{ReportFdaSamplingData}},
\code{\link{ReportFdaSopData}},
\code{\link{ReportFdaSummaryData}},
\code{\link{ReportFdaWeightAtAgeData}},
\code{\link{StockSplittingParameters}},
\code{\link{TemporalDefinition}},
\code{\link{TripPartition}},
\code{\link{WeightConversionTable}},
\code{\link{is.AgeErrorMatrix}()},
\code{\link{is.AreaPosition}()},
\code{\link{is.CarNeighbours}()},
\code{\link{is.LengthConversionTable}()},
\code{\link{is.ParameterConvergenceData}()},
\code{\link{is.ParameterizationSummaryData}()},
\code{\link{is.RecaCatchAtAge}()},
\code{\link{is.RecaData}()},
\code{\link{is.RecaParameterData}()},
\code{\link{is.RecaResult}()},
\code{\link{is.ReportFdaByAgeData}()},
\code{\link{is.ReportFdaSOP}()},
\code{\link{is.ReportFdaSamplingData}()},
\code{\link{is.StockSplittingParameters}()},
\code{\link{is.TemporalDefinition}()},
\code{\link{is.Translation}()},
\code{\link{is.WeightConversionTable}()}
}
\concept{Data types}
