% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{ListBioticDifference}
\alias{ListBioticDifference}
\title{List Difference}
\usage{
ListBioticDifference(StoxBioticData, BioticData)
}
\arguments{
\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} data set.}

\item{BioticData}{\code{\link[RstoxData]{BioticData}} that has been read from a NMDbiotic v 3.x file}
}
\value{
\code{\link[RstoxData]{BioticData}} with data that is not found in 'StoxBioticData'.
}
\description{
List the difference between a NMDbiotic v 3.x data set and a StoxBiotic data set. Appropriate when BioticData is read from NMDbiotic.
}
\details{
Lists all missions, fishstations, catchsamples and individuals from a NMDbiotic v 3.x data set
that is not present in a given StoxBiotic data set.

This may be used to summarize the effect of filters and data conversions.
}
\concept{StoX-functions}
\concept{data QA functions}
\concept{nmdbiotic functions}
