% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationsFdir2017-datadoc.R
\docType{data}
\name{locationsFdir2017}
\alias{locationsFdir2017}
\title{Location codes (FDIR to 2017 incl.)}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with location names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}}.
}
\usage{
data(locationsFdir2017)
}
\description{
Definition for location coding system defined by the Norwegian directorate of Fisheries up until 2017 (inclusive).
Polygons are defined in WGS84 coordinates (unprojected).
}
\examples{
data(locationsFdir2017)
data(locationsFdir2018)
sp::plot(locationsFdir2018, ylim=c(57,60), xlim=c(0,25), border="blue")
sp::plot(locationsFdir2017, add=TRUE)
}
\seealso{
Other area code polygons: 
\code{\link{ICESareas}},
\code{\link{ICESdivision}},
\code{\link{ICESrectangles}},
\code{\link{ICESsubArea}},
\code{\link{ICESsubDivision}},
\code{\link{ICESunit}},
\code{\link{NAFOareas}},
\code{\link{locationsFdir2018}},
\code{\link{mainareaFdir2017}},
\code{\link{mainareaFdir2018}}
}
\concept{area code polygons}
\keyword{datasets}
