% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationsFdir2017-datadoc.R
\docType{data}
\name{locationsFdir2017}
\alias{locationsFdir2017}
\title{Location codes}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with location names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}}.
}
\usage{
data(locationsFdir2017)
}
\description{
Definition for location coding system defined by the Norwegian directorate of Fisheries up until 2017 (inclusive).
Polygons are defined in WGS84 coordinates (unprojected).
}
\examples{
data(locationsFdir2017)
data(locationsFdir2018)
sp::plot(locationsFdir2018, ylim=c(57,60), xlim=c(0,25), border="blue")
sp::plot(locationsFdir2017, add=TRUE)
}
\keyword{datasets}
