% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationsFdir2018-datadoc.R
\docType{data}
\name{locationsFdir2018}
\alias{locationsFdir2018}
\title{Location codes}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with location names identified in the column 'polygonName'. See \code{\link[RstoxBase]{StratumPolygon}}.
}
\usage{
data(locationsFdir2018)
}
\description{
Definition for location coding system defined by the Norwegian directorate of Fisheries  (2018 revision).
This revision has been in use for their fishery statistics as of 2018 (inclusive).
Polygons are defined in WGS84 coordinates (unprojected).
}
\examples{
data(locationsFdir2017)
data(locationsFdir2018)
sp::plot(locationsFdir2018, ylim=c(57,60), xlim=c(0,25), border="blue")
sp::plot(locationsFdir2017, add=TRUE)
}
\keyword{datasets}
