% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationsFdir2018-datadoc.R
\docType{data}
\name{locationsFdir2018}
\alias{locationsFdir2018}
\title{Location codes (FDIR from 2018 incl.)}
\format{
\code{\link[sf]{sf}} with area names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}} (v.2).
}
\usage{
data(locationsFdir2018)
}
\description{
Definition for location coding system defined by the Norwegian directorate of Fisheries  (2018 revision).
This revision has been in use for their fishery statistics as of 2018 (inclusive).
Polygons are defined in WGS84 coordinates (unprojected).
}
\examples{
 RstoxFDA::plotAreaComparison(RstoxFDA::mainareaFdir2017, 
     RstoxFDA::mainareaFdir2018, 
     xlim=c(0,20), ylim=c(50,60))
}
\concept{area code polygons}
\keyword{datasets}
