% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainareaFdir2017-datadoc.R
\docType{data}
\name{mainareaFdir2017}
\alias{mainareaFdir2017}
\title{Main areas (FDIR to 2017 incl.)}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with area names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}}.
}
\usage{
data(mainareaFdir2017)
}
\description{
Definition for area coding system defined by the Norwegian directorate of Fisheries up until 2017 (inclusive).
Polygons are defined in WGS84 coordinates (unprojected).
}
\examples{
 RstoxFDA::plotAreaComparison(RstoxFDA::mainareaFdir2017, 
     RstoxFDA::mainareaFdir2018, 
     xlim=c(0,20), ylim=c(50,60))
}
\concept{area code polygons}
\keyword{datasets}
