% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainareaFdir2017-datadoc.R
\docType{data}
\name{mainareaFdir2017}
\alias{mainareaFdir2017}
\title{Main areas}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with area names identified in the column 'polygonName'. See \code{\link[RstoxBase]{StratumPolygon}}.
}
\usage{
data(mainareaFdir2017)
}
\description{
Definition for area coding system defined by the Norwegian directorate of Fisheries up until 2017 (inclusive).
Polygons are defined in WGS84 coordinates (unprojected).
}
\examples{
data(mainareaFdir2017)
data(mainareaFdir2018)
sp::plot(mainareaFdir2017)
sp::plot(mainareaFdir2018, border=blue, add=T)
}
\keyword{datasets}
