% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainareaFdir2018-datadoc.R
\docType{data}
\name{mainareaFdir2018}
\alias{mainareaFdir2018}
\title{Main areas}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with area names identified in the column 'polygonName'. See \code{\link[RstoxBase]{StratumPolygon}}.
}
\usage{
data(mainareaFdir2018)
}
\description{
Definition for area coding system defined by the Norwegian directorate of Fisheries (2018 revision).
This revision has been in use for their fishery statistics as of 2018 (inclusive).
Polygons are defined in WGS84 coordinates (unprojected).
}
\examples{
data(mainareaFdir2018)
sp::plot(mainareaFdir2018)
}
\keyword{datasets}
