% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ParameterizationSummaryData}
\alias{ParameterizationSummaryData}
\title{Summary statistics for simulated parameters (ParameterizationSummaryData)}
\description{
Summary statistics for (potentially multi-chained) simulated parameters, 
 such as MCMC simulations with Reca.
 'chains' in this respect refers to statistically independent simulations.
 
 list with two members 'ParameterSummary' and 'RunParameters',
 which both all \code{\link[data.table]{data.table}}s. 'ParameterSummary'
 contains parameter statistics for simulations and have the following columns:
 \describe{
  \item{Parameter}{Identifies the parameter that is summarized.}
  \item{Mean}{Mean of the parameter}
  \item{Variance}{Variance of the parameter}
  \item{chainId}{Identifies the parameterization chain.}
 }
 
 RunParameters summarizes global control parameters for each chain and has columns:
 \describe{
  \item{chainId}{Identifies the parameterization chain.}
  \item{Iterations}{The number of iterations for parameter simulation}
 }
}
\seealso{
\code{\link[RstoxFDA]{RunRecaModels}} for running Reca-analysis

\code{\link[RstoxFDA]{ReportRecaParameterStatistics}} 
 for creating ParameterizationSummaryData from Reca-simulations.

Other Data types: 
\code{\link{AgeErrorMatrix}},
\code{\link{AreaPosition}},
\code{\link{CarNeighbours}},
\code{\link{KommunePolygons}},
\code{\link{LandingData}},
\code{\link{LandingsArchiveData}},
\code{\link{LengthConversionTable}},
\code{\link{LocodeTable}},
\code{\link{LstLogbookData}},
\code{\link{MetierTable}},
\code{\link{ParameterConvergenceData}},
\code{\link{RecaCatchAtAge}},
\code{\link{RecaData}},
\code{\link{RecaParameterData}},
\code{\link{RecaResult}},
\code{\link{ReportFdaByAgeData}},
\code{\link{ReportFdaCatchAtAgeCovarianceData}},
\code{\link{ReportFdaCatchAtAgeData}},
\code{\link{ReportFdaCatchAtLengthAndAgeData}},
\code{\link{ReportFdaCatchAtLengthData}},
\code{\link{ReportFdaLandingData}},
\code{\link{ReportFdaLengthAtAgeData}},
\code{\link{ReportFdaSamplingData}},
\code{\link{ReportFdaSopData}},
\code{\link{ReportFdaSummaryData}},
\code{\link{ReportFdaWeightAtAgeData}},
\code{\link{StockSplittingParameters}},
\code{\link{TemporalDefinition}},
\code{\link{TripPartition}},
\code{\link{WeightConversionTable}},
\code{\link{is.AgeErrorMatrix}()},
\code{\link{is.AreaPosition}()},
\code{\link{is.CarNeighbours}()},
\code{\link{is.LengthConversionTable}()},
\code{\link{is.ParameterConvergenceData}()},
\code{\link{is.ParameterizationSummaryData}()},
\code{\link{is.RecaCatchAtAge}()},
\code{\link{is.RecaData}()},
\code{\link{is.RecaParameterData}()},
\code{\link{is.RecaResult}()},
\code{\link{is.ReportFdaByAgeData}()},
\code{\link{is.ReportFdaSOP}()},
\code{\link{is.ReportFdaSamplingData}()},
\code{\link{is.StockSplittingParameters}()},
\code{\link{is.TemporalDefinition}()},
\code{\link{is.Translation}()},
\code{\link{is.WeightConversionTable}()}
}
\concept{Data types}
