% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plotArea}
\alias{plotArea}
\title{Plot area definitions on a map}
\usage{
plotArea(
  data = NULL,
  latCol = NULL,
  lonCol = NULL,
  groupCol = NULL,
  areaDef,
  areaNameCol = "StratumName",
  areaLabels = is.null(data),
  xlim = NULL,
  ylim = NULL,
  areaLabelSize = 2,
  pointColor = "darkred",
  pointShape = 23,
  pointSize = 1,
  title = "",
  projection = NULL,
  polygonColor = "blue"
)
}
\arguments{
\item{data}{data.frame with any points to be plotted}

\item{latCol}{character() identifing column in 'data' that specify latitudes (WGS84)}

\item{lonCol}{character() identifing column in 'data' that specify longitudes (WGS84)}

\item{groupCol}{character() identifying column in 'daat' that specify grouping of points}

\item{areaDef}{\code{\link[sp]{SpatialPolygonsDataFrame}}}

\item{areaNameCol}{identifies column in 'areaDef' with label names for the areas}

\item{areaLabels}{logical whether to plot area labels}

\item{xlim}{x axis limits in degrees longitude}

\item{ylim}{y axis limits in degrees latitude}

\item{areaLabelSize}{size for any area labels}

\item{pointColor}{color of any points to be plotted.}

\item{pointShape}{ggplot2 shape for any points to be plotted}

\item{pointSize}{size for any points to be plotted}

\item{title}{plot title}

\item{projection}{proj4string or EPSG code specifying the desired projection, see \code{\link[sf]{st_crs}}. Defaults to mercator projection.}

\item{polygonColor}{color to be used for plotting polygons.}
}
\description{
Plots area polygons on a map, may point a set of positions as well.
}
\details{
Map ranges to plot are specified by 'xlim' and 'ylim'. These should always be given in lon/lat.
 if these are not given they will be derived in order to include all positions in 'data'.
 if 'data is not given they will be derived from the bounding box of 'areaDef'.
 To allow for flexible projection, some space will be added around 'xlim' and 'ylim'.
 
 Colors can be specified as understood by ggplot2. E.g. one of those listed by \code{\link[grDevices]{colors()}}.
}
\examples{
 # plot ICES areas with default projection
 data(ICESareas)
 plotArea(areaDef=ICESareas)

 data(mainareaFdir2018)
 data(NAFOareas)
 # plot mainarea and NAFO areas combined in a Lambert Conformal Conic projection.
 plotArea(title="Main area + NAFO",
          areaDef=rbind(mainareaFdir2018[,c("StratumName")],
                  NAFOareas[,c("StratumName")]),
          projection="+proj=lcc +lat_1=43 +lat_2=62 +lat_0=30 
          +lon_0=10 +x_0=0 +y_0=0 +ellps=intl +units=m +no_defs")
}
