% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxFunctions.R
\name{PrepareReca}
\alias{PrepareReca}
\title{Prepare data for Reca.}
\usage{
PrepareReca(StoxBioticData, StoxLandingData, fixedEffects, randomEffects,
  carEffect = NULL, CarNeighbours = NULL, AgeErrorMatrix = NULL,
  stockSplitting = F, ClassificationErrorMatrix = NULL)
}
\arguments{
\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} data with samples from fisheries and approriate columns appended for identifying corresponding landings.}

\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} data with landings from fisheries and approriate columns appended for identifying corresponding samples}

\item{fixedEffects}{character() vector identifying column names that should be treated as fixed effects}

\item{randomEffects}{character() vector identifying column names that should be treated as fixed effects}

\item{carEffect}{character() identifying the column name that should be treated as CAR-effect (conditional autoregressive effect)}

\item{CarNeighbours}{\code{\link[RstoxFDA]{CarNeighbours}} identifies which values of the carEffect are to be considered as neighbours.}

\item{AgeErrorMatrix}{\code{\link[RstoxFDA]{AgeErrorMatrix}} specifies the probabilities of misreading ages.}

\item{stockSplitting}{logical() whether to run estimate results for separate stocks in the data (coastal cod-analysis)}

\item{ClassificationErrorMatrix}{\code{\link[RstoxFDA]{ClassificationErrorMatrix}} specifies the probability of misclassifying stock for an individual.}
}
\value{
\code{\link[RstoxFDA]{RecaData}}
}
\description{
Prepare data for Reca.
}
\details{
Performs data checks and data conversions,
 and stores some data-related parameters in preparation for running.
 \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}
 via \code{\link[RstoxFDA]{runReca}}.
}
