% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalysisFunctions.R
\name{PrepareRecaEstimate}
\alias{PrepareRecaEstimate}
\title{Prepare data for Reca.}
\usage{
PrepareRecaEstimate(
  StoxBioticData,
  StoxLandingData,
  fixedEffects = NULL,
  randomEffects = NULL,
  UseCarEffect = F,
  carEffect = NULL,
  CarNeighbours = NULL,
  UseAgingError = F,
  AgeErrorMatrix = NULL,
  minAge = integer(),
  maxAge = integer(),
  maxLength = numeric(),
  lengthResolution = numeric(),
  temporalResolution = c("Quarter", "Month"),
  hatchDay = integer()
)
}
\arguments{
\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} data with samples from fisheries
and approriate columns appended for identifying corresponding landings.}

\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} data with landings from fisheries
and approriate columns appended for identifying corresponding samples}

\item{fixedEffects}{optional, vector identifying column names that should be treated as fixed effects. Defaults to no fixed effects.}

\item{randomEffects}{optional, vector identifying column names that should be treated as random effects. Defaults to no random effects.}

\item{UseCarEffect}{logical indicating whether CAR-effect (conditional autoregressive effect) should be used.}

\item{carEffect}{optional, character identifying the column name that should be treated as CAR-effect.
Mandatory if CAR-effect is TRUE
(conditional autoregressive effect).}

\item{CarNeighbours}{\code{\link[RstoxFDA]{CarNeighbours}}, mandatory if 'carEffect' is given.
Identifies which values of the carEffect are to be considered as neighbours.}

\item{UseAgingError}{logical identifying if error-aging parameters should be incorporated in the model}

\item{AgeErrorMatrix}{\code{\link[RstoxFDA]{AgeErrorMatrix}}, optional, specifies the probabilities of misreading ages.
mandatory if UseAgingError is TRUE.}

\item{minAge}{optional, must match dimensions of any 'AgeErrorMatrix'.
If not provided it will be derived from data.}

\item{maxAge}{optional, must match dimensions of any 'AgeErrorMatrix'.
If not provided it will be derived from data.}

\item{maxLength}{optional, maximal fish length in data in cm.
If not provided it will be derived from data.}

\item{lengthResolution}{optional, resolution for length measurements in cm.
If not provided modal value from data is used.}

\item{temporalResolution}{default "Quarter", code for temporal resolution in landings: "Month" or "Quarter".
Regulates temporal resolution for calculating fractional ages of fish.
Not to be confused with any temporal covariate.}

\item{hatchDay}{defaults to 1 representing Jan 1st.
encoding the day of the year when fish is consider to transition from one age to the next.}
}
\value{
\code{\link[RstoxFDA]{RecaData}} Data prepared for running Reca.
}
\description{
StoX-function.
 Performs data checks and data conversions,
 and stores some data-related parameters in preparation for running
 \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}
 via \code{\link[RstoxFDA]{RunRecaEstimate}}.
}
\details{
A covariate indentifying haul is always included in Reca. Do not add haul-identifiers as covariates.
}
