% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalysisFunctions.R
\name{PrepareRecaEstimate}
\alias{PrepareRecaEstimate}
\title{Prepare data for Reca.}
\usage{
PrepareRecaEstimate(
  StoxBioticData,
  StoxLandingData,
  FixedEffects = character(),
  RandomEffects = character(),
  UseCarEffect = F,
  CarEffect = character(),
  CarNeighbours,
  UseAgingError = F,
  AgeErrorMatrix,
  CellEffect = c("Off", "All"),
  MinAge = integer(),
  MaxAge = integer(),
  MaxLength = numeric(),
  LengthResolution = numeric(),
  HatchDay = integer()
)
}
\arguments{
\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} data with samples from fisheries
and approriate columns appended for identifying corresponding landings.}

\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} data with landings from fisheries
and approriate columns appended for identifying corresponding samples}

\item{FixedEffects}{optional, vector identifying column names that should be treated as fixed effects. Defaults to no fixed effects.}

\item{RandomEffects}{optional, vector identifying column names that should be treated as random effects. Defaults to no random effects.}

\item{UseCarEffect}{logical indicating whether CAR-effect (conditional autoregressive effect) should be used.}

\item{CarEffect}{optional, character identifying the column name that should be treated as CAR-effect.
Mandatory if CAR-effect is TRUE
(conditional autoregressive effect).}

\item{CarNeighbours}{\code{\link[RstoxFDA]{CarNeighbours}}, mandatory if 'carEffect' is given.
Identifies which values of the carEffect are to be considered as neighbours.}

\item{UseAgingError}{If true error-aging parameters will be incorporated in the model}

\item{AgeErrorMatrix}{\code{\link[RstoxFDA]{AgeErrorMatrix}}, optional, specifies the probabilities of misreading ages.
mandatory if UseAgingError is TRUE.}

\item{CellEffect}{Configures the cell effect. If 'All', an interaction term will be added with all covariates that exist in landings (whether they are fixed or random effects)}

\item{MinAge}{optional, must match dimensions of any 'AgeErrorMatrix'.
If not provided it will be derived from data.}

\item{MaxAge}{optional, must match dimensions of any 'AgeErrorMatrix'.
If not provided it will be derived from data.}

\item{MaxLength}{optional, maximal fish length in data in cm.
If not provided it will be derived from data.}

\item{LengthResolution}{optional, resolution for length measurements in cm.
If not provided modal value from data is used.}

\item{HatchDay}{defaults to 1 representing Jan 1st.
encoding the day of the year when fish is consider to transition from one age to the next.}
}
\value{
\code{\link[RstoxFDA]{RecaData}} Data prepared for running Reca.
}
\description{
Performs data checks and data conversions,
 and stores some data-related parameters in preparation for running
 \code{\link[Reca]{eca.estimate}}
 via \code{\link[RstoxFDA]{ParameterizeRecaModels}}.
}
\details{
A covariate indentifying haul or landing is always included in Reca. Do not add haul-identifiers as covariates.
}
