% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxFunctions.R
\name{RecaData}
\alias{RecaData}
\title{RecaData}
\description{
Data and some data parameters prepared for running
\code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}
via \code{\link[RstoxFDA]{RunReca}}.
}
\details{
\describe{
 \item{AgeLength}{input needed for \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}}
 \item{WeightLength}{input needed for \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}}
 \item{Landings}{input needed for \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}}
 \item{GlobalParameters}{input needed for \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}. see details}
 \item{CovariateMaps}{Mapping of values for each covariate in landings and samples (including non-configurable catchId) to integer value used in R-ECA.}
}
}
