% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recaDAta-datadoc.R
\docType{data}
\name{recaDataExample}
\alias{recaDataExample}
\title{Example data for Reca}
\format{
see \code{\link[RstoxFDA]{prepRECA}}.
}
\usage{
data(recaDataExample)
}
\description{
Example is produced by \code{\link[RstoxFDA]{prepRECA}} and is suited for input to \code{\link[RstoxFDA]{runRECA}}
}
\examples{
data(recaDataExample)
\dontrun{recaResult <- runRECA(recaData, 500, 5000)}
}
\keyword{datasets}
