% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{RecaParameterData}
\alias{RecaParameterData}
\title{Reca Parameter Data (RecaParameterData)}
\description{
Data and some data parameters prepared for running
various report functions that invoke \code{\link[Reca]{eca.predict}}.
}
\section{model fit}{

 For inspection or analysis of model fit, the lists 'FitProportionAtAge', 
 'FitLengthGivenAge' and 'FitWeightGivenLength' is of interest. For stock-splitting analysis,
 the lists FitLengthGivenAgeCC and FitWeightGivenLengthCC will be added as well, corresponding to one of the stocks.
 These lists correspond to the three Reca-models and contain:
 \describe{
 \item{LogLikeliehood}{A \code{\link[data.table]{data.table}} 
   tabulating the logarithm of the likeliehood of the parameter set for each iteration}
 \item{...}{A \code{\link[data.table]{data.table}} for each of the model effects (e.g. covariates).}
 }
 
 In addition to configurable covariates, the models always contain a constant effect (named 'constant'),
 a catch or haul effect (named 'catchSample') and effects for fish measurments (named 'fish'). 
 Where relevant the following parameters may be tabulated for each effect:
 \describe{
 \item{Age}{Identifying the age the effect applies to}
 \item{Level}{Identifying the value or level of the covariate the effect applies to}
 \item{Iteration}{Identifying the iteration the fit is provided for}
 \item{AgeIndex}{Age identifier used internally in Reca}
 \item{LevelIndex}{Level identifier used internally in Reca}
 \item{Slope}{The value of the regression slope}
 \item{tau_Slope}{The value of tau parameter for the regression slope}
 \item{ar_Slope}{The value of regression slope of a the autoregressive coefficient associated with the effect}
 \item{Intercept}{The value of the regression intercept}
 \item{tau_Intercept}{The value of tau parameter for the regression intercept}
 \item{ar_Intercept}{The value of the regression intercept of a autoregressive coefficient associated with the effect}
 }
 Consult Hirst et.al. 2005 for description of the parameters
 
 @section other data:
 The lists 'AgeLength', 'WeightLength', 'Landings', 'GlobalParameters' and 'CovariateMaps'
 may be passed to \code{\link[Reca]{eca.predict}} in functions consuming output from this function. All in all
 the following lists can be accessed on RecaParameterData objects:
 \describe{
 \item{FitProportionAtAge}{list of data tables with parameters for for the Proportion-at-age model}
 \item{FitLengthGivenAge}{list of data tables with parameters for for the Length-given-age model}
 \item{FitWeightGivenLength}{list of data tables with parameters for for the Weight-given-length model}
 \item{AgeLength}{input needed for \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}}
 \item{WeightLength}{input needed for \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}}
 \item{Landings}{input needed for \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}}
 \item{GlobalParameters}{input needed for \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}. see details}
 \item{CovariateMaps}{Mapping of values for each covariate in landings and samples (including non-configurable catchId) to integer value used in R-ECA.}
}
}

