% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{RecaParameterData}
\alias{RecaParameterData}
\title{Reca Parameter Data (RecaParameterData)}
\description{
Data and some data parameters prepared for running
various report functions that invoke \code{\link[Reca]{eca.predict}}
}
\details{
\describe{
 \item{FitProportionAtAge}{list of data tables with parameters for for the Proportion-at-age model}
 \item{FitLengthGivenAge}{list of data tables with parameters for for the Length-given-age model}
 \item{FitWeightGivenLength}{list of data tables with parameters for for the Weight-given-length model}
 \item{AgeLength}{input needed for \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}}
 \item{WeightLength}{input needed for \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}}
 \item{Landings}{input needed for \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}}
 \item{GlobalParameters}{input needed for \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}. see details}
 \item{CovariateMaps}{Mapping of values for each covariate in landings and samples (including non-configurable catchId) to integer value used in R-ECA.}
}
}
