% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recaResult-datadoc.R
\docType{data}
\name{recaPrediction}
\alias{recaPrediction}
\title{Example result from Reca}
\format{
see element prediction in return from \code{\link[RstoxFDA]{runRECA}}.
}
\usage{
data(recaPrediction)
}
\description{
Example is produced by \code{\link[RstoxFDA]{runRECA}}
}
\examples{
data(recaPrediction)
plotCatchAtAge(recaPrediction)
}
\keyword{datasets}
