% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{RecaResult}
\alias{RecaResult}
\title{Reca Results (RecaResult)}
\description{
Results from running
\code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}
via \code{\link[RstoxFDA]{RunRecaEstimate}}.
}
\details{
\describe{
 \item{input}{All input data and parameters provided to \code{\link[Reca]{eca.estimate}} and \code{\link[Reca]{eca.predict}}}
 \item{fit}{as returned by \code{\link[Reca]{eca.estimate}}}
 \item{prediction}{as returned by \code{\link[Reca]{eca.predict}}}
 \item{covariateMaps}{list() mapping from Reca covariate encoding to values fed to \code{\link[RstoxFDA]{PrepareRecaEstimate}}. As in \code{\link[RstoxFDA]{RecaData}}}
}
}
