% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportFdaByAgeData}
\alias{ReportFdaByAgeData}
\title{Age group statistics (ReportFdaByAgeData)}
\description{
Results from catch at age estimations. The results may be presented
 decomposed on combinations of aggregation variables, such as gear, area, stock etc.
 
 list with two members 'FdaReport' and 'GroupingVariables'.
 'FdaReport' is a \code{\link[data.table]{data.table}} with the columns:
 \describe{
  \item{AgeGroup}{character. The age group the estimate is reported for. May be age or plus group}
  \item{Age}{integer. The lower age the estimate is reported for. May be an age or lower limit of plus group (inclusive)}
  \item{<Statistic>}{A reported statistic}
  \item{SD}{Standard deviation for the reported statistic.}
  \item{Low}{The lower limit of the estimated interval for the reported statistic.}
  \item{High}{The higher limit of the estimated interval for the reported statistic.}
  \item{...}{Any aggregation variables. The names of these are listed in 'GroupingVariables'}
 }
 'GroupingVariables' is a \code{\link[data.table]{data.table}} with a column containing the names of any aggregation variables.
}
