% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportFdaCatchAtAgeData}
\alias{ReportFdaCatchAtAgeData}
\title{Fisheries dependent Catch At Age Report (ReportFdaCatchAtAgeData)}
\description{
A \code{\link[RstoxFDA]{ReportFdaByAgeData}} object with the reported <Statistic> being:
 
 \describe{
  \item{CacthAtAge}{The total catch at age in numbers.}
 }
}
