% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportFdaCatchAtLengthData}
\alias{ReportFdaCatchAtLengthData}
\title{Fisheries dependent Catch At Age Report (ReportFdaCatchAtLengthData)}
\description{
A list with two members: 'NbyLength', and 'GroupingVariables'.
 
 \describe{
  \item{NbyLength}{A \code{\link[RstoxFDA]{ReportFdaByAgeData}} table with reported <Statistic> being 'CacthAtLength': the total catch at length in numbers.}
  \item{GroupingVariables}{Any specified Grouping variables.}
 }
 
 Units are configurable, and can be inspected by ~\code{\link[RstoxData]{getUnit}}
 
 'GroupingVariables' is a \code{\link[data.table]{data.table}} with a column containing the names of any aggregation variables.
}
