% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxReportFunctions.R
\name{ReportFdaLandings}
\alias{ReportFdaLandings}
\title{Report FDA landings}
\usage{
ReportFdaLandings(
  StoxLandingData,
  GroupingVariables = character(),
  Decimals = integer(),
  Unit = RstoxData::getUnitOptions("mass", conversionRange = c(1, 1e+12))
)
}
\arguments{
\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} data with landings from fisheries
and approriate columns added for identifying corresponding samples}

\item{GroupingVariables}{Columns of 'StoxBioticData' and 'StoxLandingData' that partitions the fisheries. If not provided, a single row for all landings will be produced.}

\item{Decimals}{integer specifying the number of decimals to report for 'LandedRoundWeight' and 'WeightOfSampledCatches'. Defaults to zero.}

\item{Unit}{unit for the weights 'LandedRoundWeight' and 'WeightOfSampledCatches'. Defaults to 'kg'}
}
\value{
\code{\link[RstoxFDA]{ReportFdaLandingData}}
}
\description{
Report landings in partitions of the fisheries.
}
\details{
Landings are reported partitioned on the provided 'GroupingVariables'.
 
 Landings are sorted by decreasing weight, except if the column 'CatchDate' is used as GroupingVariable.
 In that case Landings are sorted by increasing date, after sorting on other Grouping variables.
 
 Rounding of numbers according to the argument 'Decimals' is done with \code{\link[base]{round}},
 so that negative numbers specify rounding to powers of ten, and rounding of the digit 5 is towards the even digit.

 The units considered valid for weights are those listed for quantity 'mass' in \code{\link[RstoxData]{StoxUnits}}
}
