% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportFdaLengthAtAgeData}
\alias{ReportFdaLengthAtAgeData}
\title{Fisheries dependent Length At Age Report (ReportFdaLengthAtAgeData)}
\description{
A list with two members: 'MeanLengthByAge', and 'GroupingVariables'.
 
 \describe{
  \item{MeanLengthByAge}{A \code{\link[RstoxFDA]{ReportFdaData}} table with reported <Statistic> being 'MeanIndividualLength': the mean length.}
  \item{GroupingVariables}{Any specified Grouping variables.}
 }
 
 Units are configurable, and can be inspected by ~\code{\link[RstoxData]{getUnit}}
 
 'GroupingVariables' is a \code{\link[data.table]{data.table}} with a column containing the names of any aggregation variables.
 
 Note that the summary statistics are reported for summaries of mean lengths, 
 so that e.g. SD report the standard deviation of the means,
 and does not characterize the length distribution of fish.
}
\concept{Data types}
