% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxReportFunctions.R
\name{ReportFdaSampling}
\alias{ReportFdaSampling}
\title{Report FDA sampling}
\usage{
ReportFdaSampling(
  StoxBioticData,
  StoxLandingData,
  GroupingVariables = character(),
  Decimals = integer(),
  Unit = RstoxData::getUnitOptions("mass", conversionRange = c(1, 1e+12))
)
}
\arguments{
\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} data with samples from fisheries
and approriate columns added for identifying corresponding landings.}

\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} data with landings from fisheries
and approriate columns added for identifying corresponding samples}

\item{GroupingVariables}{Columns of 'StoxBioticData' and 'StoxLandingData' that partitions the fisheries. Defaults to all column names that are found in both inputs.}

\item{Decimals}{integer specifying the number of decimals to report for 'LandedRoundWeight' and 'WeightOfSampledCatches'. Defaults to zero.}

\item{Unit}{unit for the weights 'LandedRoundWeight' and 'WeightOfSampledCatches'. Defaults to 'kg'}
}
\value{
\code{\link[RstoxFDA]{ReportFdaSamplingData}}
}
\description{
Report sampling of fisheries against landings in partitions of the fisheries.
}
\details{
Sampling is reported partitioned on the provided 'GroupingVariables'.
 If samples are encoded in partitions of the fisheries with no landings. 'LandedRoundWeight' will be NA.
 This may be due to recording errors or filtering errors, but it may also be due to comparison of similar
 but unequal category-definitios. For instance area are coded in landings as dominant area for a fishing trip,
 while at-sea sampling will record area of fishing operation, and the catch from that area by subsequently be landed
 with another area listed as dominant area.
 
 Rounding of numbers according to the argument 'Decimals' is done with \code{\link[base]{round}},
 so that negative numbers specify rounding to powers of ten, and rounding of the digit 5 is towards the even digit.

 The units considered valid for weights are those listed for quantity 'mass' in \code{\link[RstoxData]{StoxUnits}}
}
