% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxReportFunctions.R
\name{ReportFdaSOP}
\alias{ReportFdaSOP}
\title{Report SOP test}
\usage{
ReportFdaSOP(
  ReportFdaCatchAtAgeData,
  ReportFdaWeightAtAgeData,
  StoxLandingData,
  GroupingVariables = character()
)
}
\arguments{
\item{ReportFdaCatchAtAgeData}{\code{\link[RstoxFDA]{ReportFdaCatchAtAgeData}} with estimates of total catch at age}

\item{ReportFdaWeightAtAgeData}{\code{\link[RstoxFDA]{ReportFdaWeightAtAgeData}} with estimates of mean weight at age for individual fish}

\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} data with landings from fisheries}

\item{GroupingVariables}{Columns of 'StoxLandingData' that partitions the landings into groups SOP tests should be reported for.}
}
\value{
\code{\link[RstoxFDA]{ReportFdaSopData}}
}
\description{
Report sum-of-product test (SOP-test) for catch estimates

 Mean weight at age and estimated catch at age is used to compute total catches
 and the relative difference to reported landings are reported.
 
 The report will be generated for landings decomposed on the provided 'GroupingVariables'
 which must be available in both 'ReportFdaCatchAtAgeData' and 'ReportFdaWeightAtAgeData'
 and 'StoxLandingData'.
 
 'ReportFdaCatchAtAgeData' and 'ReportFdaWeightAtAgeData' must be decomposed on the same
 'GroupingVariables' and must be reported for the same age groups
}
\seealso{
\code{\link[RstoxFDA]{ReportRecaWeightAtAge}} and \code{\link[RstoxFDA]{ReportRecaCatchAtAge}} for some ways of preparing 'ReportFdaWeightAtAgeData' and 'ReportFdaCatchAtAgeData'.
 \code{\link[RstoxData]{StoxLanding}} and \code{\link[RstoxData]{FilterStoxLanding}} for ways of preparing 'StoxLandingData'.
}
