% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxReportFunctions.R
\name{ReportFdaSOP}
\alias{ReportFdaSOP}
\title{Report SOP test}
\usage{
ReportFdaSOP(
  ReportFdaCatchAtAgeData,
  ReportFdaWeightAtAgeData,
  StoxLandingData,
  GroupingVariables = character(),
  DecimalWeight = integer(),
  DecimalFraction = integer(),
  UnitFraction = RstoxData::getUnitOptions("fraction")
)
}
\arguments{
\item{ReportFdaCatchAtAgeData}{\code{\link[RstoxFDA]{ReportFdaCatchAtAgeData}} with estimates of total catch at age}

\item{ReportFdaWeightAtAgeData}{\code{\link[RstoxFDA]{ReportFdaWeightAtAgeData}} with estimates of mean weight at age for individual fish}

\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} data with landings from fisheries}

\item{GroupingVariables}{Columns of 'StoxLandingData' that partitions the landings into groups SOP tests should be reported for.}

\item{DecimalWeight}{integer specifying the number of decimals to report for weights: 'TotalWeightEstimated', 'LandedWeight', and 'Difference'. Defaults to 0}

\item{DecimalFraction}{integer specifying the number of decimals to report for 'RelativeDifference'. Defaults to 3.}

\item{UnitFraction}{unit for the RelativeDifference. E.g. '0.' for decimal notation or '%' for percent.}
}
\value{
\code{\link[RstoxFDA]{ReportFdaSopData}}
}
\description{
Report sum-of-product test (SOP-test) for catch estimates.

 Mean weight at age and estimated catch (numbers) at age is used to compute total catches
 and the relative difference to reported landings are reported.
 
 The report will be generated for landings decomposed on the provided 'GroupingVariables'
 which must be available in both 'ReportFdaCatchAtAgeData' and 'ReportFdaWeightAtAgeData'
 and 'StoxLandingData'.
 
 'ReportFdaCatchAtAgeData' and 'ReportFdaWeightAtAgeData' must be decomposed on the same
 'GroupingVariables' and must be reported for the same age groups
 
 Rounding of numbers according to the argument 'Decimals' is done with \code{\link[base]{round}},
 so that negative numbers specify rounding to powers of ten, and rounding of the digit 5 is towards the even digit.
 
 The units considered valid for mean lengths are those listed for quantity 'fraction' in \code{\link[RstoxData]{StoxUnits}}
}
\seealso{
\code{\link[RstoxFDA]{ReportRecaWeightAtAge}} and \code{\link[RstoxFDA]{ReportRecaCatchAtAge}} for some ways of preparing 'ReportFdaWeightAtAgeData' and 'ReportFdaCatchAtAgeData'.
 \code{\link[RstoxData]{StoxLanding}} and \code{\link[RstoxData]{FilterStoxLanding}} for ways of preparing 'StoxLandingData'.
}
