% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportFdaSopData}
\alias{ReportFdaSopData}
\title{Sum of Products report (ReportFdaSopData)}
\description{
Sum of Products report (SOP-report), comparing the total landed weight of fish
 with the product of mean weight at age estimates and total number 
 at age estimates 
 
 list with two members 'SopReport' and 'GroupingVariables'.
 'SopReport' is a \code{\link[data.table]{data.table}} with the columns:
 \describe{
  \item{TotalWeightEstimated}{Total round weight (kg) estimated}
  \item{LandedWeight}{Landed round weight (kg) reported}
  \item{Difference}{The difference between estimated and reported landed weight}
  \item{RelativeDifference}{The difference between estimated and reported landed weight relative to reported weight}
  \item{...}{Any aggregation variables. The names of these are listed in 'GroupingVariables'}
 }
 'GroupingVariables' is a \code{\link[data.table]{data.table}} with a column containing the names of any aggregation variables.
}
