% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportFdaSummaryData}
\alias{ReportFdaSummaryData}
\title{Total catch statistics (ReportFdaSummaryData)}
\description{
Results from catch estimations. The results may be presented
 decomposed on combinations of aggregation variables, such as gear, area, stock etc.
 
 list with six members 'MeanAge', 'MeanWeight', 'MeanLength', 'TotalWeight',
 'TotalNumber', and 'GroupingVariables'.
 'MeanAge', 'MeanWeight', 'MeanLength', 'TotalWeight',
 'TotalNumber' are \code{\link[data.table]{data.table}}s with the columns:
 \describe{
  \item{<Statistic>}{The reported statistic, either 'MeanIndividualAge', 'MeanIndividualWeight', 'MeanIndividualLength', 'TotalWeight', or 'TotalNumber'}
  \item{SD}{Standard deviation for the reported statistic.}
  \item{Low}{The lower limit of the estimated interval for the reported statistic.}
  \item{High}{The higher limit of the estimated interval for the reported statistic.}
  \item{...}{Any aggregation variables. The names of these are listed in 'GroupingVariables'}
 }
 'GroupingVariables' is a \code{\link[data.table]{data.table}} with a column containing the names of any aggregation variables.
}
