% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxReportFunctions.R
\name{ReportRecaCatchAtAge}
\alias{ReportRecaCatchAtAge}
\title{Report catch at age}
\usage{
ReportRecaCatchAtAge(RecaCatchAtAge)
}
\arguments{
\item{RecaCatchAtAge}{Results from MCMC simulations (\code{\link[RstoxFDA]{RecaCatchAtAge}}).}
}
\value{
\code{\link[RstoxFDA]{ReportRecaCatchAtAgeData}}
}
\description{
Tabulates summary statistics from MCMC simulations using Reca.
 Typically obtained with \code{\link[RstoxFDA]{RunRecaModels}}.
 
 If 'RecaCatchAtAge' contains estimate for a set of aggregation variables, such as
 area, gear, stock, etc., summary statistics will be presented similarly.
}
