% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{ReportRecaCatchAtAgeData}
\alias{ReportRecaCatchAtAgeData}
\title{Reca Catch At Age Report (ReportRecaCatchAtAgeData)}
\description{
A \code{\link[RstoxFDA]{ReportRecaData}} object with the reported <Statistic> being:
 
 \describe{
  \item{CacthAtAge}{The total catch at age in numbers.}
 }
}
