% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxReportFunctions.R
\name{ReportRecaCatchStatistics}
\alias{ReportRecaCatchStatistics}
\title{Report catch statistics}
\usage{
ReportRecaCatchStatistics(
  RecaCatchAtAge,
  IntervalWidth = numeric(),
  DecimalOptions = FALSE,
  DecimalTotalNumber = integer(),
  DecimalTotalWeight = integer(),
  DecimalMeanAge = integer(),
  DecimalMeanWeight = integer(),
  DecimalMeanLength = integer(),
  UnitOptions = FALSE,
  UnitTotalNumber = c("Mi", "i", "ki"),
  UnitTotalWeight = c("kt", "t", "kg"),
  UnitMeanWeight = c("kg", "g"),
  UnitMeanLength = c("cm", "mm", "m")
)
}
\arguments{
\item{RecaCatchAtAge}{Results from MCMC simulations (\code{\link[RstoxFDA]{RecaCatchAtAge}}).}

\item{IntervalWidth}{The width of the reported credible interval. Defaults to 0.9 for 90 per cent credible intervals.}

\item{DecimalOptions}{logical determining whether the StoX user interface should show decimal options}

\item{DecimalTotalNumber}{integer specifying the number of decimals to report for 'TotalNumber', and the corresponding 'SD', 'Low' and 'High'. to 0}

\item{DecimalTotalWeight}{integer specifying the number of decimals to report for 'TotalWeightDefaults', and the corresponding 'SD', 'Low' and 'High'. to 0}

\item{DecimalMeanAge}{integer specifying the number of decimals to report for 'MeanIndividualAge', and the corresponding 'SD', 'Low' and 'High'. Defaults to 1}

\item{DecimalMeanWeight}{integer specifying the number of decimals to report for 'MeanIndividualWeight', and the corresponding 'SD', 'Low' and 'High'. Defaults to 3}

\item{DecimalMeanLength}{integer specifying the number of decimals to report for 'MeanIndividualLength', and the corresponding 'SD', 'Low' and 'High'. Defaults to 2}

\item{UnitOptions}{logical determining whether the StoX user interface should show unit options}

\item{UnitTotalNumber}{unit for total catch in numbers. Defaults to Mi (millions)}

\item{UnitTotalWeight}{unit for weight of total catch. Defaults to kt}

\item{UnitMeanWeight}{unit for mean weight. Defaults to kg}

\item{UnitMeanLength}{unit for mean length. Defaults to cm.}
}
\value{
\code{\link[RstoxFDA]{ReportFdaSummaryData}}
}
\description{
Report summary statistics for landed catches from MCMC simulations using Reca.
 Summary statistics are obtained from the posterior distribution, and
 the interval is reported as equal-tailed credible intervals are reported.
 
 Weights are reported in kg, lengths in cm.
 
 If 'RecaCatchAtAge' contains estimate for a set of aggregation variables, such as
 area, gear, stock, etc., summary statistics will be presented similarly.
 
 Rounding of numbers according to the argument 'Decimals' is done with \code{\link[base]{round}},
 so that negative numbers specify rounding to powers of ten, and rounding of the digit 5 is towards the even digit.
 
 The units considered valid for mean lengths are those listed for quantity 'length' in ~\code{\link[RstoxData]{StoxUnits}}
 The units considered valid for weights are those listed for quantity 'mass' in ~\code{\link[RstoxData]{StoxUnits}}
 The units considered valid for total catch in numbers are those listed for quantity 'cardinality' in ~\code{\link[RstoxData]{StoxUnits}}
}
\seealso{
\code{\link[RstoxFDA]{RunRecaModels}} for running Reca-analysis
}
