% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxReportFunctions.R
\name{ReportRecaCatchStatistics}
\alias{ReportRecaCatchStatistics}
\title{Report catch statistics}
\usage{
ReportRecaCatchStatistics(
  RecaCatchAtAge,
  IntervalWidth = numeric(),
  UseDefaultDecimalOptions = TRUE,
  DecimalTotalNumber = integer(),
  DecimalTotalWeight = integer(),
  DecimalMeanAge = integer(),
  DecimalMeanWeight = integer(),
  DecimalMeanLength = integer(),
  UseDefaultUnitOptions = TRUE,
  UnitTotalNumber = RstoxData::getUnitOptions("cardinality", conversionRange = c(1,
    1e+12)),
  UnitTotalWeight = RstoxData::getUnitOptions("mass", conversionRange = c(1, 1e+12)),
  UnitMeanWeight = RstoxData::getUnitOptions("mass", conversionRange = c(1e-04, 10)),
  UnitMeanLength = RstoxData::getUnitOptions("length", conversionRange = c(1e-04, 10))
)
}
\arguments{
\item{RecaCatchAtAge}{Results from MCMC simulations (\code{\link[RstoxFDA]{RecaCatchAtAge}}).}

\item{IntervalWidth}{The width of the reported credible interval. Defaults to 0.9 for 90 per cent credible intervals.}

\item{UseDefaultDecimalOptions}{logical determining whether to use default decimal options.}

\item{DecimalTotalNumber}{integer specifying the number of decimals to report for 'TotalNumber', and the corresponding 'SD', 'Low' and 'High'. to 0}

\item{DecimalTotalWeight}{integer specifying the number of decimals to report for 'TotalWeightDefaults', and the corresponding 'SD', 'Low' and 'High'. to 0}

\item{DecimalMeanAge}{integer specifying the number of decimals to report for 'MeanIndividualAge', and the corresponding 'SD', 'Low' and 'High'. Defaults to 1}

\item{DecimalMeanWeight}{integer specifying the number of decimals to report for 'MeanIndividualWeight', and the corresponding 'SD', 'Low' and 'High'. Defaults to 3}

\item{DecimalMeanLength}{integer specifying the number of decimals to report for 'MeanIndividualLength', and the corresponding 'SD', 'Low' and 'High'. Defaults to 2}

\item{UseDefaultUnitOptions}{logical determining whether to use default unit options.}

\item{UnitTotalNumber}{unit for total catch in numbers. Defaults to Mi (millions)}

\item{UnitTotalWeight}{unit for weight of total catch. Defaults to kt}

\item{UnitMeanWeight}{unit for mean weight. Defaults to kg}

\item{UnitMeanLength}{unit for mean length. Defaults to cm.}
}
\value{
\code{\link[RstoxFDA]{ReportFdaSummaryData}}
}
\description{
Report summary statistics for landed catches from MCMC simulations using Reca.
This function reports estimated total catch, mean length, mean weight, and mean age.

If 'RecaCatchAtAge' contains estimate for a set of aggregation variables, such as
area, gear, stock, etc., summary statistics will be presented similarly.

The units considered valid for mean lengths are those listed for quantity 'length' in \code{\link[RstoxData]{StoxUnits}}
The units considered valid for weights are those listed for quantity 'mass' in \code{\link[RstoxData]{StoxUnits}}
The units considered valid for total catch in numbers are those listed for quantity 'cardinality' in \code{\link[RstoxData]{StoxUnits}}

Summary statistics are obtained from the posterior distribution, and
the interval is reported as equal-tailed credible intervals are reported.

Rounding of numbers according to the argument 'Decimals' is done with \code{\link[base]{round}},
so that negative numbers specify rounding to powers of ten, and rounding of the digit 5 is towards the even digit.
}
\seealso{
\code{\link[RstoxFDA]{RunRecaModels}} for running Reca-analysis
\code{\link[RstoxFDA]{ReportRecaCatchAtAge}} for reporting catch, mean length and mean weight by age groups
\code{\link[RstoxFDA]{ReportRecaCatchAtLength}} for reporting length distributions
\code{\link[RstoxFDA]{ReportRecaCatchAtLengthAndAge}} for reporting catch by age-group and length-group combinations.
}
\concept{StoX-Reca functions}
\concept{StoX-functions}
