% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxReportFunctions.R
\name{ReportRecaLengthAtAge}
\alias{ReportRecaLengthAtAge}
\title{Report length at age}
\usage{
ReportRecaLengthAtAge(RecaCatchAtAge, PlusGroup = integer())
}
\arguments{
\item{RecaCatchAtAge}{Results from MCMC simulations (\code{\link[RstoxFDA]{RecaCatchAtAge}}).}

\item{PlusGroup}{If given, ages 'PlusGroup' or older are included in a plus group.}
}
\value{
\code{\link[RstoxFDA]{ReportFdaLengthAtAgeData}}
}
\description{
Tabulates summary statistics for mean length at age from MCMC simulations using Reca.
 MCMC simulations are typically obtained with \code{\link[RstoxFDA]{RunRecaModels}}.
 Summary statistics are obtained from the posterior distribution, and
 credible intervals are equal-tailed intervals.
 
 If 'RecaCatchAtAge' contains estimate for a set of aggregation variables, such as
 area, gear, stock, etc., summary statistics will be presented similarly.
}
\seealso{
\code{\link[RstoxFDA]{RunRecaModels}} for running Reca-analysis
}
