% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxReportFunctions.R
\name{ReportRecaLengthAtAge}
\alias{ReportRecaLengthAtAge}
\title{Report length at age}
\usage{
ReportRecaLengthAtAge(
  RecaCatchAtAge,
  PlusGroup = integer(),
  IntervalWidth = numeric(),
  Decimals = integer(),
  Threshold = numeric(),
  Unit = RstoxData::getUnitOptions("length", conversionRange = c(1e-07, 10))
)
}
\arguments{
\item{RecaCatchAtAge}{Results from MCMC simulations (\code{\link[RstoxFDA]{RecaCatchAtAge}}).}

\item{PlusGroup}{If given, ages 'PlusGroup' or older are included in a plus group.}

\item{IntervalWidth}{The width of the reported credible interval. Defaults to 0.9 for 90 per cent credible intervals.}

\item{Decimals}{integer specifying the number of decimals to report for 'MeanIndividualLength', 'SD', 'Low' and 'High'. Defaults to 1.}

\item{Threshold}{threshold for reporting mean weight. Rows with an estimated Catch At Age (number of individuals) lower than this will have NA reported for their mean length Defaults to 0.}

\item{Unit}{unit for 'MeanIndividualLength', 'SD', 'Low' and 'High'}
}
\value{
\code{\link[RstoxFDA]{ReportFdaLengthAtAgeData}}
}
\description{
Tabulates summary statistics for mean length at age from MCMC simulations using Reca.
 MCMC simulations are typically obtained with \code{\link[RstoxFDA]{RunRecaModels}}.
 Summary statistics are obtained from the posterior distribution, and
 the interval is reported as equal-tailed credible intervals are reported.
 
 Mean length for plus-groups are a weighted by the relative catch-at-age in each composite age group.
 For iterations where all of the plus-group ages have a zero catch at age, this weight is not defined,
 and summary statistics are obtained from the remaining iterations.
 
 If 'RecaCatchAtAge' contains estimate for a set of aggregation variables, such as
 area, gear, stock, etc., summary statistics will be presented similarly.
 
 Rounding of numbers according to the argument 'Decimals' is done with \code{\link[base]{round}},
 so that negative numbers specify rounding to powers of ten, and rounding of the digit 5 is towards the even digit.
 
 The units considered valid for mean lengths are those listed for quantity 'length' in \code{\link[RstoxData]{StoxUnits}}
}
\seealso{
\code{\link[RstoxFDA]{RunRecaModels}} for running Reca-analysis

Other Reca functions: 
\code{\link{AddStratumStoxLanding}()},
\code{\link{DefineAgeErrorMatrix}()},
\code{\link{DefineStockSplittingParameters}()},
\code{\link{ParameterizeRecaModels}()},
\code{\link{PrepareRecaEstimate}()},
\code{\link{ReportRecaCatchAtAgeCovariance}()},
\code{\link{ReportRecaCatchAtAge}()},
\code{\link{ReportRecaCatchAtLengthAndAge}()},
\code{\link{ReportRecaCatchAtLength}()},
\code{\link{ReportRecaCatchStatistics}()},
\code{\link{ReportRecaParameterStatistics}()},
\code{\link{ReportRecaWeightAtAge}()},
\code{\link{RunRecaModels}()}
}
\concept{Reca functions}
