% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalysisFunctions.R
\name{RunRecaModels}
\alias{RunRecaModels}
\title{Run Reca Models}
\usage{
RunRecaModels(RecaParameterData, Caa.burnin = numeric(), Seed = numeric())
}
\arguments{
\item{Caa.burnin}{see documentation for \code{\link[Reca]{eca.predict}}. Defaults to 0.}

\item{Seed}{see documentation for \code{\link[Reca]{eca.estimate}}. Defaults to random seed.}
}
\value{
RecaCatchAtAge
}
\description{
Run Reca Models
}
\details{
Runs prediction for parameterized Reca models
}
