% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxAnalysisFunctions.R
\name{RunRecaModels}
\alias{RunRecaModels}
\title{Run Reca Models}
\usage{
RunRecaModels(
  RecaParameterData,
  StoxLandingData,
  GroupingVariables = character(),
  TemporalResolution = c("Quarter", "Month"),
  Caa.burnin = numeric(),
  Seed = numeric(),
  CollapseLength = TRUE
)
}
\arguments{
\item{RecaParameterData}{Parameters for Reca models.}

\item{StoxLandingData}{Landings data (\code{\link[RstoxData]{StoxLandingData}}).}

\item{GroupingVariables}{character vector identifying columns in 'StoxLandingData' that results should be provided for.}

\item{TemporalResolution}{default "Quarter", code for temporal resolution in landings: "Month" or "Quarter".
Regulates temporal resolution for calculating fractional ages of fish.
Not to be confused with any temporal covariate.}

\item{Caa.burnin}{see documentation for \code{\link[Reca]{eca.predict}}. Defaults to 0.}

\item{Seed}{see documentation for \code{\link[Reca]{eca.estimate}}. Defaults to seed stored in 'RecaParameterData'.}

\item{CollapseLength}{indicates whether length groups should be collapsed in result. Defaults to TRUE. See details.}
}
\value{
\code{\link[RstoxFDA]{RecaCatchAtAge}}
}
\description{
Runs prediction (catch-at-age estimate) for parameterized Reca models.
}
\details{
Parameters may be obtained with \code{\link[RstoxFDA]{ParameterizeRecaModels}}.
If the function-parameter 'GroupingVariables' is provided, predictions will be provided for corresponding partitions of landings.
The parameter 'StoxLandingData' may differ from the landings used in parameterisation (passed to \code{\link[RstoxFDA]{ParameterizeRecaModels}}),
as long as all not additional values / levels for the model covariates / effects are introduced.

If The models are configured for stock-splitting analysis. The variable 'Stock' will be added to 'GroupingVariables' in the return value (\code{\link[RstoxFDA]{RecaCatchAtAge}})

If the 'GroupingVariables' specify a very large number of partitions of the landings, this function may
exhaust available computer memory.
\subsection{CollapseLength}{

By default length groups are collapsed into one length group in the result. This does not facilitate reporting
length resolved data, such as length distributions.
The full age-length prediction may be extracted by setting the parameter 'CollapseLength' to FALSE.
If this is used in combination with several grouping variables, there is some risk of exhausting available computer memory.
}
}
\seealso{
\code{\link[RstoxFDA]{ParameterizeRecaModels}} for model parameterisation,
\code{\link[RstoxFDA]{ReportRecaCatchAtAge}},
\code{\link[RstoxFDA]{ReportRecaLengthAtAge}},
\code{\link[RstoxFDA]{ReportRecaWeightAtAge}} for compiling reports of predictions / estimates.
}
\concept{StoX-Reca functions}
\concept{StoX-functions}
