% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{SetShortGearBiotic}
\alias{SetShortGearBiotic}
\title{Set short gear codes biotic}
\usage{
SetShortGearBiotic(BioticData)
}
\arguments{
\item{BioticData}{\code{\link[RstoxData]{BioticData}} data for which short gear codes should be set}
}
\value{
\code{\link[RstoxData]{BioticData}}
}
\description{
Set short gear codes for all stations. Appropriate when BioticData is read from NMDbiotic.
}
\details{
Set the column 'gear' on the table 'fishstation' in \code{\link[RstoxData]{BioticData}} to the first two characters / digits, 
 for all rows where the gear code currently consist of four characters.
 
 The gear codes in NMDBiotic are conventionally organised hierarchically, so that the two first characters denote a gear group.
 The two character/digit system is more convenient to work with in many circumstances.
 
 'gear' is on the table 'fishstation' in data originating from NMDBiotic (http://www.imr.no/formats/nmdbiotic/).
 For bioticdata that does not conform to this, no modifications are done.
}
\seealso{
\code{\link{RstoxData}{RstoxData::StoxBiotic}} For converting \code{\link[RstoxData]{BioticData}} to \code{\link[RstoxData]{StoxBioticData}}.
}
\concept{StoX-functions}
\concept{gear coding functions}
\concept{nmdbiotic functions}
