% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxBaselineFunctions.R
\name{SetStartDateBiotic}
\alias{SetStartDateBiotic}
\title{Set startdate Biotic}
\usage{
SetStartDateBiotic(BioticData, Overwrite = F)
}
\arguments{
\item{BioticData}{\code{\link[RstoxData]{BioticData}} data for which time should be set}

\item{Overwrite}{if True any existing values in stationstartdate will be overwritten.}
}
\value{
\code{\link[RstoxData]{BioticData}}
}
\description{
Set start date from stop date.
}
\details{
Set the column 'stationstartdate' on the table 'fishstation' in \code{\link[RstoxData]{BioticData}}
 to the value of 'stationstopdate'.
 These columns are on the table 'fishstation' in data originating from NMDBiotic (http://www.imr.no/formats/nmdbiotic/).
 For bioticdata that does not conform to this, no modifications are done.
 Setting stationstartdate facilitates conversion to \code{\link[RstoxData]{StoxBioticData}} 
 using \code{\link[RstoxData]{StoxBiotic}}, and is applicable if daily resolution of date and time is not critical
 for subsequent analysis.
}
