% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxFunctions.R
\docType{data}
\name{stoxFunctionAttributes}
\alias{stoxFunctionAttributes}
\title{Function specification for inclusion in StoX GUI}
\format{An object of class \code{list} of length 1.}
\usage{
stoxFunctionAttributes
}
\description{
Function specification for inclusion in StoX GUI
}
\keyword{datasets}
