% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{TemporalCategories}
\alias{TemporalCategories}
\title{Temporal Categories (TemporalCategories)}
\description{
Table (\code{\link[data.table]{data.table}}) defining a categorical variable for grouping data based on date.
}
\details{
\describe{
  \item{temporalCategory}{character() Value of the temporal category}
  \item{startDay}{integer() Day of month for first day in the temporal category (1-based)}
  \item{startMonth}{integer() Month for first day in the temporal category (1-based)}
  \item{year}{integer() Year for which the category is defined, NA for seasonal definitions or for definition for a single unspecified year.}
 }

 Start and end of year is not implied as category delimitations when not included.
 If 1st of January is not definied as the start of a category,
 it is taken to be included in the last category of the preceding year.
}
