% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{TemporalDefinition}
\alias{TemporalDefinition}
\title{Temporal Categories (TemporalDefinition)}
\description{
Table (\code{\link[data.table]{data.table}}) defining a categorical variable for grouping data based on date.
}
\details{
\describe{
  \item{TemporalCategory}{character() Value of the temporal category}
  \item{StartDay}{integer() Day of month for first day in the temporal category (1-based)}
  \item{StartMonth}{integer() Month for first day in the temporal category (1-based)}
  \item{StartYear}{optional integer() Year for which the category is defined, omit this column for seasonal definitions.}
 }

 Start and end of year is not implied as category delimitations when not included.
 If 1st of January is not defined as the start of a category,
 it is taken to be included in the last category of the preceding year.
}
