% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoxDataTypes.R
\name{TripPartition}
\alias{TripPartition}
\title{Trip Partition}
\description{
Partitioning of catch from a trip.
}
\details{
list with two members 'fractions' and 'groupDefinition'

'fractions' is a \code{\link[data.table]{data.table}} with columns:
\describe{
 \item{tripid}{trip identifier}
 \item{species}{species identifier}
 \item{groupid}{identifies the group the fraction is provided for. Groups are further specified in 'groupDefinition'}
 \item{fraction}{fraction of the total catch of species in the given group}
}

'groupDefinition' is a \code{\link[data.table]{data.table}} with columns:
\describe{
 \item{groupid}{identifies the group the defintion is provided for.}
 \item{...}{one or more custom columns definint the group}
}
}
\concept{Data types}
